Reworked into separate library/spec, reducing module code
