+++ lib/msf/core/exploit/remote/http/rails_active_storage_vips.rb
@@ -1,0 +1,31 @@
+# frozen_string_literal: true
+
+module Msf
+  class Exploit
+    class Remote
+      module HTTP
+        # Helpers for the Rails Active Storage Vips file-read and signed-variation exploit chain.
+        module RailsActiveStorageVips
+          DEFAULT_ENVIRONMENT_MAX_BYTES = 65_536
+          DEFAULT_CREDENTIALS_MAX_BYTES = 262_144
+          DEFAULT_MAX_READ_REQUESTS = 4096
+          REPRESENTATION_REDIRECT_LIMIT = 3
+
+          class FlowError < StandardError; end
+          class ConfigError < StandardError; end
+          class DataError < StandardError; end
+          class ReadBudgetError < ConfigError; end
+          class TriggerError < StandardError; end
+          class UnreachableError < StandardError; end
+
+          include Msf::Exploit::Remote::HttpClient
+          include ActiveStorage
+          include FileReader
+          include MessageVerifier
+          include RailsSecrets
+          include Request
+        end
+      end
+    end
+  end
+end
