+++ lib/msf/core/exploit/remote/http/rails_active_storage_vips/png_decoder.rb
@@ -1,0 +1,167 @@
+# frozen_string_literal: true
+
+require 'zlib'
+
+module Msf::Exploit::Remote::HTTP::RailsActiveStorageVips
+  # Minimal PNG decoder for the grayscale representation responses.
+  # Chunk and filtering rules: https://www.w3.org/TR/png-3/
+  class PngDecoder
+    PNG_SIGNATURE = "\x89PNG\r\n\x1a\n".b
+    MAX_PIXELS = 1_048_576
+    MAX_DATA_BYTES = 4_194_304
+    MAX_CHUNKS = 1024
+    MAX_IDAT_BYTES = 3_145_728
+
+    # Decodes a bounded, 8-bit, non-interlaced grayscale PNG image.
+    #
+    # @param data [String] the binary PNG representation body
+    # @return [Hash{Symbol => Integer, String}] the image dimensions, channel count, and grayscale pixels
+    # @raise [TriggerError] if the body is malformed, unsupported, or exceeds a resource limit
+    def self.decode(data)
+      raise TriggerError, 'Representation body is not a PNG image' unless data.is_a?(String) && data.start_with?(PNG_SIGNATURE)
+      raise TriggerError, 'Representation PNG exceeded the maximum encoded size' if data.bytesize > MAX_DATA_BYTES
+
+      offset = PNG_SIGNATURE.bytesize
+      idat = String.new.b
+      chunk_count = 0
+      width = nil
+      height = nil
+      bit_depth = nil
+      color_type = nil
+      compression = nil
+      filter_method = nil
+      interlace = nil
+      saw_iend = false
+
+      while offset + 12 <= data.bytesize
+        chunk_count += 1
+        raise TriggerError, 'Representation PNG contained too many chunks' if chunk_count > MAX_CHUNKS
+
+        length = data.byteslice(offset, 4).unpack1('N')
+        raise TriggerError, 'Representation PNG contains an oversized chunk' if length > data.bytesize - offset - 12
+
+        chunk_type = data.byteslice(offset + 4, 4)
+        chunk_data = data.byteslice(offset + 8, length)
+        chunk_crc = data.byteslice(offset + 8 + length, 4).unpack1('N')
+        expected_crc = Zlib.crc32(chunk_type + chunk_data)
+        raise TriggerError, "Representation PNG #{chunk_type.inspect} chunk failed its CRC check" unless chunk_crc == expected_crc
+
+        case chunk_type
+        when 'IHDR'
+          raise TriggerError, 'Representation PNG contained an invalid IHDR chunk' unless length == 13 && width.nil?
+
+          width, height, bit_depth, color_type, compression, filter_method, interlace = chunk_data.unpack('NNC5')
+        when 'IDAT'
+          raise TriggerError, 'Representation PNG image data exceeded the compressed size limit' if idat.bytesize + length > MAX_IDAT_BYTES
+
+          idat << chunk_data
+        when 'IEND'
+          saw_iend = true
+          break
+        end
+
+        offset += 12 + length
+      end
+
+      raise TriggerError, 'Representation PNG is missing IHDR data' unless width && height
+      raise TriggerError, 'Representation PNG dimensions were invalid' unless width.positive? && height.positive? && width * height <= MAX_PIXELS
+      raise TriggerError, "Unsupported PNG bit depth #{bit_depth}" unless bit_depth == 8
+      raise TriggerError, "Unsupported PNG color type #{color_type}" unless color_type == 0
+      raise TriggerError, "Unsupported PNG compression method #{compression}" unless compression.zero?
+      raise TriggerError, "Unsupported PNG filter method #{filter_method}" unless filter_method.zero?
+      raise TriggerError, 'Interlaced PNG responses are not supported' unless interlace.zero?
+      raise TriggerError, 'Representation PNG is missing image data' if idat.empty?
+      raise TriggerError, 'Representation PNG is missing IEND data' unless saw_iend
+
+      stride = width
+      expected_size = height * (stride + 1)
+      raw = inflate_limited(idat, expected_size)
+      raise TriggerError, 'Representation PNG scanline data had an unexpected length' unless raw.bytesize == expected_size
+
+      previous = Array.new(stride, 0)
+      pixels = String.new.b
+      cursor = 0
+      height.times do
+        filter = raw.getbyte(cursor)
+        cursor += 1
+        row = raw.byteslice(cursor, stride).bytes
+        cursor += stride
+        decoded = unfilter(filter, row, previous)
+        pixels << decoded.pack('C*')
+        previous = decoded
+      end
+
+      { width: width, height: height, channels: 1, pixels: pixels }
+    rescue Zlib::Error => e
+      raise TriggerError, "Representation PNG decompression failed: #{e.message}"
+    end
+
+    def self.inflate_limited(compressed, expected_size)
+      inflater = Zlib::Inflate.new
+      inflated = String.new(capacity: expected_size).b
+      limit = expected_size + 1
+      inflater.inflate(compressed) do |chunk|
+        remaining = limit - inflated.bytesize
+        if chunk.bytesize > remaining
+          raise TriggerError, 'Representation PNG scanline data exceeded the expected length'
+        end
+
+        inflated << chunk
+      end
+      raise TriggerError, 'Representation PNG zlib stream was truncated' unless inflater.finished?
+      raise TriggerError, 'Representation PNG zlib stream contained trailing data' unless inflater.total_in == compressed.bytesize
+
+      inflated
+    ensure
+      inflater&.close unless inflater&.closed?
+    end
+
+    def self.unfilter(filter, row, previous)
+      case filter
+      when 0
+        row
+      when 1
+        decoded = []
+        row.each_index do |index|
+          left = index.zero? ? 0 : decoded[index - 1]
+          decoded << ((row[index] + left) & 0xff)
+        end
+        decoded
+      when 2
+        row.each_index.map { |index| (row[index] + previous[index]) & 0xff }
+      when 3
+        decoded = []
+        row.each_index do |index|
+          left = index.zero? ? 0 : decoded[index - 1]
+          decoded << ((row[index] + ((left + previous[index]) / 2)) & 0xff)
+        end
+        decoded
+      when 4
+        decoded = []
+        row.each_index do |index|
+          left = index.zero? ? 0 : decoded[index - 1]
+          upper = previous[index]
+          upper_left = index.zero? ? 0 : previous[index - 1]
+          decoded << ((row[index] + paeth(left, upper, upper_left)) & 0xff)
+        end
+        decoded
+      else
+        raise TriggerError, "Unsupported PNG scanline filter #{filter}"
+      end
+    end
+
+    def self.paeth(left, upper, upper_left)
+      estimate = left + upper - upper_left
+      left_distance = (estimate - left).abs
+      upper_distance = (estimate - upper).abs
+      upper_left_distance = (estimate - upper_left).abs
+
+      return left if left_distance <= upper_distance && left_distance <= upper_left_distance
+      return upper if upper_distance <= upper_left_distance
+
+      upper_left
+    end
+
+    private_class_method :inflate_limited, :unfilter, :paeth
+  end
+end
