+++ spec/lib/msf/core/exploit/remote/http/rails_active_storage_vips/active_storage_spec.rb
@@ -1,0 +1,286 @@
+# frozen_string_literal: true
+
+require 'spec_helper'
+require_relative 'support'
+
+RSpec.describe Msf::Exploit::Remote::HTTP::RailsActiveStorageVips::ActiveStorage do
+  include RailsActiveStorageVipsSpecSupport
+  include_context 'Msf::Simple::Framework'
+
+  subject(:exploit) { build_rails_active_storage_vips_host }
+
+  let(:namespace) { Msf::Exploit::Remote::HTTP::RailsActiveStorageVips }
+  let(:flow_error) { namespace::FlowError }
+  let(:config_error) { namespace::ConfigError }
+
+  before do
+    allow(exploit).to receive(:framework).and_return(framework)
+  end
+
+  describe '#report_active_storage_service' do
+    before do
+      exploit.datastore['RHOST'] = '192.0.2.10'
+      exploit.datastore['RPORT'] = 8080
+      exploit.datastore['TARGETURI'] = '/application'
+    end
+
+    it 'reports and memoizes the Rails application above HTTP and TCP' do
+      service = double('Rails service')
+      common = { host: '192.0.2.10', port: 8080, proto: 'tcp' }
+      expect(exploit).to receive(:report_service).once.with(
+        common.merge(
+          name: 'rails',
+          info: 'Ruby on Rails Active Storage',
+          resource: { uri: '/application' },
+          parents: common.merge(name: 'http', parents: common.merge(name: 'tcp', parents: nil))
+        )
+      ).and_return(service)
+
+      expect(exploit.send(:report_active_storage_service)).to be(service)
+      expect(exploit.send(:report_active_storage_service)).to be(service)
+    end
+
+    it 'reports HTTPS above SSL and TCP' do
+      exploit.datastore['SSL'] = true
+      service = double('Rails TLS service')
+      common = { host: '192.0.2.10', port: 8080, proto: 'tcp' }
+      expect(exploit).to receive(:report_service).with(
+        common.merge(
+          name: 'rails',
+          info: 'Ruby on Rails Active Storage',
+          resource: { uri: '/application' },
+          parents: common.merge(
+            name: 'https',
+            parents: common.merge(name: 'ssl', parents: common.merge(name: 'tcp', parents: nil))
+          )
+        )
+      ).and_return(service)
+
+      expect(exploit.send(:report_active_storage_service)).to be(service)
+    end
+  end
+
+  describe '#report_vuln' do
+    it 'links framework-generated vulnerability reports to the Rails service' do
+      service = double('Rails service')
+      vulnerability = double('Vulnerability', id: 7)
+      framework_db = framework.db
+      allow(exploit).to receive(:report_active_storage_service).and_return(service)
+      allow(exploit).to receive(:db).and_return(true)
+      allow(exploit).to receive(:myworkspace).and_return(:test_workspace)
+      allow(exploit).to receive(:mytask).and_return(nil)
+      expect(framework_db).to receive(:report_vuln).with(
+        hash_including(
+          host: '192.0.2.10',
+          name: exploit.fullname,
+          service: service
+        )
+      ).and_return(vulnerability)
+      allow(framework_db).to receive(:report_vuln_attempt)
+
+      result = exploit.report_vuln(host: '192.0.2.10', name: exploit.fullname)
+
+      expect(result).to be(vulnerability)
+      expect(exploit.instance_variable_get(:@vulnerability_reported)).to be(true)
+    end
+  end
+
+  describe '#service_details' do
+    it 'associates inferred session vulnerabilities with the Rails service' do
+      exploit.datastore['RHOST'] = '192.0.2.10'
+      exploit.datastore['RPORT'] = 8080
+
+      expect(exploit.service_details).to eq(
+        origin_type: :service,
+        protocol: 'tcp',
+        service_name: 'rails',
+        address: '192.0.2.10',
+        port: 8080
+      )
+    end
+  end
+
+  describe '#validate_options!' do
+    it 'rejects invalid local limits and malformed external origins' do
+      invalid_options = {
+        'REPRESENTATION_INDEX' => -1,
+        'EnvironmentMaxBytes' => 1023,
+        'CredentialsMaxBytes' => 1023,
+        'MaxReadRequests' => 0,
+        'ExternalStorageOrigins' => 'https://user:pass@storage.example'
+      }
+
+      invalid_options.each do |name, value|
+        original = exploit.datastore[name]
+        exploit.datastore[name] = value
+        expect { exploit.send(:validate_options!) }.to raise_error(config_error)
+        exploit.datastore[name] = original
+      end
+    end
+  end
+
+  describe '#representation_paths' do
+    it 'extracts modern and legacy representation URLs from src and srcset attributes' do
+      doc = Nokogiri::HTML(<<~HTML)
+        <img src="/assets/avatar.png"
+             srcset="/assets/avatar-small.png 1x, /rails/active_storage/representations/redirect/blob/variation/avatar.png 2x">
+        <img src="/rails/active_storage/representations/proxy/blob/variation/photo.png"
+             srcset="/rails/active_storage/representations/proxy/blob/variation/photo.png 1x, /assets/photo-large.png 2x">
+        <img src="/rails/active_storage/representations/blob/variation/legacy.png">
+      HTML
+
+      expect(exploit.send(:representation_paths, doc)).to eq(
+        [
+          '/rails/active_storage/representations/proxy/blob/variation/photo.png',
+          '/rails/active_storage/representations/blob/variation/legacy.png',
+          '/rails/active_storage/representations/redirect/blob/variation/avatar.png'
+        ]
+      )
+    end
+
+    it 'ignores non-hierarchical image sources' do
+      %w[data:image/png;base64,AAAA javascript:void(0) mailto:user@example.com].each do |source|
+        expect(exploit.send(:representation_path?, source)).to be(false)
+      end
+    end
+
+    it 'normalizes a selected non-hierarchical representation path to a flow error' do
+      expect do
+        exploit.send(:select_representation_path, ['mailto:user@example.com'], 'REPRESENTATIONURI')
+      end.to raise_error(flow_error, /non-hierarchical/i)
+    end
+
+    it 'substitutes signed IDs and variations in modern and legacy routes' do
+      modern = '/rails/active_storage/representations/redirect/old_blob/old_variation/image.png'
+      legacy = '/rails/active_storage/representations/old_blob/old_variation/image.png'
+
+      expect(exploit.send(:substitute_representation_blob, modern, 'new_blob', 'new.png')).to eq(
+        '/rails/active_storage/representations/redirect/new_blob/old_variation/new.png'
+      )
+      expect(exploit.send(:substitute_variation_key, legacy, 'new_variation')).to eq(
+        '/rails/active_storage/representations/old_blob/new_variation/image.png'
+      )
+    end
+
+    it 'uses the Active Storage route when the application base path also contains representations' do
+      path = '/representations/application/rails/active_storage/representations/redirect/old_blob/old_variation/image.png'
+
+      expect(exploit.send(:substitute_representation_blob, path, 'new_blob', 'new.png')).to eq(
+        '/representations/application/rails/active_storage/representations/redirect/new_blob/old_variation/new.png'
+      )
+    end
+
+    it 'uses the Active Storage route when the representation filename is representations' do
+      path = '/rails/active_storage/representations/redirect/old_blob/old_variation/representations'
+
+      expect(exploit.send(:substitute_variation_key, path, 'new_variation')).to eq(
+        '/rails/active_storage/representations/redirect/old_blob/new_variation/representations'
+      )
+    end
+
+    it 'constructs both standard representation route layouts' do
+      expect(exploit.send(:standard_representation_path, 'blob', 'variation', 'safe.png')).to eq(
+        '/rails/active_storage/representations/redirect/blob/variation/safe.png'
+      )
+      expect(exploit.send(:standard_representation_path, 'blob', 'variation', 'safe.png', route: :legacy)).to eq(
+        '/rails/active_storage/representations/blob/variation/safe.png'
+      )
+    end
+  end
+
+  describe '#direct_upload' do
+    it 'rejects successful responses whose JSON root is not an object' do
+      expect(exploit).not_to receive(:send_request_to_url)
+
+      ['[]', 'null', 'true'].each do |body|
+        allow(exploit).to receive(:send_request_cgi).and_return(http_response(200, body: body))
+
+        expect do
+          exploit.send(
+            :direct_upload,
+            csrf_token: 'csrf',
+            filename: 'image.png',
+            content_type: 'image/png',
+            content: 'PNG'
+          )
+        end.to raise_error(flow_error, /JSON object/i)
+      end
+    end
+
+    it 'reports Active Storage before uploading the blob content' do
+      service = double('Rails service')
+      create_response = http_response(
+        200,
+        body: {
+          signed_id: 'signed-id',
+          direct_upload: {
+            url: 'https://storage.example/object',
+            headers: {}
+          }
+        }.to_json
+      )
+      upload_response = http_response(204)
+      expect(exploit).to receive(:send_request_cgi).ordered.and_return(create_response)
+      expect(exploit).to receive(:report_active_storage_service).ordered.and_return(service)
+      expect(exploit).to receive(:send_request_to_url).ordered.with(
+        'https://storage.example/object',
+        method: 'PUT',
+        data: 'PNG',
+        headers: {}
+      ).and_return(upload_response)
+
+      signed_id = exploit.send(
+        :direct_upload,
+        csrf_token: 'csrf',
+        filename: 'image.png',
+        content_type: 'image/png',
+        content: 'PNG'
+      )
+
+      expect(signed_id).to eq('signed-id')
+      expect(exploit.instance_variable_get(:@active_storage_service)).to be(service)
+    end
+  end
+
+  describe '#submit_safe_upload' do
+    before do
+      exploit.datastore['RHOST'] = 'app.example'
+      exploit.datastore['RPORT'] = 443
+      exploit.datastore['SSL'] = true
+      exploit.datastore['VHOST'] = 'app.example'
+      exploit.datastore['SUBMITURI'] = '/uploads'
+      exploit.datastore['ATTACHMENT_FIELD'] = 'upload[avatar]'
+    end
+
+    it 'follows a bounded same-origin redirect to the representation page' do
+      redirect = http_response(302, location: '/uploads/7')
+      final = http_response(
+        200,
+        body: '<img src="/rails/active_storage/representations/blob/variation/image.png">'
+      )
+
+      expect(exploit).to receive(:send_request_cgi).ordered.and_return(redirect)
+      expect(exploit).to receive(:send_request_cgi).ordered do |options|
+        expect(options).to include('method' => 'GET', 'uri' => '/uploads/7')
+        final
+      end
+
+      expect(exploit.send(:submit_safe_upload, 'csrf', 'signed')).to eq(
+        '/rails/active_storage/representations/blob/variation/image.png'
+      )
+    end
+
+    it 'rejects a cross-origin form redirect' do
+      redirect = http_response(302, location: 'https://storage.example/uploads/7')
+      allow(exploit).to receive(:send_request_cgi).and_return(redirect)
+
+      expect { exploit.send(:submit_safe_upload, 'csrf', 'signed') }.to raise_error(flow_error, /different origin/i)
+    end
+
+    it 'normalizes an opaque redirect location to a module flow error' do
+      allow(exploit).to receive(:send_request_cgi).and_return(http_response(302, location: 'mailto:a'))
+
+      expect { exploit.send(:submit_safe_upload, 'csrf', 'signed') }.to raise_error(flow_error, /redirect was invalid/i)
+    end
+  end
+end
