+++ spec/lib/msf/core/exploit/remote/http/rails_active_storage_vips/png_decoder_spec.rb
@@ -1,0 +1,130 @@
+# frozen_string_literal: true
+
+require 'spec_helper'
+require_relative 'support'
+
+RSpec.describe Msf::Exploit::Remote::HTTP::RailsActiveStorageVips::PngDecoder do
+  include RailsActiveStorageVipsSpecSupport
+
+  let(:trigger_error) { Msf::Exploit::Remote::HTTP::RailsActiveStorageVips::TriggerError }
+
+  describe '.decode' do
+    it 'decodes all grayscale PNG scanline filters' do
+      cases = {
+        none: {
+          width: 3,
+          height: 1,
+          scanlines: [0, 10, 20, 30].pack('C*'),
+          expected: [10, 20, 30].pack('C*')
+        },
+        sub: {
+          width: 3,
+          height: 1,
+          scanlines: [1, 10, 10, 10].pack('C*'),
+          expected: [10, 20, 30].pack('C*')
+        },
+        up: {
+          width: 3,
+          height: 2,
+          scanlines: [0, 10, 20, 30, 2, 5, 5, 5].pack('C*'),
+          expected: [10, 20, 30, 15, 25, 35].pack('C*')
+        },
+        average: {
+          width: 3,
+          height: 2,
+          scanlines: [0, 10, 20, 30, 3, 10, 8, 8].pack('C*'),
+          expected: [10, 20, 30, 15, 25, 35].pack('C*')
+        },
+        paeth: {
+          width: 3,
+          height: 2,
+          scanlines: [0, 10, 100, 150, 4, 90, 206, 236].pack('C*'),
+          expected: [10, 100, 150, 100, 50, 80].pack('C*')
+        }
+      }
+
+      cases.each do |name, test_case|
+        png = grayscale_png(
+          width: test_case[:width],
+          height: test_case[:height],
+          scanlines: test_case[:scanlines]
+        )
+
+        expect(described_class.decode(png)).to eq(
+          {
+            width: test_case[:width],
+            height: test_case[:height],
+            channels: 1,
+            pixels: test_case[:expected]
+          }
+        ), "failed to decode the #{name} filter"
+      end
+    end
+
+    it 'rejects data without the PNG signature' do
+      expect { described_class.decode('not a png'.b) }.to raise_error(trigger_error, /not a PNG/i)
+    end
+
+    it 'rejects truncated chunks and IHDR data' do
+      truncated_chunk = described_class::PNG_SIGNATURE + [13].pack('N') + 'IHDR'.b + ("\x00".b * 9)
+      short_ihdr = described_class::PNG_SIGNATURE + png_chunk('IHDR'.b, "\x00".b * 5) + png_chunk('IEND'.b, ''.b)
+
+      expect { described_class.decode(truncated_chunk) }.to raise_error(trigger_error)
+      expect { described_class.decode(short_ihdr) }.to raise_error(trigger_error)
+    end
+
+    it 'rejects unsupported grayscale image properties' do
+      unsupported_properties = [
+        { bit_depth: 16 },
+        { color_type: 2 },
+        { compression: 1 },
+        { filter: 1 },
+        { interlace: 1 }
+      ]
+
+      unsupported_properties.each do |properties|
+        png = grayscale_png(width: 1, height: 1, scanlines: "\x00\x41".b, **properties)
+        expect { described_class.decode(png) }.to raise_error(trigger_error)
+      end
+    end
+
+    it 'rejects an unsupported scanline filter' do
+      png = grayscale_png(width: 1, height: 1, scanlines: "\x05\x41".b)
+
+      expect { described_class.decode(png) }.to raise_error(trigger_error, /scanline filter/i)
+    end
+
+    it 'normalizes corrupt and truncated zlib streams to TriggerError' do
+      corrupt = grayscale_png(width: 1, height: 1, idat: 'not zlib'.b)
+      truncated = grayscale_png(width: 1, height: 1, idat: "\x78\x9c".b)
+
+      expect { described_class.decode(corrupt) }.to raise_error(trigger_error, /decompression/i)
+      expect { described_class.decode(truncated) }.to raise_error(trigger_error, /(decompression|zlib)/i)
+    end
+
+    it 'rejects truncated inflated scanline data' do
+      png = grayscale_png(width: 3, height: 1, scanlines: "\x00\x41".b)
+
+      expect { described_class.decode(png) }.to raise_error(trigger_error, /scanline data/i)
+    end
+
+    it 'bounds decompression and rejects trailing compressed data' do
+      bomb = grayscale_png(width: 1, height: 1, idat: Zlib::Deflate.deflate('A' * 4_194_304))
+      trailing = grayscale_png(width: 1, height: 1, idat: Zlib::Deflate.deflate("\x00\x41".b) + 'junk')
+
+      expect { described_class.decode(bomb) }.to raise_error(trigger_error, /exceeded/i)
+      expect { described_class.decode(trailing) }.to raise_error(trigger_error, /trailing/i)
+    end
+
+    it 'bounds the encoded body, chunk count, and aggregate IDAT data' do
+      oversized_body = described_class::PNG_SIGNATURE + ("\x00".b * (described_class::MAX_DATA_BYTES - described_class::PNG_SIGNATURE.bytesize + 1))
+      ihdr = [1, 1, 8, 0, 0, 0, 0].pack('NNC5')
+      excessive_chunks = described_class::PNG_SIGNATURE + png_chunk('IHDR'.b, ihdr) + (png_chunk('tEXt'.b, ''.b) * described_class::MAX_CHUNKS)
+      excessive_idat = described_class::PNG_SIGNATURE + png_chunk('IHDR'.b, ihdr) + png_chunk('IDAT'.b, 'A'.b * (described_class::MAX_IDAT_BYTES + 1))
+
+      expect { described_class.decode(oversized_body) }.to raise_error(trigger_error, /encoded size/i)
+      expect { described_class.decode(excessive_chunks) }.to raise_error(trigger_error, /too many chunks/i)
+      expect { described_class.decode(excessive_idat) }.to raise_error(trigger_error, /compressed size/i)
+    end
+  end
+end
