+++ spec/lib/msf/core/exploit/remote/http/rails_active_storage_vips/rails_secrets_spec.rb
@@ -1,0 +1,207 @@
+# frozen_string_literal: true
+
+require 'spec_helper'
+require_relative 'support'
+
+RSpec.describe Msf::Exploit::Remote::HTTP::RailsActiveStorageVips::RailsSecrets do
+  include RailsActiveStorageVipsSpecSupport
+
+  subject(:host) { build_rails_active_storage_vips_host }
+
+  describe '#parse_environment' do
+    it 'parses NUL-delimited entries and preserves values containing equals signs' do
+      pixels = "SECRET_KEY_BASE=alpha=beta\x00EMPTY=\x00MALFORMED\x00DUPLICATE=first\x00DUPLICATE=last\x00".b
+
+      expect(host.send(:parse_environment, pixels)).to eq(
+        {
+          'SECRET_KEY_BASE' => 'alpha=beta',
+          'EMPTY' => '',
+          'DUPLICATE' => 'last'
+        }
+      )
+    end
+
+    it 'does not treat a key name embedded in another value as a complete entry' do
+      embedded = "OTHER=prefixSECRET_KEY_BASE=wrong\x00".b
+      anchored = embedded + "SECRET_KEY_BASE=right\x00".b
+
+      expect(host.send(:environment_contains_complete_key?, embedded)).to be(false)
+      expect(host.send(:environment_contains_complete_key?, anchored)).to be(true)
+    end
+
+    it 'does not stop at a master key before later environment entries are read' do
+      bytes = "RAILS_MASTER_KEY=#{'a' * 32}\x00RAILS_ENV=custom\x00".b
+
+      expect(host.send(:environment_contains_usable_key?, bytes, '/proc/self/environ', 'signed-id')).to be(false)
+    end
+
+    it 'stops only after SECRET_KEY_BASE validates against the signed blob ID' do
+      bytes = "SECRET_KEY_BASE=validated-secret\x00".b
+      validated = ['validated-secret', '/proc/self/environ', 'sha256', { encoded: 'token' }]
+      yielded = nil
+      allow(host).to receive(:validated_secret_candidate).and_return(validated)
+
+      result = host.send(:environment_contains_usable_key?, bytes, '/proc/self/environ', 'signed-id') do |candidate|
+        yielded = candidate
+      end
+
+      expect(result).to be(true)
+      expect(yielded).to eq(validated)
+    end
+  end
+
+  describe '#recover_secret_key_base' do
+    it 'links recovered environment loot to the Rails service' do
+      host.datastore['RHOST'] = '192.0.2.10'
+      service = double('Rails service')
+      bytes = "SECRET_KEY_BASE=validated-secret\x00".b
+      token_info = { encoded: 'token' }
+      validated = ['validated-secret', '/proc/self/environ', 'sha256', token_info]
+      context = {
+        mode: :raw,
+        representation: { signed_id: 'signed-id' }
+      }
+      allow(host).to receive(:try_read_text_file) do |_read_context, path, **_options|
+        path == '/proc/self/environ' ? bytes : nil
+      end
+      allow(host).to receive(:validated_secret_candidate)
+        .with('validated-secret', '/proc/self/environ', 'signed-id')
+        .and_return(validated)
+      allow(host).to receive(:report_active_storage_service).and_return(service)
+      expect(host).to receive(:store_loot).with(
+        'rails.process.environ',
+        'application/octet-stream',
+        host.rhost,
+        bytes,
+        'environ.bin',
+        'Recovered /proc/self/environ bytes',
+        service
+      ).and_return('/tmp/rails-process-environ.bin')
+
+      expect(host.send(:recover_secret_key_base, context)).to eq(validated)
+    end
+  end
+
+  describe '#decrypt_rails_credentials' do
+    let(:master_key) { '00112233445566778899aabbccddeeff' }
+    let(:encrypted_credentials) do
+      'L91+5oK6ok2VS0z8Yug1smZ69BUW0J3otX65PIRv40ZC4LM4PHsskwDI///UK9Olqw==' \
+        '--AAECAwQFBgcICQoL--1BoUyMzI7zRVEcKrsUPuzA=='
+    end
+    let(:plaintext_hex) do
+      '040849222c2d2d2d0a7365637265745f6b65795f626173653a2063726564656e7469616c2d7365637265740a063a064554'
+    end
+
+    it 'decrypts a Rails-compatible AES-128-GCM credentials envelope' do
+      plaintext = host.send(:decrypt_rails_credentials, encrypted_credentials, master_key)
+
+      expect(plaintext).to eq([plaintext_hex].pack('H*'))
+    end
+
+    it 'rejects a modified authentication tag' do
+      parts = encrypted_credentials.split('--')
+      tag = Base64.strict_decode64(parts.fetch(2))
+      tag.setbyte(0, tag.getbyte(0) ^ 1)
+      parts[2] = Base64.strict_encode64(tag)
+
+      expect(host.send(:decrypt_rails_credentials, parts.join('--'), master_key)).to be_nil
+    end
+
+    it 'rejects non-Rails GCM IV and authentication tag lengths' do
+      parts = encrypted_credentials.split('--')
+      iv = Base64.strict_decode64(parts.fetch(1))
+      tag = Base64.strict_decode64(parts.fetch(2))
+
+      [iv.byteslice(0, 11), iv + "\x00".b].each do |invalid_iv|
+        candidate = parts.dup
+        candidate[1] = Base64.strict_encode64(invalid_iv)
+        expect(host.send(:decrypt_rails_credentials, candidate.join('--'), master_key)).to be_nil
+      end
+      [tag.byteslice(0, 15), tag + "\x00".b].each do |invalid_tag|
+        candidate = parts.dup
+        candidate[2] = Base64.strict_encode64(invalid_tag)
+        expect(host.send(:decrypt_rails_credentials, candidate.join('--'), master_key)).to be_nil
+      end
+    end
+
+    it 'rejects malformed envelopes and invalid master keys' do
+      expect(host.send(:decrypt_rails_credentials, 'not--enough-parts', master_key)).to be_nil
+      expect(host.send(:decrypt_rails_credentials, '%%%--%%%--%%%', master_key)).to be_nil
+      expect(host.send(:decrypt_rails_credentials, encrypted_credentials, 'not-a-master-key')).to be_nil
+    end
+  end
+
+  describe 'legacy secrets.yml recovery' do
+    let(:environments) { %w[production staging development test] }
+    let(:secrets_key) { '000102030405060708090a0b0c0d0e0f' }
+
+    # Rails::Secrets encrypts secrets.yml.enc with ActiveSupport::MessageEncryptor
+    # (aes-128-gcm, Marshal serializer, no purpose), producing the same padded
+    # Base64 ciphertext--iv--tag envelope wrapped around a Marshal string.
+    def secrets_yml_enc(yaml, key_hex)
+      cipher = OpenSSL::Cipher.new('aes-128-gcm')
+      cipher.encrypt
+      cipher.key = [key_hex].pack('H*')
+      iv = "\x00".b * 12
+      cipher.iv = iv
+      cipher.auth_data = ''
+      encrypted = cipher.update(Marshal.dump(yaml)) + cipher.final
+      [encrypted, iv, cipher.auth_tag].map { |part| Base64.strict_encode64(part) }.join('--')
+    end
+
+    it 'decrypts and extracts secret_key_base from an encrypted secrets.yml.enc' do
+      yaml = "shared:\n  a: 1\nproduction:\n  secret_key_base: #{'c' * 128}\n"
+      envelope = secrets_yml_enc(yaml, secrets_key)
+
+      plaintext = host.send(:decrypt_rails_credentials, envelope, secrets_key)
+      unwrapped = host.send(:unwrap_marshaled_string, plaintext)
+
+      expect(host.send(:secret_key_base_candidates_from_secrets_yaml, unwrapped, environments)).to eq(['c' * 128])
+    end
+
+    it 'reads environment-keyed plaintext secrets.yml and skips unrendered ERB' do
+      yaml = "production:\n  secret_key_base: <%= ENV[\"SECRET_KEY_BASE\"] %>\ndevelopment:\n  secret_key_base: #{'d' * 128}\n"
+
+      expect(host.send(:secret_key_base_candidates_from_secrets_yaml, yaml, environments)).to eq(['d' * 128])
+    end
+
+    it 'applies environment values over shared values and retains the shared fallback' do
+      yaml = "shared:\n  secret_key_base: #{'s' * 128}\nproduction:\n  secret_key_base: #{'p' * 128}\ndevelopment:\n  other: value\n"
+
+      expect(host.send(:secret_key_base_candidates_from_secrets_yaml, yaml, environments)).to eq(['p' * 128, 's' * 128])
+    end
+
+    it 'supports safe YAML aliases used for shared secrets' do
+      yaml = "shared: &shared\n  secret_key_base: #{'s' * 128}\nproduction:\n  <<: *shared\n"
+
+      expect(host.send(:secret_key_base_candidates_from_secrets_yaml, yaml, ['production'])).to eq(['s' * 128])
+    end
+
+    it 'tries later environment candidates after a signature mismatch' do
+      production_secret = 'p' * 128
+      development_secret = 'd' * 128
+      yaml = "production:\n  secret_key_base: #{production_secret}\ndevelopment:\n  secret_key_base: #{development_secret}\n"
+      context = { representation: { signed_id: 'signed-id' } }
+      validated = [development_secret, '/proc/self/cwd/config/secrets.yml', 'sha1', { encoded: 'token' }]
+      allow(host).to receive(:try_read_text_file) do |_context, path, **_kwargs|
+        path == '/proc/self/cwd/config/secrets.yml' ? yaml.b : nil
+      end
+      expect(host).to receive(:validated_secret_candidate).ordered.with(production_secret, '/proc/self/cwd/config/secrets.yml', 'signed-id').and_return(nil)
+      expect(host).to receive(:validated_secret_candidate).ordered.with(development_secret, '/proc/self/cwd/config/secrets.yml', 'signed-id').and_return(validated)
+
+      expect(host.send(:recover_legacy_secrets, context, %w[production development], [])).to eq(validated)
+    end
+
+    it 'returns an empty array for documents with no usable secret and for malformed YAML' do
+      expect(host.send(:secret_key_base_candidates_from_secrets_yaml, "production:\n  a: 1\n", environments)).to be_empty
+      expect(host.send(:secret_key_base_candidates_from_secrets_yaml, '!!invalid: [', environments)).to be_empty
+    end
+
+    it 'rejects deeply nested target-controlled YAML without exhausting the Ruby stack' do
+      nested_yaml = ('[' * 2000) + '0' + (']' * 2000)
+
+      expect(host.send(:secret_key_base_candidates_from_secrets_yaml, nested_yaml, environments)).to be_empty
+      expect(host.send(:extract_secret_key_base_from_plaintext, nested_yaml)).to be_nil
+    end
+  end
+end
