+++ spec/lib/msf/core/exploit/remote/http/rails_active_storage_vips/request_spec.rb
@@ -1,0 +1,337 @@
+# frozen_string_literal: true
+
+require 'spec_helper'
+require_relative 'support'
+
+RSpec.describe Msf::Exploit::Remote::HTTP::RailsActiveStorageVips::Request do
+  include RailsActiveStorageVipsSpecSupport
+
+  subject(:exploit) { build_rails_active_storage_vips_host }
+
+  let(:namespace) { Msf::Exploit::Remote::HTTP::RailsActiveStorageVips }
+  let(:png_decoder) { namespace::PngDecoder }
+  let(:flow_error) { namespace::FlowError }
+  let(:config_error) { namespace::ConfigError }
+  let(:trigger_error) { namespace::TriggerError }
+
+  describe '#request_representation' do
+    before do
+      exploit.datastore['RHOST'] = 'app.example'
+      exploit.datastore['RPORT'] = 443
+      exploit.datastore['SSL'] = true
+      exploit.datastore['VHOST'] = 'app.example'
+      exploit.datastore['COOKIE'] = 'session=application-secret'
+      exploit.datastore['HttpUsername'] = 'application-user'
+      exploit.datastore['HttpPassword'] = 'application-password'
+      exploit.datastore['ExternalStorageOrigins'] = 'https://storage.example'
+    end
+
+    it 'keeps cross-origin redirects isolated when the chain returns to the application origin' do
+      application_redirect = http_response(302, location: 'https://storage.example/object?signature=public')
+      storage_redirect = http_response(302, location: 'https://app.example/rails/active_storage/object')
+      final_response = http_response(200)
+      sanitized_requests = []
+      sanitized_responses = [storage_redirect, final_response]
+
+      expect(exploit).to receive(:send_request_cgi).once do |options|
+        expect(options).to include(
+          'method' => 'GET',
+          'uri' => '/rails/active_storage/representations/redirect/blob/variation/image.png',
+          'cookie' => 'session=application-secret'
+        )
+        application_redirect
+      end
+      allow(exploit).to receive(:send_sanitized_request) do |options|
+        sanitized_requests << options
+        sanitized_responses.shift
+      end
+
+      response = exploit.send(
+        :request_representation,
+        'https://app.example/rails/active_storage/representations/redirect/blob/variation/image.png'
+      )
+
+      expect(response).to be(final_response)
+      expect(sanitized_requests.map { |request| request['rhost'] }).to eq(%w[storage.example app.example])
+      sanitized_requests.each do |request|
+        expect(request.keys).not_to include('cookie', 'authorization', 'username', 'password')
+        expect(request.fetch('headers')).to eq('Accept' => 'image/png', 'Connection' => 'close')
+      end
+    end
+
+    it 'rejects a redirect chain that exceeds the bounded hop count' do
+      redirect = http_response(302, location: '/rails/active_storage/representations/redirect/next')
+      limit = namespace::REPRESENTATION_REDIRECT_LIMIT
+
+      expect(exploit).to receive(:send_request_cgi).exactly(limit + 1).times.and_return(redirect)
+
+      expect do
+        exploit.send(
+          :request_representation,
+          'https://app.example/rails/active_storage/representations/redirect/blob/variation/image.png'
+        )
+      end.to raise_error(flow_error, /redirect limit/i)
+    end
+
+    it 'rejects malformed redirect locations as a module flow error' do
+      expect(exploit).to receive(:send_request_cgi).once.and_return(http_response(302, location: 'mailto:a'))
+
+      expect do
+        exploit.send(
+          :request_representation,
+          'https://app.example/rails/active_storage/representations/redirect/blob/variation/image.png'
+        )
+      end.to raise_error(flow_error, /redirect was invalid/i)
+    end
+
+    it 'rejects opaque redirect locations as a module flow error' do
+      expect(exploit).to receive(:send_request_cgi).once.and_return(http_response(302, location: 'mailto:user@example.com'))
+
+      expect do
+        exploit.send(
+          :request_representation,
+          'https://app.example/rails/active_storage/representations/redirect/blob/variation/image.png'
+        )
+      end.to raise_error(flow_error, /not an HTTP\(S\) URL/i)
+    end
+
+    it 'requires an exact operator-approved origin before following an external redirect' do
+      exploit.datastore['ExternalStorageOrigins'] = nil
+      expect(exploit).to receive(:send_request_cgi).once.and_return(http_response(302, location: 'https://storage.example/object'))
+      expect(exploit).not_to receive(:send_sanitized_request)
+
+      expect do
+        exploit.send(
+          :request_representation,
+          'https://app.example/rails/active_storage/representations/redirect/blob/variation/image.png'
+        )
+      end.to raise_error(config_error, /ExternalStorageOrigins/)
+    end
+
+    it 'rejects oversized final and redirect response bodies' do
+      oversized = 'A'.b * (png_decoder::MAX_DATA_BYTES + 1)
+      [http_response(200, body: oversized), http_response(302, location: '/next', body: oversized)].each do |response|
+        allow(exploit).to receive(:send_request_cgi).and_return(response)
+
+        expect do
+          exploit.send(
+            :request_representation,
+            'https://app.example/rails/active_storage/representations/redirect/blob/variation/image.png'
+          )
+        end.to raise_error(flow_error, /exceeded/i)
+      end
+    end
+
+    it 'recognizes an absolute IPv6 literal URL as same-origin' do
+      exploit.datastore['RHOST'] = '2001:db8::10'
+      exploit.datastore['RPORT'] = 3003
+      exploit.datastore['SSL'] = false
+      exploit.datastore['VHOST'] = nil
+      response = http_response(200)
+
+      expect(exploit).not_to receive(:send_sanitized_request)
+      expect(exploit).to receive(:send_request_cgi).with(
+        hash_including(
+          'method' => 'GET',
+          'uri' => '/rails/active_storage/representations/redirect/blob/variation/image.png'
+        )
+      ).and_return(response)
+
+      result = exploit.send(
+        :request_representation,
+        'http://[2001:db8::10]:3003/rails/active_storage/representations/redirect/blob/variation/image.png'
+      )
+
+      expect(result).to be(response)
+    end
+  end
+
+  describe '#send_request_to_url' do
+    it 'passes an unbracketed IPv6 hostname to the sanitized HTTP client' do
+      exploit.datastore['ExternalStorageOrigins'] = 'https://[2001:db8::20]'
+      response = http_response(204)
+      expect(exploit).to receive(:send_sanitized_request).with(
+        hash_including(
+          'rhost' => '2001:db8::20',
+          'rport' => 443,
+          'SSL' => true,
+          'vhost' => '2001:db8::20'
+        )
+      ).and_return(response)
+
+      result = exploit.send(
+        :send_request_to_url,
+        'https://[2001:db8::20]/object',
+        method: 'PUT',
+        data: 'content',
+        headers: {}
+      )
+
+      expect(result).to be(response)
+    end
+
+    it 'allows stock storage headers but rejects authority, framing, authentication, and control headers' do
+      exploit.datastore['RHOST'] = 'app.example'
+      exploit.datastore['RPORT'] = 443
+      exploit.datastore['SSL'] = true
+      valid_headers = {
+        'Content-Type' => 'image/png',
+        'Content-MD5' => 'checksum',
+        'Content-Disposition' => 'inline',
+        'Cache-Control' => 'max-age=3600',
+        'x-amz-meta-test' => 'one',
+        'x-goog-meta-test' => 'two',
+        'x-ms-blob-type' => 'BlockBlob',
+        'X-Optional' => nil
+      }
+      response = http_response(204)
+      expect(exploit).to receive(:send_request_cgi).with(
+        hash_including(
+          'headers' => valid_headers.compact.merge('Connection' => 'close')
+        )
+      ).and_return(response)
+
+      expect(
+        exploit.send(
+          :send_request_to_url,
+          'https://app.example/object',
+          method: 'PUT',
+          data: 'content',
+          headers: valid_headers
+        )
+      ).to be(response)
+
+      invalid_headers = [
+        { 'Host' => 'internal.example' },
+        { 'content-length' => '1' },
+        { 'Transfer-Encoding' => 'chunked' },
+        { 'Authorization' => 'Bearer reflected-secret' },
+        { 'Cookie' => 'session=reflected-secret' },
+        { "X-Bad\r\nInjected" => 'value' },
+        { 'X-Bad' => "value\r\nInjected: true" },
+        { 'X-Bad' => 7 },
+        { 'X-Test' => 'one', 'x-test' => 'two' },
+        { 'X-Test' => nil, 'x-test' => 'two' },
+        (0..described_class::MAX_UPLOAD_HEADERS).to_h { |index| ["X-Header-#{index}", 'value'] },
+        { 'X-Large' => 'A' * (described_class::MAX_UPLOAD_HEADER_VALUE_BYTES + 1) },
+        (0...9).to_h { |index| ["X-Large-#{index}", 'A' * described_class::MAX_UPLOAD_HEADER_VALUE_BYTES] }
+      ]
+      invalid_headers.each do |headers|
+        expect do
+          exploit.send(
+            :send_request_to_url,
+            'https://app.example/object',
+            method: 'PUT',
+            data: 'content',
+            headers: headers
+          )
+        end.to raise_error(flow_error)
+      end
+    end
+  end
+
+  describe '#send_sanitized_request' do
+    it 'uses a fresh client with cookies and application authentication disabled' do
+      exploit.datastore['COOKIE'] = 'session=application-secret'
+      exploit.datastore['HttpUsername'] = 'application-user'
+      exploit.datastore['HttpPassword'] = 'application-password'
+      logger = instance_double(Rex::Proto::Http::HttpLoggerSubscriber)
+      client = instance_double(Rex::Proto::Http::Client)
+      request = instance_double(Rex::Proto::Http::Request)
+      response = http_response(200)
+      request_options = nil
+
+      allow(Rex::Proto::Http::HttpLoggerSubscriber).to receive(:new).and_return(logger)
+      expect(Rex::Proto::Http::Client).to receive(:new) do |*arguments, **keywords|
+        expect(arguments.values_at(0, 1, 3, 6, 7)).to eq(['storage.example', 443, true, '', ''])
+        expect(keywords).to include(kerberos_authenticator: nil, subscriber: logger)
+        client
+      end
+      expect(client).to receive(:set_config).with(
+        hash_including(
+          'vhost' => 'storage.example',
+          'ssl_server_name_indication' => 'storage.example',
+          'raw_headers' => ''
+        )
+      )
+      expect(client).to receive(:request_cgi) do |options|
+        request_options = options
+        request
+      end
+      expect(client).to receive(:_send_recv).with(request, 20).and_return(response)
+      expect(client).to receive(:close)
+
+      result = exploit.send(
+        :send_sanitized_request,
+        {
+          'method' => 'GET',
+          'uri' => '/object',
+          'headers' => { 'Accept' => 'image/png' },
+          'cookie' => 'must-not-survive',
+          'authorization' => 'must-not-survive',
+          'username' => 'must-not-survive',
+          'password' => 'must-not-survive',
+          'rhost' => 'storage.example',
+          'rport' => 443,
+          'SSL' => true,
+          'vhost' => 'storage.example'
+        }
+      )
+
+      expect(result).to be(response)
+      expect(request_options).to include(
+        'cookie' => nil,
+        'raw_headers' => '',
+        'authorization' => nil,
+        'username' => '',
+        'password' => '',
+        'preferred_auth' => 'None',
+        kerberos_authenticator: false
+      )
+      expect(request_options.to_s).not_to include('application-secret', 'application-user', 'application-password', 'must-not-survive')
+    end
+  end
+
+  describe '#sanitized_request_timeout' do
+    it 'uses a finite fallback when HttpClientTimeout is unset or non-positive' do
+      [nil, 0, -1].each do |configured_timeout|
+        exploit.datastore['HttpClientTimeout'] = configured_timeout
+
+        expect(exploit.send(:sanitized_request_timeout)).to eq(20)
+      end
+    end
+
+    it 'honors a positive HttpClientTimeout' do
+      exploit.datastore['HttpClientTimeout'] = 7.5
+
+      expect(exploit.send(:sanitized_request_timeout)).to eq(7.5)
+    end
+  end
+
+  describe '#trigger_variation' do
+    before do
+      exploit.datastore['RHOST'] = 'app.example'
+      exploit.datastore['RPORT'] = 443
+      exploit.datastore['SSL'] = true
+      exploit.datastore['VHOST'] = 'app.example'
+    end
+
+    it 'treats redirects and definitive client-side rejection statuses as trigger failures' do
+      [300, 301, 302, 303, 307, 308, 400, 401, 403, 404, 405, 414, 422, 429, 499].each do |status|
+        allow(exploit).to receive(:send_request_cgi).and_return(http_response(status))
+
+        expect do
+          exploit.send(:trigger_variation, '/rails/active_storage/representations/redirect/blob/forged/image.png')
+        end.to raise_error(trigger_error, /rejected with HTTP #{status}/)
+      end
+    end
+
+    it 'allows an ambiguous server error because command execution may have occurred first' do
+      allow(exploit).to receive(:send_request_cgi).and_return(http_response(500))
+
+      expect do
+        exploit.send(:trigger_variation, '/rails/active_storage/representations/redirect/blob/forged/image.png')
+      end.not_to raise_error
+    end
+  end
+end
