+++ spec/lib/msf/core/exploit/remote/http/rails_active_storage_vips/support.rb
@@ -1,0 +1,76 @@
+# frozen_string_literal: true
+
+require 'base64'
+require 'zlib'
+require 'msf/core/exploit/remote/http/rails_active_storage_vips'
+
+# Shared host and data builders for the Rails Active Storage Vips library specs.
+module RailsActiveStorageVipsSpecSupport
+  DEFAULT_DATASTORE = {
+    'RHOST' => '192.0.2.10',
+    'RPORT' => 3000,
+    'SSL' => false,
+    'TARGETURI' => '/',
+    'LANDINGURI' => '/',
+    'SUBMITURI' => '/posts',
+    'DIRECTUPLOADURI' => '/rails/active_storage/direct_uploads',
+    'ATTACHMENT_FIELD' => 'post[image]',
+    'REPRESENTATION_INDEX' => 0,
+    'KEY_GENERATOR_DIGEST' => 'auto',
+    'VERIFIER_DIGEST' => 'auto',
+    'MESSAGE_SERIALIZER' => 'auto',
+    'EnvironmentMaxBytes' => 65_536,
+    'CredentialsMaxBytes' => 262_144,
+    'MaxReadRequests' => 4096,
+    'ExternalStorageOrigins' => nil
+  }.freeze
+
+  def build_rails_active_storage_vips_host
+    host_class = Class.new(Msf::Exploit::Remote) do
+      include Msf::Exploit::Remote::HTTP::RailsActiveStorageVips
+    end
+
+    host_class.new.tap do |host|
+      DEFAULT_DATASTORE.each { |name, value| host.datastore[name] = value }
+    end
+  end
+
+  def png_chunk(type, data)
+    [data.bytesize].pack('N') + type + data + [Zlib.crc32(type + data)].pack('N')
+  end
+
+  def grayscale_png(width:, height:, scanlines: ''.b, idat: nil, **properties)
+    properties = {
+      bit_depth: 8,
+      color_type: 0,
+      compression: 0,
+      filter: 0,
+      interlace: 0
+    }.merge(properties)
+    ihdr = [width, height, *properties.values_at(:bit_depth, :color_type, :compression, :filter, :interlace)].pack('NNC5')
+    compressed = idat.nil? ? Zlib::Deflate.deflate(scanlines) : idat
+
+    described_class::PNG_SIGNATURE + png_chunk('IHDR'.b, ihdr) + png_chunk('IDAT'.b, compressed) + png_chunk('IEND'.b, ''.b)
+  end
+
+  def http_response(code, location: nil, body: nil)
+    response = Rex::Proto::Http::Response.new(code)
+    response.headers['Location'] = location if location
+    response.body = body if body
+    response
+  end
+
+  def striped_pixels(layout, source)
+    dimension = layout.fetch(:dimension)
+    columns = (1...(dimension - 1)).step(2).first(layout.fetch(:data_columns))
+    pixels = "\x00".b * (dimension * dimension)
+
+    columns.each_with_index do |column, column_index|
+      dimension.times do |row|
+        pixels.setbyte((row * dimension) + column, source.getbyte((column_index * dimension) + row))
+      end
+    end
+
+    pixels
+  end
+end
